% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustGamma.R
\name{adjustGamma}
\alias{adjustGamma}
\title{Gamma correction of the amplitude}
\usage{
\S4method{adjustGamma}{GPR}(x, a = 1, b = 1, track = TRUE)
}
\description{
This function compute the gamma correction.
}
