% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplEnv.R
\name{amplEnv}
\alias{amplEnv}
\alias{amplEnv,GPR-method}
\title{Amplitude envelope}
\usage{
amplEnv(x, method = c("peak", "hilbert"), npad = 100, threshold = 2)

\S4method{amplEnv}{GPR}(x, method = c("peak", "hilbert"), npad = 100, threshold = 2)
}
\arguments{
\item{x}{An object of the class GPR.}

\item{method}{(\code{character[1]}) Method to use. See details.}

\item{npad}{(\code{integer[1]}) Only for \code{method = "hilbert"}. Positive
value defining the number of values to
pad \code{x} (the padding help to minimize the Gibbs effect at
the beginning and end of the data caused by the Hilbert
transform).}

\item{threshold}{(\code{numeric[1]}) Threshold value for peak detection. The
larger the value, the longer the computation time.}
}
\description{
Estimate for each trace the amplitude envelope with the Hilbert
transform (instataneous amplitude).
}
\details{
Two methods:
\itemize{
\item \code{hilbert} Envelope based on the Hilbert transform
\item  \code{peak} The local maxima of the absolute values of the
signal are first estimated. The envelope is determined using spline
interpolation over local maxima.
}
}
