% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle.R
\name{angle}
\alias{angle}
\alias{angle,GPR-method}
\alias{angle,GPRsurvey-method}
\title{Angle of the GPR data}
\usage{
angle(x)

\S4method{angle}{GPR}(x)

\S4method{angle}{GPRsurvey}(x)
}
\arguments{
\item{x}{[\code{GPR|GPRsurvey}]}
}
\value{
[\code{numeric(1)}] The angle of the oriented bounding box.
}
\description{
The angle is computed based on the orientation of the oriented bounding
box (\code{\link{obbox}}).
}
\concept{spatial computation}
