% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ann.R
\name{ann}
\alias{ann}
\alias{ann<-}
\alias{ann,GPR-method}
\alias{ann<-,GPR-method}
\title{Annotations of the GPR data}
\usage{
ann(x)

ann(x) <- value

\S4method{ann}{GPR}(x)

\S4method{ann}{GPR}(x) <- value
}
\arguments{
\item{x}{[\code{GPR}] An object of the class GPR.}

\item{value}{[\code{character}]}
}
\value{
[\code{GPR}]
}
\description{
Annotations are currently used to store the intersections of GPR lines.
}
\concept{getters/setters}
