% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{antfreq}
\alias{antfreq}
\alias{antfreq<-}
\title{Antenna frequency in MHz}
\usage{
\S4method{antfreq}{GPR}(x)

\S4method{antfreq}{GPR}(x) <- value
}
\description{
Antenna frequency of the antennas in MHz
}
