% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/getterSetter.R
\docType{methods}
\name{antsep}
\alias{antsep}
\alias{antsep<-}
\title{Antenna separation}
\usage{
\S4method{antsep}{GPR}(x)

\S4method{antsep}{GPR}(x) <- value
}
\description{
For common-offset GPR data, the antenna separation is constant.
For multi-offset GPR data (e.g., CMP, WARR), the antenna separation increase
from trace to trace.
}
