% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{antSepFromAntFreq}
\alias{antSepFromAntFreq}
\title{Estimate antenna separation from antenna frequency}
\usage{
antSepFromAntFreq(antfreq, verbose = TRUE)
}
\arguments{
\item{antfreq}{[\code{numeric(1)}] Antenna frequency.}

\item{verbose}{[\code{logical(1)}] If \code{TRUE} the function gives a
message.}
}
\value{
[\code{numeric(1)}] Antenna separation
}
\description{
Estimate antenna separation from antenna frequency
}
