% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\docType{methods}
\name{apply,GPR-method}
\alias{apply,GPR-method}
\title{Apply a function along the rows (samples per trace) or columns (traces)}
\usage{
\S4method{apply}{GPR}(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{[\code{GPR}]}

\item{MARGIN}{[\code{integer}] A vector giving the subscripts which the 
function will be applied over
(see \code{\link{apply}}).}

\item{FUN}{[\code{function}] The function to be applied
(see \code{\link{apply}}).}

\item{...}{Additional parameters to be passed (see \code{\link{apply}}).}
}
\description{
Apply a function along the rows (samples per trace) or columns (traces)
}
