% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupArith_GPRvirtual.R
\docType{methods}
\name{Arith,GPRvirtual,ANY-method}
\alias{Arith,GPRvirtual,ANY-method}
\alias{Arith}
\alias{Arith,GPRvirtual,GPRvirtual-method}
\alias{Arith,ANY,GPRvirtual-method}
\title{Basic arithmetical functions}
\usage{
\S4method{Arith}{GPRvirtual,ANY}(e1, e2)

\S4method{Arith}{GPRvirtual,GPRvirtual}(e1, e2)

\S4method{Arith}{ANY,GPRvirtual}(e1, e2)
}
\arguments{
\item{e1}{An object of the class GPRvirtual}

\item{e2}{An object of the class GPRvirtual}
}
\description{
Basic arithmetical functions
}
