% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{readGPR}
\alias{readGPR}
\title{Basic mathematical functions}
\usage{
\S4method{Math}{GPR}(x)
}
\arguments{
\item{x}{An object of the class RGPR.}
}
\description{
Basic mathematical functions
}
\examples{
data(frenkeLine00)
A <- frenkeLine00
}

