% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R, R/coercion_spatial.R
\name{as.sf}
\alias{as.sf}
\alias{as.sf,GPR-method}
\alias{as.sf,GPRsurvey-method}
\title{Coerce object to an sf POINT object}
\usage{
as.sf(x)

\S4method{as.sf}{GPR}(x)

\S4method{as.sf}{GPRsurvey}(x)

as.sf(x)

\S4method{as.sf}{GPR}(x)

\S4method{as.sf}{GPRsurvey}(x)
}
\arguments{
\item{x}{[\code{class GPR|GPRsurvey}]}
}
\value{
[\code{class sf}] Geometry type is \code{POINT}.

[\code{class sf}] Geometry type is \code{POINT}.
}
\description{
Coerce object to a simple feature (sf) geometry POINT. 
Only coordinates are coerced, all additional information is lost.

Coerce object to a simple feature (sf) geometry POINT. 
Only coordinates are coerced, no additional information.
}
\concept{coercion}
