% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion_spatial.R
\name{as.spatialLines}
\alias{as.spatialLines}
\alias{as.spatialLines,GPRsurvey-method}
\alias{as.spatialLines,GPR-method}
\title{Coerce object to an object of the class sf lines}
\usage{
as.spatialLines(x)

\S4method{as.spatialLines}{GPRsurvey}(x)

\S4method{as.spatialLines}{GPR}(x)
}
\arguments{
\item{x}{(\verb{class GPR|GPRsurvey})}
}
\value{
(\verb{SpatialLines object})
}
\description{
Coerce object to an object of the class sf lines
}
\concept{coercion}
