% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atan2.R
\name{atan2}
\alias{atan2}
\alias{atan2,GPRvirtual,ANY-method}
\alias{atan2,GPRvirtual,GPRvirtual-method}
\alias{atan2,ANY,GPRvirtual-method}
\title{Arc-tangent}
\usage{
atan2(y, x)

\S4method{atan2}{GPRvirtual,ANY}(y, x)

\S4method{atan2}{GPRvirtual,GPRvirtual}(y, x)

\S4method{atan2}{ANY,GPRvirtual}(y, x)
}
\arguments{
\item{y}{(\code{GPR|numeric|matrix|complex})}

\item{x}{(\code{GPR|numeric|matrix|complex})}
}
\value{
(\code{GPR})
}
\description{
The arc-tangent of two arguments atan2(y, x) returns the angle between
the x-axis and the vector from the origin to (x, y), i.e.,
for positive arguments atan2(y, x) == atan(y/x).
}
\concept{arithmetic}
