% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{bits2volt}
\alias{bits2volt}
\title{Bytes to volt conversion
      
Convert bytes to volt values}
\usage{
bits2volt(Vmax = 50, Vmin = 50, nbits = 16)
}
\arguments{
\item{Vmax}{[\code{numeric(1)}] Maximal nominal analog input voltage. 
If \code{Vmax = NULL} it returns \code{1} (no bytes to volt 
transformation)}

\item{Vmin}{[\code{numeric(1)}] Minimal nominal analog input voltage. 
If missing, then \code{Vmin = -Vmax}.}

\item{nbits}{[\code{integer(1)} Number of bits.}
}
\description{
Bytes to volt conversion
      
Convert bytes to volt values
}
