% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer}
\alias{buffer}
\alias{buffer,GPR-method}
\alias{buffer,GPRsurvey-method}
\alias{buffer,sfc-method}
\alias{buffer,sf-method}
\title{Buffer the GPR lines}
\usage{
buffer(x, d, combine = TRUE)

\S4method{buffer}{GPR}(x, d)

\S4method{buffer}{GPRsurvey}(x, d, combine = TRUE)

\S4method{buffer}{sfc}(x, d, combine = TRUE)

\S4method{buffer}{sf}(x, d, combine = TRUE)
}
\arguments{
\item{x}{[\code{GPR|GPRsurvey}] An object of the class \code{GPR} or
\code{GPRsurvey}}

\item{d}{[\code{numeric(1)}] buffer distance}

\item{combine}{[\code{logical(1)}] If \code{TRUE}, returns the buffer for
all the GPR lines together. Otherwise it returns the buffer
for each line.}
}
\value{
[\code{sfc}] Polygon as a simple feature geometry list-column.
}
\description{
Returns the buffered lines as polygon(s)
}
