% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{byte2volt}
\alias{byte2volt}
\title{Bytes to volt conversion
      
Convert bytes to volt values}
\usage{
byte2volt(V = c(-50, 50), nBytes = 16)
}
\description{
Bytes to volt conversion
      
Convert bytes to volt values
}
