% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{capFirstLetter}
\alias{capFirstLetter}
\title{Capitalize the first letter in a word string in R}
\usage{
capFirstLetter(x)
}
\arguments{
\item{x}{[character(1)] A character string}
}
\value{
The character string with first letter capitalised
}
\description{
Capitalize the first letter in a word string in R
}
