% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_checkArgs.R
\name{checkArg}
\alias{checkArg}
\title{Check argument}
\usage{
checkArg(x, u, type, y, ...)
}
\arguments{
\item{x}{Object to test.}

\item{u}{[\code{character}] Message.}

\item{type}{[\code{character}] Code for what to check.}

\item{y}{Additional argument for checking (e.g., length).}

\item{...}{Additional argument for checking.}
}
\description{
To be used inside a function.
}
