% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\docType{methods}
\name{clip}
\alias{clip}
\alias{clip,GPR-method}
\title{Clip the amplitude}
\usage{
clip(x, xclip = NULL, track = TRUE)

\S4method{clip}{GPR}(x, xclip = NULL, track = TRUE)
}
\arguments{
\item{x}{[\code{GPR}] GPR object.}

\item{xclip}{[\code{numeric(1|2)}] Value above and below which the signal has 
to be clipped. If \code{xclip} is a length-one
vector, the signal outside the range 
\code{-xclip}, \code{xclip} will be clipped.
If \code{xclip} is a length-one
vector, the signal outside the range 
\code{xclip[1]}, \code{xclip[2]} will be 
clipped.}

\item{track}{[\code{logical(1)}] If \code{TRUE}, processing will be tracked.}
}
\value{
[\code{GPR class}] Clipped GPR object.
}
\description{
Clip the amplitude

Clip the amplitude
}
