% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clippedData.R
\name{clippedData}
\alias{clippedData}
\alias{clippedData,GPR-method}
\title{Return the clipped values of the GPR signal as GPR object}
\usage{
clippedData(obj, dlim = NULL, verbose = TRUE)

\S4method{clippedData}{GPR}(obj, dlim = NULL, verbose = TRUE)
}
\arguments{
\item{obj}{(\code{GPR})}

\item{dlim}{(\code{numeric[2]}) The min and max clipped values if they are
a priori known. If \code{dlim = NULL}, the function tries to
get the the clipped values from \code{metadata(obj)$clip}. If
there are no clipped values, the function tries to estimate
them.}

\item{verbose}{(\code{logical[1]}) If \code{FALSE}, all messages and warnings are
suppressed (use with care).}
}
\value{
(\code{GPR}) The object with the clipped values.
}
\description{
Max and min values
}
\concept{signal processing}
