% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{closestTr}
\alias{closestTr}
\title{Closest trace}
\usage{
closestTr(x, y)
}
\arguments{
\item{x}{Object of the class GPR.}

\item{y}{Length-two numeric vector of the (x, y)-coordinates.}
}
\value{
Indice (integer) of the closest trace.
}
\description{
Return the indice of the closest trace to the point \code{y}
}
