% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{CMPAnalysis}
\alias{CMPAnalysis}
\title{Common mid-point (CMP) analysis}
\usage{
\S4method{CMPAnalysis}{GPR}(x, method = c("semblance", "winsemblance",
  "wincoherence", "wincoherence2"), v = NULL, asep = NULL, w = NULL)
}
\arguments{
\item{x}{An object of the class \code{GPR}}

\item{method}{A length-one character vector}

\item{v}{A numeric vector defining at which velocities the analysis is
performed}

\item{asep}{A length-n numeric vector defining the antenna separation for
each trace (n = number of traces)}

\item{w}{A length-one numeric vector defining the window length for the
methods 'wincoherence' and 'wincoherence2'.}
}
\description{
either use 'rec' and 'trans' to compute the distance between the antennas
or give the distance between the antennas (asep)
or seq(x@antsep, by = x@dx, length.out = length(x))
}
