% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMPhyperbolas.R
\name{CMPhyperbolas}
\alias{CMPhyperbolas}
\alias{CMPhyperbolas,GPR-method}
\title{Returns hyperbolas associated with Vrms velociies}
\usage{
CMPhyperbolas(x)

\S4method{CMPhyperbolas}{GPR}(x)
}
\arguments{
\item{x}{[\code{GPR class}] A CMP object of the class \code{GPR}}
}
\value{
[\code{list}] A list element key \code{antsep} containing a numeric 
       vector of \eqn{n} antenna separation values, and element \code{twt} 
       containing a \eqn{n \times m} matrix, where \eqn{m} is the
       number of hyperbolas (i.e., the number of velocities).
}
\description{
Returns hyperbolas associated with root-mean-square velocities stored in
the CMP data.
}
\examples{
\dontrun{
HPB <- hyperbolaFromVrms(x)
plot(x, barscale = FALSE, main = "CMP")
matplot(HPB$antsep, HPB$twt, type = "l", col = "green", 
        lwd = 2, add = TRUE, lty = 1)
}        
}
