% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRvirtual.R, R/ClassGPR.R
\docType{methods}
\name{as.vector,GPRvirtual-method}
\alias{as.vector,GPRvirtual-method}
\alias{as.numeric,GPRvirtual-method}
\alias{as.matrix,GPR-method}
\title{Coercion functions from \code{GPR}}
\usage{
\S4method{as.vector}{GPRvirtual}(x, mode = "any")

\S4method{as.numeric}{GPRvirtual}(x, ...)

\S4method{as.matrix}{GPR}(x)
}
\description{
Coerce an object of the class \code{GPR} into an object of the class
\code{matrix}, \code{vector}, \code{numeric}
}
\examples{
data(frenkeLine00)
x <- as.matrix(frenkeLine00)
x <- as.vector(frenkeLine00)
x <- as.numeric(frenkeLine00)
}
