% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\name{as.GPR.matrix}
\alias{as.GPR.matrix}
\alias{as.GPR.list}
\title{Coercion functions to \code{GPR}}
\usage{
as.GPR.matrix(x, ...)

as.GPR.list(x, ...)
}
\description{
Coerce a \code{matrix}, a \code{list} into an object of the class \code{GPR}.
}
\examples{
x <- matrix(rnorm(50), nrow = 10, ncol = 5)
y <- as.GPR.matrix(x)
y2 <- as(x, "GPR")
x <- list(data = matrix(rnorm(50), nrow = 10, ncol = 5),
          x    = seq(0, 10, length.out = 5),
          z    = seq(0, 50, length.out = 10))
y <- as.GPR.list(x)
y2 <- as(x, "GPR")
}
