% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupCompare_GPRvirtual.R
\name{Compare,GPRvirtual,ANY-method}
\alias{Compare,GPRvirtual,ANY-method}
\alias{Compare}
\alias{Compare,GPRvirtual,GPRvirtual-method}
\alias{Compare,ANY,GPRvirtual-method}
\title{Comparison operations}
\usage{
\S4method{Compare}{GPRvirtual,ANY}(e1, e2)

\S4method{Compare}{GPRvirtual,GPRvirtual}(e1, e2)

\S4method{Compare}{ANY,GPRvirtual}(e1, e2)
}
\arguments{
\item{e1}{An object of the class GPRvirtual}

\item{e2}{An object of the class GPRvirtual}
}
\description{
Methods for the base Compare methods \link[methods:S4groupGeneric]{S4groupGeneric}: "==", ">", "<", "!=", "<=", ">="
}
