% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComplex_GPRvirtual.R
\name{Complex,GPRvirtual-method}
\alias{Complex,GPRvirtual-method}
\title{Complex operations}
\usage{
\S4method{Complex}{GPRvirtual}(z)
}
\arguments{
\item{z}{An object of the class GPRvirtual}
}
\description{
Methods for the base Complex methods \link[methods]{S4groupGeneric}: "Arg", "Conj", "Im", "Mod", "Re"
}
