% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{conv1D}
\alias{conv1D}
\alias{conv2D}
\title{Trace convolution (1D)}
\usage{
\S4method{conv1D}{GPR}(x, w)

\S4method{conv2D}{GPR}(x, w)
}
\arguments{
\item{x}{A GPR data}

\item{w}{A numeric vector defining a wavelet or a matrix with number of
columns equal to the number of traces.}

\item{x}{A GPR data}

\item{w}{A numeric matrix with smaller dimension than the GPR data.}
}
\value{
The convolved GPR data.

The convolved GPR data.
}
\description{
Convolution of the GPR traces with a wavelet

Convolution of the GPR data with a kernel
}

