% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTimeToDepth.R
\name{convertTimeToDepth}
\alias{convertTimeToDepth}
\title{Time to depth conversion}
\usage{
convertTimeToDepth(
  x,
  dz = NULL,
  zmax = NULL,
  method = c("pchip", "linear", "nearest", "cubic", "spline"),
  track = TRUE
)

\S4method{convertTimeToDepth}{GPR}(
  x,
  dz = NULL,
  zmax = NULL,
  method = c("pchip", "linear", "nearest", "cubic", "spline"),
  track = TRUE
)
}
\arguments{
\item{dz}{vertical resolution of the migrated data. 
If \code{dz = NULL}, then \code{dz} is set equal to the 
smallest depth resolution inferred from the data.}

\item{zmax}{maximum depth for the time to depth conversion. If 
\code{zmax = NULL}, then \code{zmax} is set equal to the 
largest depth inferred from the data.}

\item{method}{method for the interpolation 
(see \code{\link[signal]{interp1}}).}
}
\description{
Convert the two-way travel time of the recorded waves into depth. It does
not account for the topography. To add the topography, 
use \code{\link{migrate}} instead.
}
