% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexhull.R
\name{convexhull}
\alias{convexhull}
\alias{convexhull,GPRsurvey-method}
\alias{convexhull,GPR-method}
\alias{convexhull,matrix-method}
\title{Convex hull of GPR data}
\usage{
convexhull(x, verbose = FALSE)

\S4method{convexhull}{GPRsurvey}(x, verbose = FALSE)

\S4method{convexhull}{GPR}(x, verbose = FALSE)

\S4method{convexhull}{matrix}(x, verbose = FALSE)
}
\arguments{
\item{x}{[\code{GPR|GPRsurvey}]}

\item{verbose}{[\code{logical(1)}] If \code{FALSE}, all messages and warnings 
are suppressed (use with care).}
}
\description{
Return the convex hull.
}
