% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.R
\name{convmtx}
\alias{convmtx}
\title{Returns the convolution matrix
 
Returns the convolution matrix, A, 
such that the product of A and a vector, x, 
is the convolution of y and x.}
\usage{
convmtx(w, n)
}
\value{
If y is a column vector of length m, A is m-by-nf.
}
\description{
Returns the convolution matrix
 
Returns the convolution matrix, A, 
such that the product of A and a vector, x, 
is the convolution of y and x.
}
