% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convUnit.R
\name{convUnit}
\alias{convUnit}
\title{Utility function for unit conversion}
\usage{
convUnit(x, unitIn, unitOut)
}
\arguments{
\item{x}{[\code{numeric}] Numeric vector of values to be converted}

\item{unitIn}{[\code{character}] Current unit of the \code{x} 
(see \code{\link[units]{valid_udunits}} for a list of valid units).}

\item{unitOut}{[\code{character}] Unit in which to convert \code{x} 
(see \code{\link[units]{valid_udunits}} for a list of valid units).}
}
\description{
Convert numerical values from one unit to another unit
}
\examples{
# Conversion from feets to meters
v_ft <- 1:10
v_m <- convUnit(v_ft, unitIn = "ft", unitOut = "m")

# Conversion from meters to milimeters
v_mm <- convUnit(v_m, unitIn = "m", unitOut = "mm")
}
