% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{coord}
\alias{coord}
\alias{coord<-}
\alias{coord,GPR-method}
\alias{coord<-,GPR-method}
\title{Coordinates of the GPR data}
\usage{
\S4method{coord}{GPR}(x, i, ...)

\S4method{coord}{GPR}(x) <- value
}
\description{
Coordinates of the GPR data
}
