% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\docType{methods}
\name{coord}
\alias{coord}
\alias{coord<-}
\alias{coord<-,GPR-method}
\alias{coord,GPR-method}
\alias{coord,GPRsurvey-method}
\alias{coord<-,GPRsurvey-method}
\title{Trace coordinates}
\usage{
coord(x)

coord(x) <- value

\S4method{coord}{GPR}(x)

\S4method{coord}{GPR}(x) <- value

\S4method{coord}{GPRsurvey}(x)

\S4method{coord}{GPRsurvey}(x) <- value
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}

\item{value}{[\code{matrix(n,3)|list}] coordinates (x, y, z)}
}
\value{
[\code{GPR class}] An object of the class \code{GPR}
}
\description{
Return or update the trace coordinates (x, y, z). Not that you cannot 
change the number of coordinates with \code{coord}.
}
\details{
Modified slots class GPR
\itemize{
  \item \code{coord} the trace coordinates
  \item \code{x} the local trace position (along profile)
}
Modified slots class GPRsurvey
\itemize{
  \item \code{coords} the trace coordinates
  \item \code{xlengths} the local trace position (along profile)
  \item \code{intersections} the local trace position (along profile)
}
}
