% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRsurvey.R
\docType{methods}
\name{coordref}
\alias{coordref}
\alias{coordref<-}
\alias{coordref,GPRsurvey-method}
\title{Define a local reference coordinate}
\usage{
\S4method{coordref}{GPRsurvey}(x)
}
\description{
Define a local reference coordinate
}
