% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getterSetter.R
\name{coordref}
\alias{coordref}
\alias{coordref<-}
\title{Coordref of the GPR data}
\usage{
\S4method{coordref}{GPRsurvey}(x)

\S4method{coordref}{GPRsurvey}(x) <- value
}
\description{
Coordref of the GPR data
}
