% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRsurvey.R
\docType{methods}
\name{coords}
\alias{coords}
\alias{coords<-}
\alias{coords,GPRsurvey-method}
\alias{coords<-,GPRsurvey-method}
\title{Return coordinates}
\usage{
\S4method{coords}{GPRsurvey}(x, i)

\S4method{coords}{GPRsurvey}(x) <- value
}
\description{
Return coordinates
}
