% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{corAntElev}
\alias{corAntElev}
\title{Correct for the elevation of antenna above the ground}
\usage{
\S4method{corAntElev}{GPR}(x, c0 = 0.3)
}
\description{
For air-lauched GPR or airborn GPR
}
