% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctNMO.R
\docType{methods}
\name{correctNMO}
\alias{correctNMO}
\alias{correctNMO,GPR-method}
\title{Normal Move-Out correction}
\usage{
correctNMO(x, thrs = NULL, v = NULL, method = c("linear", "nearest",
  "pchip", "cubic", "spline"))

\S4method{correctNMO}{GPR}(x, thrs = NULL, v = NULL,
  method = c("linear", "nearest", "pchip", "cubic", "spline"))
}
\arguments{
\item{x}{An object of the class \code{GPR}}

\item{thrs}{[\code{numeric(1)|NULL}] Definite the threshold for muting
(i.e., suppressing) the values where the NMO-stretching is
above the threshold. Setting \code{thrs = NULL}, the full data
will be used.}

\item{v}{A length-one numeric vector defining the radar wave velocity in 
the ground}

\item{method}{[\code{character(1)}] Interpolation method to be applied:
one of \code{pchip}, \code{linear}, \code{nearest}, 
\code{spline}, \code{cubic} 
(see also \code{\link[signal]{interp1}}).}
}
\description{
Remove the Normal Move-Out (NMO) from the trace given a constant velocity: 
this is a non-linear 
correction of the time axis that requires interpolation. Note that
only the conventional NMO correction is currently implemented. The 
conventional NMO introduces a streching effect. A nonstretch NMO will
be implemented in a near future. The Normal Move-out is defined as the
difference between the two-way time at a given offset and the two-way 
zero-offset time.
}
\details{
Assuming a horizontal reflecting plane and homogeneous medium, the two-way
bistatic travel time of the reflected wave 
for an antenna separation \eqn{x} follows directly from the Pythagorean 
theorem:
\deqn{t_{TWT}(x,z) = \sqrt{\frac{x^2}{v^2} + \frac{4z^2}{v^2}}}
where \eqn{t_{TWT}(x)} is the two-way travel time at antenna
separation \eqn{x} of the wave reflected at depth \eqn{z} with propagation
velocity \eqn{v}. This equation defines an hyperbola (keep \eqn{z} constant,
increase the antenna separation \eqn{x} and you obtain a hyperbola similar
to the reflection signals you obtain with common-mid point survey).
The idea behind NMO-correction is to correct the signal for the antenna 
separation (offset) and therefore to transform the signal to the signal we 
would have recorded with zero offset (\eqn{x = 0}). We write the vertical
two-way traveltime at zero offset 
\deqn{t_0 = t_{TWT}(x = 0) = \frac{2z}{v}}
Therefore, the NMO-correction \eqn{\Delta_{NMO}} is
\deqn{\Delta_{NMO} = t_{TWT}(x) - t_0}  
\deqn{\Delta_{NMO} = t_0 (\sqrt{1 + \frac{x^2}{v^2 t_0^2}} - 1)}
}
\references{
\itemize{
  \item{Tillard and Dubois (1995) Analysis of GPR data: wave propagation
        velocity determination. Journal of Applied Geophysics, 33:77-91}
  \item{Shatilo and Aminzadeh (2000) Constant normal-moveout (CNMO) 
        correction: a technique and test results. Geophysical Prospecting,
        473-488}
}
}
