% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCubeFromGrid.R
\name{createCubeFromGrid}
\alias{createCubeFromGrid}
\alias{createCubeFromGrid,GPRsurvey-method}
\title{Create a GPRcube from a GPR grid (GPRsurvey)}
\usage{
createCubeFromGrid(
  obj,
  dz = NULL,
  direction = c("x", "y"),
  name = "",
  desc = "",
  track = TRUE
)

\S4method{createCubeFromGrid}{GPRsurvey}(
  obj,
  dz = NULL,
  direction = c("x", "y"),
  name = "",
  desc = "",
  track = TRUE
)
}
\arguments{
\item{obj}{(\code{GPRsurvey}) An object of the class GPRsurvey.}

\item{dz}{(\code{NULL|numeric[1]}) Vertical resolution (if NULL, the mean
vertical resolution of the data is taken).}

\item{name}{(\code{character[1]}) Name of the GPRcube.}

\item{desc}{(\code{character[1]}) Description of the GPRcube.}

\item{track}{(\code{logical[1]}) Should the processing step be tracked?}
}
\value{
(\code{GPRcube}) An object of the class GPRcube.
}
\description{
Create a GPRcube from a GPR grid (GPRsurvey)
}
\concept{processing}
