% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{Crop GPR data}
\usage{
\S4method{crop}{GPR}(x, xlim = NULL, ylim = NULL, track = TRUE)
}
\arguments{
\item{x}{GPR data}

\item{xlim}{[\code{numeric(2)|NULL}] Trace position range. If equal to 
\code{NULL} the range is equal to the data range.}

\item{ylim}{[\code{numeric(2)|NULL}] two-way travel time/depth range. 
If equal to \code{NULL} the range is equal to the data range.}
}
\description{
Crop GPR data based on trace position (\code{xlim}) and/or 
two-way travel time/depth (\code{ylim}). One of the arguments \code{xlim} and
\code{ylim} can be set equal to \code{NULL}. In this case, the whole data
range wi
}
