% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsUnit.R
\name{crsUnit}
\alias{crsUnit}
\title{Unit of Coordinate Reference System (CRS)}
\usage{
crsUnit(x)
}
\arguments{
\item{x}{[\code{character}] CRS (one or more)}
}
\value{
[\code{character}] The unit abbreviation (except for degree)
}
\description{
Returns the unit of the CRS
}
