% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declip.R
\name{declip}
\alias{declip}
\alias{declip,GPR-method}
\title{Declip the GPR signal}
\usage{
declip(obj, drange = NULL, lambda = 1, mu = 0, objclip = NULL)

\S4method{declip}{GPR}(obj, drange = NULL, lambda = 1, mu = 0, objclip = NULL)
}
\arguments{
\item{obj}{(\code{GPR})}

\item{drange}{(\code{numeric[2]}) the desired range of the declipped values
(for the negatively and positively clipped values) used as
a soft constraint whose strength depends on the value of
\code{lambda}. If \code{drange = NULL}, the desired range is
set equal to the clipped data range +/- 10 percent.}

\item{lambda}{(\code{numeric[1]}) Positive value defining the strength of
signal range constraint. The larger \code{lamda} is the larger
the constraint on the signal range.
If \code{lambda = 0}, there is no constraint on the data range
and \code{drange} is not used.}

\item{mu}{(\code{numeric[1]}) Positive value that add some noise to
stabilize the matrix inversion used in the least-square
approach (could be useful if the matrix to invert is singular,
use small value, e.g., 0.00001 of the signal amplitude).}

\item{objclip}{(\code{matrix[m,n]}) The clipped values, a matrix
with the Same dimension as \code{obj}, with \code{1} for
the positively clipped values, \code{-1} for the negatively
clipped values and \code{0} everywhere else. If
\code{objclip = NULL}, the clipped values are estimated with
the function \code{\link[=clippedData]{clippedData()}}.}
}
\value{
(\code{GPR})
}
\description{
Use constrained least squares. Based on the code of Ivan Selesnick:
we minimize the energy of the the third derivative. This encourages the
filled in data to have the form of a parabola (second order polynomial),
because the third derivative of a parabola is zero.".
}
\details{
M. J. Harvilla and R. M. Stern, "Efficient audio declipping using regularized least squares," 2015 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP), South Brisbane, QLD, Australia, 2015, pp. 221-225, doi: 10.1109/ICASSP.2015.7177964.
}
\concept{signal processing}
