% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{deconv}
\alias{deconv}
\title{Deconvolution}
\usage{
\S4method{deconv}{GPR}(x, method = c("spiking", "wavelet", "min-phase",
  "mixed-phase"), ...)
}
\arguments{
\item{method}{Type of deconvolution method.}

\item{...}{additional arguments, see \code{details}.}
}
\value{
A list containing the deconvolued GPR data (and possibly other
variables.
}
\description{
A generic function to perform different types of convolution
}
\section{Spiking and mixed-phase deconvolution}{

The required arguments for \code{method = "spiking"} and 
\code{method = "mixed-phase"} are:
\itemize{
  \item \code{W}: A length-two numeric vector defining the time/depth window
                  for which the wavelet is estimated
  \item \code{wtr}: A length-one numeric vector defining the number of 
                    neighorough traces to be combine into a "super trace"
                    (the total number of traces is \code{2*wtr + 1}).
  \item \code{nf} A length-one numeric vector defining the filter length.
  \item \code{mu} A length-one numeric vector defining the amount of noise.
}
}

\section{Wavelet deconvolution}{

The required arguments for \code{method = "wavelet"} are:
\itemize{
  \item \code{h}: A numeric vector corresponding to the wavelet used to
                  deconvolve the GPR data.
  \item \code{mu} A length-one numeric vector defining the amount of noise.
}
}

