% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{deconvolve}
\alias{deconvolve}
\title{Deconvolution with known wavelet}
\usage{
deconvolve(y, h, mu = 1e-04, type = c("matrix", "FFT"), regDeriv = FALSE)
}
\value{
Vector with same length as \code{y}
}
\description{
convolution model: \eqn{y = h \times x} where h and y are known, 
x is unknown. Spectral or matrix-based.
}
