% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{deconvolveF}
\alias{deconvolveF}
\title{spectral deconvolution with known wavelet}
\usage{
deconvolveF(y, h, mu = 1e-04)
}
\value{
Vector with same length as \code{y}
}
\description{
convolution model: \eqn{y = h \times x} where h and y are known, 
x is unknown
}
