% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{deconvolveMtx}
\alias{deconvolveMtx}
\title{Matrix based deconvolution}
\usage{
deconvolveMtx(y, h, lambda)
}
\value{
Vector with same length as \code{y}
}
\description{
same as \code{deconvolveF} but matrix based (same results)
}
