% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{deconvolveSparse}
\alias{deconvolveSparse}
\title{Sparse deconvolution}
\usage{
deconvolveSparse(
  y,
  w,
  Nit = 50,
  eps = 0.01,
  sigma = 0.1,
  thr = NULL,
  phifun,
  wfun,
  ...
)
}
\description{
Reference: Penalty and Threshold Functions for Sparse Signal Processing
Ivan Selesnick, NYU-Poly, selesi@poly.edu, 2012
http://eeweb.poly.edu/iselesni/lecture_notes/
}
