% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate.R, R/delineations.R,
%   R/rmDelineations.R, R/identifyDelineations.R, R/exportDelineations.R
\name{delineate}
\alias{delineate}
\alias{delineate,GPR-method}
\alias{plotDelineations}
\alias{plotDelineations,GPR-method}
\alias{delineations}
\alias{delineations,GPR-method}
\alias{rmDelineations<-}
\alias{rmDelineations<-,GPR-method}
\alias{identifyDelineations}
\alias{identifyDelineations,GPR-method}
\alias{exportDelineations}
\alias{exportDelineations,GPR-method}
\title{Delineate structure on GPR data}
\usage{
delineate(x, name = NULL, values = NULL, n = 10000, plotDel = NULL, ...)

\S4method{delineate}{GPR}(x, name = NULL, values = NULL, n = 10000, plotDel = NULL, ...)

plotDelineations(
  x,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  ...
)

\S4method{plotDelineations}{GPR}(
  x,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  ...
)

delineations(obj, name = NULL)

\S4method{delineations}{GPR}(obj, name = NULL)

rmDelineations(x) <- value

\S4method{rmDelineations}{GPR}(x) <- value

identifyDelineations(
  obj,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  ...
)

\S4method{identifyDelineations}{GPR}(
  obj,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  ...
)

exportDelineations(obj, dsn = ".")

\S4method{exportDelineations}{GPR}(obj, dsn = ".")
}
\arguments{
\item{x}{(\code{GPR})}

\item{name}{(\code{character[n]}) Names of the delineations.}

\item{values}{(\code{list}) list of x and y coordinates (optional). If \code{NULL},
\code{locator()} function is run.}

\item{n}{(\code{numeric[1]}) the maximum number of points to locate. Valid values start at 1.}

\item{plotDel}{(\code{logical[1]}) If \code{TRUE}plot delineation.}

\item{...}{Additional parameters (not yet used)}

\item{method}{(\code{character[1]}) Interpolation method.}

\item{col}{(\code{character[n]}) \code{n} colors for the \code{n} delineations}

\item{obj}{(\code{GPR* object}) An object of the class \code{GPR}}

\item{value}{(\code{integer[n]|all}) Either a vector of indexes of delineations to
to remove or \code{all} to remove all delineations,}

\item{dsn}{(\verb{character(1)|connection object}) data source name:
either the filepath to the GPR data (character),
or an open file connection.}
}
\description{
Print and invisible returns the GPR data delineations.

Works only close to the points !!!
}
