% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{depth}
\alias{depth}
\alias{depth<-}
\alias{depth}
\alias{depth<-}
\title{Depth/time of the GPR data}
\usage{
\S4method{depth}{GPR}(x)

\S4method{depth}{GPR}(x) <- value
}
\description{
Depth/time of the GPR data
}
