% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{depthunit}
\alias{depthunit}
\alias{depthunit<-}
\alias{depthunit}
\alias{depthunit<-}
\title{Depth unit of the GPR data}
\usage{
\S4method{depthunit}{GPR}(x)

\S4method{depthunit}{GPR}(x) <- value
}
\description{
Depth unit of the GPR data
}
