% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{description}
\alias{description}
\alias{description<-}
\title{Description of the GPR data}
\usage{
\S4method{description}{GPR}(x)

\S4method{description}{GPR}(x) <- value
}
\description{
Description of the GPR data
}
