% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\docType{methods}
\name{dim,GPR-method}
\alias{dim,GPR-method}
\title{Dimensions of a GPR object 
Return the number of rows (samples per trace) and 
columns (samples per trace).}
\usage{
\S4method{dim}{GPR}(x)
}
\arguments{
\item{x}{[\code{GPR}]}
}
\description{
Dimensions of a GPR object 
Return the number of rows (samples per trace) and 
columns (samples per trace).
}
