% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\name{dim,GPRvirtual-method}
\alias{dim,GPRvirtual-method}
\title{Dimensions of a GPR object}
\usage{
\S4method{dim}{GPRvirtual}(x)
}
\arguments{
\item{x}{(\code{GPR})}
}
\description{
Return the number of rows (samples per trace) and
columns (samples per trace).
}
\concept{array}
