% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{distTensors}
\alias{distTensors}
\title{Distance between structure tensors}
\usage{
distTensors(J1, J2, method = c("geodesic", "log-Euclidean", "angular"),
  normalise = FALSE)
}
\description{
Distance between structure tensors
}

