% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmStripes.R
\name{.estimateStripeStrength}
\alias{.estimateStripeStrength}
\title{Estimate stripe strength per column}
\usage{
.estimateStripeStrength(S)
}
\arguments{
\item{S}{Numeric matrix.}
}
\value{
Numeric vector of length \code{ncol(S)} representing
relative stripe strength per column.
}
\description{
Computes a quantitative estimate of striping intensity
for each column of a gridded image by measuring
high-frequency perpendicular differences.
}
\details{
This metric is used to drive adaptive smoothing.
}
