% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_update.R
\name{.h5_lock_release}
\alias{.h5_lock_release}
\title{Release a lock acquired with \code{.h5_lock_acquire()}}
\usage{
.h5_lock_release(lockdir)
}
\arguments{
\item{lockdir}{(\code{character(1)}) Value returned by \code{.h5_lock_acquire()}.}
}
\description{
Release a lock acquired with \code{.h5_lock_acquire()}
}
\keyword{internal}
