% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_update.R
\name{.h5_read_r_object}
\alias{.h5_read_r_object}
\title{Read back an object written with \code{.h5_write_r_object()}}
\usage{
.h5_read_r_object(grp, name)
}
\arguments{
\item{grp}{An open \code{hdf5r} group.}

\item{name}{(\code{character(1)}) Dataset name.}
}
\value{
The original R object.
}
\description{
Read back an object written with \code{.h5_write_r_object()}
}
\keyword{internal}
