% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_update.R
\name{.h5_temp_path}
\alias{.h5_temp_path}
\title{Build a temporary path in the same directory as \code{dsn}}
\usage{
.h5_temp_path(dsn)
}
\description{
Using the same directory guarantees \code{file.rename()} is an atomic,
same-filesystem operation when the file is later swapped into place.
}
\keyword{internal}
