% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGPR.R
\name{.maybe_set_cmp_mode}
\alias{.maybe_set_cmp_mode}
\title{Set CMP mode fields when the acquisition mode is CMP}
\usage{
.maybe_set_cmp_mode(x)
}
\arguments{
\item{x}{GPR object.}
}
\value{
Updated GPR object.
}
\description{
Set CMP mode fields when the acquisition mode is CMP
}
\keyword{internal}
