% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGPRsurvey.R
\name{.read_GPR_line_hdf5}
\alias{.read_GPR_line_hdf5}
\title{Read a single GPR line from an HDF5 file}
\usage{
.read_GPR_line_hdf5(dsn, name)
}
\arguments{
\item{dsn}{(\code{character(1)}) Path to the HDF5 file.}

\item{name}{(\code{character(1)}) Line name (sub-group under \verb{/lines/}).}
}
\value{
Object of class \code{GPR}.
}
\description{
Read a single GPR line from an HDF5 file
}
\keyword{internal}
