% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_update.R
\name{.write_intersections_hdf5}
\alias{.write_intersections_hdf5}
\title{Write line intersection data into the HDF5 file}
\usage{
.write_intersections_hdf5(h5, obj)
}
\arguments{
\item{h5}{Open, writable \link[hdf5r:H5File]{hdf5r::H5File} handle.}

\item{obj}{\code{GPRsurvey} object, typically just returned by
\code{\link[=findIntersection]{findIntersection()}}.}
}
\description{
Writes the \verb{@intersections} slot (if non-empty) to \verb{/survey/intersections}
as one dataset per named element. Assumes \verb{/survey} already exists (i.e.
this is called after \code{.write_survey_group_hdf5()}).
}
\keyword{internal}
