% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{eigenFilter}
\alias{eigenFilter}
\title{Eigenimage filter}
\usage{
\S4method{eigenFilter}{GPR}(x, eigenvalue = NA, center = TRUE,
  scale = FALSE)
}
\arguments{
\item{x}{An object of the class GPR}

\item{eigenvalue}{An integer vector specifying the eigenvalues selected. 
\code{eigenvalue = 1:5} returns the image reconstructed 
using the first five eigenvalues while \code{c(2,4,6)}
will return the image reconstructed for these three 
specific eigenvalues. 
If \code{NA}, a plot of the eigenvalues spectrum 
will be displayed and the user will be prompted to enter 
the pair of selected eigenvalues separated by a comma. 
That is for example \code{1,2} will return the image 
reconstructred using tthe first two eigenvalues. 
If \code{NULL}, the image is reconstructed using 
all eigenvalues.}

\item{center}{Logical or numeric. If \code{TRUE}, centering is done by 
subtracting the layer means (omitting \code{NA}'s), and 
if \code{FALSE}, no centering is done. If center is a 
numeric vector with length equal to the 
\code{nlayers(x)}, then each layer of \code{x} has the 
corresponding value from center subtracted from it.}

\item{scale}{Logical or numeric. If \code{TRUE}, scaling is done by 
dividing the (centered) layers of \code{x} by their 
standard deviations if center is \code{TRUE}, and the 
root mean square otherwise. 
If scale is \code{FALSE}, no scaling is done. If scale is
a numeric vector with length equal to \code{nlayers(x)}, 
each layer of \code{x} is divided by the corresponding 
value. Scaling is done after centering.}
}
\value{
An object of the class GPR.
}
\description{
Decompose the GPR data (radargram) using singular value decomposition (SVD) 
and return the reconstructed data for the selected singular values 
(eigenvalues). This method is sometimes refered to in the litterature as 
the Karhunen-Loeve (KL) transformation or the eigenimage decomposition.
}
\examples{
 
x1 <- eigenFilter(x, eigenvalue = c(1,3))
plot(x)
plot(x1)

}
\references{
\itemize{
  \item{Textbook: Sacchi (2002) Statistical and Transform Methods
        in Geophysical Signal Processing}
        }
}
