% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenimages.R
\name{eigenimages}
\alias{eigenimages}
\title{Eigenimage}
\usage{
\S4method{eigenimages}{GPR}(x, center = TRUE, scale = FALSE, track = TRUE)
}
\arguments{
\item{x}{An object of the class GPR}

\item{center}{Logical or numeric. If \code{TRUE}, centering is done by 
subtracting the layer means (omitting \code{NA}'s), and 
if \code{FALSE}, no centering is done. If center is a 
numeric vector with length equal to the 
\code{nlayers(x)}, then each layer of \code{x} has the 
corresponding value from center subtracted from it.}

\item{scale}{Logical or numeric. If \code{TRUE}, scaling is done by 
dividing the (centered) layers of \code{x} by their 
standard deviations if center is \code{TRUE}, and the 
root mean square otherwise. 
If scale is \code{FALSE}, no scaling is done. If scale is
a numeric vector with length equal to \code{nlayers(x)}, 
each layer of \code{x} is divided by the corresponding 
value. Scaling is done after centering.}
}
\value{
An object of the class GPR.
}
\description{
Decompose the GPR data (radargram) using singular value decomposition (SVD) 
and return the reconstructed data for the selected singular values 
(eigenvalues). This method is sometimes refered to in the litterature as 
the Karhunen-Loeve (KL) transformation or the eigenimage decomposition.
}
\examples{
 
data(frenkeLine00)
x <- frenkeLine00
x1 <- eigenimage(x)
plot(x)
plot(x1[ , , 1])

}
\references{
\itemize{
  \item{Textbook: Sacchi (2002) Statistical and Transform Methods
        in Geophysical Signal Processing}
        }
}
