% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{envelope}
\alias{envelope}
\alias{plotEnvelope}
\title{Amplitude envelope}
\usage{
\S4method{envelope}{GPR}(x, npad = 100)

\S4method{plotEnvelope}{GPR}(x, npad = 100, FUN = mean, add = FALSE,
  all = FALSE, plotLog = TRUE, ...)
}
\arguments{
\item{x}{An object of the class GPR.}

\item{npad}{Integer. Padding to avoid Gibbs effect at the begining and
end of the data because of the Hilbert transform.}

\item{FUN}{DEPRECATED. A function to be applied on each row of the GPR data 
to estimate the wave amplitude as a function of time/depth.}

\item{add}{A length-one boolean vector. If TRUE the amplitude is plotted
on the previous plot. If FALSE (default) a new plot is created.}

\item{all}{A length-one boolean vector. If TRUE all trace envelope are 
plotted, if FALSE only the mean.}

\item{plotLog}{A length-one boolean vector. If TRUE the logarithm of the 
amplitude of every trace is ploted on the estimate amplitude.
Default is FALSE.}

\item{x}{An object of the class GPR.}

\item{npad}{Integer. Padding to avoid Gibbs effect at the begining and
end of the data because of the Hilbert transform.}

\item{FUN}{DEPRECATED. A function to be applied on each row of the GPR data 
to estimate the wave amplitude as a function of time/depth.}
}
\description{
Estimate for each trace the amplitude envelope with the Hilbert 
transform (instataneous amplitude).

Plot the amplitude envelope of each trace. See \code{envelope}.
}
\examples{
data(frenkeLine00)
plotAmpl(frenkeLine00, FUN = median)
}
