% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{exportCoord}
\alias{exportCoord}
\title{Export the trace coordinates.}
\usage{
\S4method{exportCoord}{GPR}(x, type = c("SpatialPoints", "SpatialLines",
  "ASCII"), fPath = NULL, folder = NULL, sep = "\\t",
  driver = "ESRI Shapefile", ...)
}
\description{
Export the trace coordinates.
}

