% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{exportPDF}
\alias{exportPDF}
\title{Export a PDF showing the GPR profile.}
\usage{
\S4method{exportPDF}{GPR}(x, fPath = NULL, addTopo = FALSE,
  clip = NULL, normalize = NULL, nupspl = NULL, type = "wiggles",
  ...)
}
\description{
Export a PDF showing the GPR profile.
}
