% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPNG.R
\name{exportPNG,GPR-method}
\alias{exportPNG,GPR-method}
\title{Export GPR data as PNG file}
\usage{
\S4method{exportPNG}{GPR}(x, target = raw(), markers = NULL, ...)
}
\arguments{
\item{x}{[\code{GPR}]}

\item{target}{Either name of the file to write, a binary connection or a 
raw vector (see \code{\link[png]{writePNG}})}

\item{markers}{[\code{integer(n)}] Pixel index set to white at marker
position (e.g., \code{markers = 1:5} to
set the first 5 pixels white at marker
position).}

\item{...}{See additional arguments to be passed to
\code{\link[png]{writePNG}})}
}
\description{
Export GPR data as black-and-white PNG file with pixel dimension equal to 
the sample dimension of the GPR data. There is the possibility to add
white pixels to identify the fiducial markers on the PNG file.
}
