% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{fid}
\alias{fid}
\alias{fid<-}
\alias{fid<-}
\alias{fid}
\title{Fiducial markers of the GPR data}
\usage{
\S4method{fid}{GPR}(x) <- value

\S4method{fid}{GPR}(x)
}
\description{
Fiducial markers of the GPR data
}
