% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/getterSetter.R
\docType{methods}
\name{filepath}
\alias{filepath}
\alias{filepath<-}
\alias{filepath,GPR-method}
\alias{filepath<-,GPR-method}
\title{Filepath of the GPR data}
\usage{
\S4method{filepath}{GPR}(x)

\S4method{filepath}{GPR}(x) <- value
}
\description{
Filepath of the GPR data
}
