% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/ClassGPRvirtual.R
\docType{methods}
\name{filepath}
\alias{filepath}
\alias{filepath<-}
\alias{filepath}
\alias{filepath<-}
\title{Filepath of the GPR data}
\usage{
\S4method{filepath}{GPRvirtual}(x)

\S4method{filepath}{GPRvirtual}(x) <- value
}
\description{
Filepath of the GPR data
}
