% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClosestCoord.R
\name{findClosestCoord}
\alias{findClosestCoord}
\title{Closest trace}
\usage{
findClosestCoord(x, y)
}
\arguments{
\item{x}{Object of the class GPR.}

\item{y}{Length-two numeric vector of the (x, y)-coordinates.}
}
\value{
Indice (integer) of the closest trace.
}
\description{
Return the indice of the closest trace to the point \code{y}
}
\concept{spatial computing}
