% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{firstBreak}
\alias{firstBreak}
\title{Time of first wave break}
\usage{
\S4method{firstBreak}{GPR}(x, method = c("coppens", "threshold", "MER"),
  thr = 0.12, w = 11, ns = NULL, bet = NULL)
}
\arguments{
\item{x}{[\code{GPR class}]\cr An object of the class \code{GPR}}

\item{method}{[\code{character(1)}]\cr Method to be applied (either
\code{coppens}, \code{threshold} or \code{MER}). 
\code{"coppens"} corresponds to the modified Coppens method, 
\code{"threshold"} to the threshold method, 
and \code{"MER"} to the modified energy ratio method.}

\item{thr}{[\code{numeric(1)}]\cr Threshold for the signal 
amplitude (in \%) at which time zero is picked (only for the
threshold method). \code{thr} ranges between 0 and 1.}

\item{w}{[\code{numeric(1)}]\cr Length of the leading window in unit of time
(only for the modified Coppens and modified energy ratio 
methods). Recommended value: about one period of the first-arrival 
waveform.}

\item{ns}{[\code{numeric(1)}]\cr Length of the edge preserving smoothing 
window in unit of time (only for the modified Coppens 
method). Recommended value: between one and two signal periods.
When \code{ns = NULL} the value of \code{ns} is set to 
\code{1.5 * w}.}

\item{bet}{[\code{numeric(1)}]\cr Stabilisation constant (only for the 
modified Coppens method). Not critical. 
When \code{bet = NULL} the value of \code{bet} is set to 
20\% of the maximal signal amplitude.}
}
\value{
[\code{numeric(n)}]\cr The time of the first wave break for every
        traces in unit of time (\code{n = ncol(x) =} number of traces).
}
\description{
Pick the time corresponding to the first break of each trace in the GPR profile.
Return a vector containing the first break times.
}
\references{
\describe{
  \item{Modified Coppens method}{Sabbione J.I. and Velis D. (2010) 
       Automatic first-breaks picking: New strategies and algorithms. 
       Geophysics, 75(4): 67-76.}
  \item{Modified Energy Ratio (MER) method}{Han L., Wong J., and John C. 
       (2010) Time picking on noisy microseismograms. In: Proceedings of the
       GeoCanada 2010 Convention - Working with the Earth, Calgary, AB, 
       Canada, p. 4}
}

#' @examples 
data("frenkeLine00")
fb <- firstbreak(frenkeLine00, w = 10)
plot(seq_along(frenkeLine00), fb)
}
\seealso{
\code{\link{firstBreakToTime0}} to convert time of first wave break
         into time-zero; 
         \code{\link{time0}} and \code{\link{setTime0}} to set time-zero;
         \code{\link{estimateTime0}} to estimate first wave break, convert
         it to time-zero and set time zero (all in one step);
         \code{\link{time0Cor}} to shift the traces such that they start
         at time-zero.
}
