% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstBreakToTime0.R
\name{firstBreakToTime0}
\alias{firstBreakToTime0}
\alias{firstBreakToTime0,GPR-method}
\title{Convert first wave break time to time-zero}
\usage{
firstBreakToTime0(x, fb, c0 = 0.299)

\S4method{firstBreakToTime0}{GPR}(x, fb, c0 = 0.299)
}
\arguments{
\item{x}{(\code{GPR})}

\item{fb}{(\code{numeric[1|n]}) Time of first break (\code{n = ncol(x)})}

\item{c0}{(\code{numeric[1]})  Propagation speed of the GPR wave
through air in unit of space per unit of time
(generally in m/ns).}
}
\description{
Account for the delay time between time of wave emission and time of first
wave break recording due to the antenna separation (offset).
}
