% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{fkFilter}
\alias{fkFilter}
\title{Frequency-wavenumber filter}
\usage{
\S4method{fkFilter}{GPR}(x, fk = NULL, L = c(5, 5), npad = 1)
}
\description{
Frequency-wavenumber filter
}
