% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frenkeCMP}
\alias{frenkeCMP}
\title{Ground-penetrating radar data: CMP.}
\format{An object of the class RGPR}
\source{
University of Basel (Switzerland)
}
\usage{
frenkeCMP
}
\description{
Surface ground-penetrating radar data recorded
the 25 April 2014 in Frenkental (Swizerland).
Coordinates: CH1903+/LV95:2'622'209.66, 1'256'907.54
Elevation: 345.8 m
}
\keyword{datasets}
