% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frenkeTopo00}
\alias{frenkeTopo00}
\title{Trace coordinates of \code{frenkeLine00}.}
\format{\code{data.frame}, three columns for the East, North and vertical
coordinates}
\source{
University of Basel (Switzerland)
}
\usage{
frenkeTopo00
}
\description{
Topographic survey recorded
the 25 April 2014 in Frenkental (Swizerland).
Coordinates: CH1903+/LV95:2'622'209.66, 1'256'907.54
Elevation: 345.8 m
}
\keyword{datasets}
