% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqSpectrum.R
\name{freqSpectrum}
\alias{freqSpectrum}
\alias{freqSpectrum,GPR-method}
\title{Frequency spectrum the traces (1D)}
\usage{
freqSpectrum(obj, plotSpec = TRUE, unwrapPhase = TRUE, ...)

\S4method{freqSpectrum}{GPR}(obj, plotSpec = TRUE, unwrapPhase = TRUE, ...)
}
\arguments{
\item{obj}{(\code{GPR* object})}

\item{plotSpec}{(\code{logical[1]}) Plot spectrum?}

\item{unwrapPhase}{(\code{logical[1]}) Unwrap phase?}

\item{...}{Additional parameters}
}
\value{
(\code{GPRset}) Sets = Amplitude and Phase as a function of frequency
}
\description{
Based on the function \code{fft()} of R base
}
