% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqSpectrum2D.R
\name{freqSpectrum2D}
\alias{freqSpectrum2D}
\alias{freqSpectrum2D,GPR-method}
\title{Wave-number frequency spectrum (2D)}
\usage{
freqSpectrum2D(obj, plotSpec = TRUE, unwrapPhase = TRUE, ...)

\S4method{freqSpectrum2D}{GPR}(obj, plotSpec = TRUE, unwrapPhase = TRUE, ...)
}
\description{
Based on the function \code{fft()} of R base
}
