% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainAGC.R
\name{gainAGC}
\alias{gainAGC}
\alias{gainAGC,GPR-method}
\title{Classical Automatic Gain Control (AGC)}
\usage{
gainAGC(
  obj,
  sig = 10,
  p = 2,
  r = 0.5,
  floorquantile = 0.05,
  return_gain = FALSE,
  track = TRUE
)

\S4method{gainAGC}{GPR}(
  obj,
  sig = 10,
  p = 2,
  r = 0.5,
  floorquantile = 0.05,
  return_gain = FALSE,
  track = TRUE
)
}
\arguments{
\item{obj}{(\code{GPR* object}) An object of the class GPR.}

\item{sig}{(\code{numeric[1]}) standard deviation of the kernel.}

\item{p}{(\code{numeric[1]}) Parameter of the power filter
(\code{p} \eqn{\geq} 0). Usually, \code{p} = 2.}

\item{r}{(\code{numeric[1]}) Root applied to the smoothed power.
Usually, \code{r} = 0.5.}

\item{floorquantile}{(\code{numeric[1]}) Quantile of local gain used as minimum
threshold to prevent extreme amplification
(default = 0.05, value between 0 and 1.).}

\item{return_gain}{(\code{logical[1]}) Should the gain be returned (instead of the gained object)?}

\item{track}{(\code{logical[1]}) Should the processing step be tracked?}
}
\value{
\verb{GPR class} An object of the class GPR.
}
\description{
\code{gainAGC} applies an AGC (Automatic Gain Control) gain to
compensate for amplitude decay and enhances visibility of weaker signals.
The function uses a Gaussian-weighted local RMS or power measure, with
local mean removal. A gain floor is applied to avoid extreme
amplification in regions of near-zero signal.
}
\details{
The trace signal is smoothed with a Gaussian filter. The smoothed trace
is substracted from the original trace, raised to power \code{p}, smoothed
by a Gaussian filter (to obtain a local weighted sum) and raised to
power \code{r} to obtain the gain.
Typical values are \code{p} = 2 and \code{r} = 0.5 which will
make gain equal to the local RMS.  The abs() function is used to allow
for arbitrary 'p' and 'r'.

Modified slots
\code{data}: trace gained.
\code{proc}: updated with function name and arguments.
}
