% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainEnv.R
\name{gainEnv}
\alias{gainEnv}
\alias{gainEnv,GPR-method}
\title{Gain Compensation Based on Envelope}
\usage{
gainEnv(
  obj,
  FUN = mean,
  floorquantile = 0.05,
  return_gain = FALSE,
  track = TRUE
)

\S4method{gainEnv}{GPR}(
  obj,
  FUN = mean,
  floorquantile = 0.05,
  return_gain = FALSE,
  track = TRUE
)
}
\arguments{
\item{obj}{(\code{GPR})] A GPR object containing traces to be gain-compensated.}

\item{FUN}{(\code{function}) A summary function applied to the envelope of each time sample across traces.
Defaults to \code{mean}. Other options could be \code{median}, \code{max}, etc.}

\item{floorquantile}{(\code{numeric[1]}) Quantile of local gain used as minimum
threshold to prevent extreme amplification
(default = 0.05, value between 0 and 1.).}

\item{return_gain}{(\code{logical[1]}) Should the gain be returned (instead of the gained object)?}

\item{track}{(\code{logical[1]}) Should the processing step be tracked?}
}
\value{
A GPR object with amplitude gain applied according to the envelope-based compensation.
}
\description{
Applies a gain to a GPR object based on the envelope of the signal.
The gain is computed from a summary statistic (e.g., mean) of the
envelope across traces and normalized to compensate for amplitude decay.
}
\details{
The function computes the envelope of each trace (using \code{amplEnv}) and summarizes it across
traces using the specified \code{FUN}. The resulting vector is normalized to its maximum value
to produce a gain vector. The gain is set to 1 for all samples before the peak of the normalized
envelope to avoid boosting early-time noise. The reciprocal of the gain is applied to the original
GPR data.

Mathematically:
\deqn{
g_0(t) = \frac{\text{FUN}(\text{Envelope}(obj))}{\max(\text{FUN}(\text{Envelope}(obj)))}, \quad
g(t) = 1/g_0(t)
}
where \(g_0(t) = 1\) for times before the peak of the envelope.
}
