% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainSEC.R
\docType{methods}
\name{getGainSEC}
\alias{getGainSEC}
\alias{gainSEC}
\title{Spreading and Exponential Compensation (SEC) gain}
\usage{
\S4method{getGainSEC}{GPR}(x, a = 0.01, b = 1, t0 = NULL,
  tend = NULL, tcst = NULL, track = TRUE)

\S4method{gainSEC}{GPR}(x, a = 0.01, b = 1, t0 = NULL, tend = NULL,
  tcst = NULL, return_gain = FALSE, track = TRUE)
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class GPR.}

\item{a}{[\code{numeric(1)}] Parameter of the exponential filter
(\code{a} \eqn{\geq} 0).}

\item{b}{[\code{numeric(1)}] Parameter of the power filter
(\code{b} \eqn{\geq} 0). Usually, \code{b = 1}.}

\item{t0}{[\code{numeric}] Start time of the gain filter
(if \code{t0 = NULL}, \code{t0} is set equal to \code{time0(x)}).}

\item{tend}{[\code{numeric(1)}] End time of the gain filter (optional)}

\item{tcst}{[\code{numeric(1)}] Constant time: the gain before 
\code{tcst} is set equal to the gain value at \code{tcst}.}
}
\value{
[\code{GPR class}] An object of the class GPR.
}
\description{
\code{getGainSEC} is deprecated.

\code{gainSEC} Applies a combination of a power and exponential time gain to compensate for 
the signal attenuation through spherical spreading losses and  
exponential ohmic dissipation of energy with depth. Usually, the power in
the power gain is set to zero.
}
\details{
Spreading and Exponential Compensation (SEC) gain can be written as 
\eqn{\exp(a \cdot t) \cdot t^b}, where \eqn{t^b} is the power gain
(set \eqn{b = 1} to get a linear gain) and \eqn{\exp(a \cdot t)} is the
exponential gain.

Modified slots
\itemize{
  \item \code{data}: trace gained.
  \item \code{proc}: updated with function name and arguments.
}
}
\seealso{
\code{\link{gainAGC}}
}
