% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{georef}
\alias{georef}
\title{Georeferencing}
\usage{
georef(x, alpha, center = c(0, 0), center2 = NULL)
}
\arguments{
\item{x}{A matrix with the first two columns corresponding
to the coordinates.}

\item{alpha}{A length-one numeric vector corresponding to 
the rotation angle in radians}

\item{center}{A length-two numeric vector.}

\item{center2}{A length-two numeric vector or 
\code{NULL} (default).}
}
\description{
Perform on a set of x,y coordinates
(1) a translation by \code{-center}, then
(2) a rotation by \code{alpha} (radian), and (3)
a translation by \code{center2}. If \code{center2}
is \code{NULL}, then \code{center2} is set equal
to \code{center}.
}

