% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDepth.R
\name{getDepth}
\alias{getDepth}
\alias{getDepth,GPR-method}
\title{Get depth}
\usage{
getDepth(obj)

\S4method{getDepth}{GPR}(obj)
}
\arguments{
\item{obj}{(\code{GPR* object}) An object of the class \code{GPR}}
}
\value{
(\code{GPR* object}) with signal as a function of depth.
}
\description{
Retunrs an object identical to \code{obj} but with values equal to the signal
depth
}
\concept{signal processing}
