% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS.R
\name{getLonLatFromGPGGA}
\alias{getLonLatFromGPGGA}
\title{Get longitude and latitude from GPGGA sentence information (NMEA)}
\usage{
getLonLatFromGPGGA(a)
}
\arguments{
\item{a}{[\code{data.frame}] GPGGA sentence information}
}
\value{
[\code{data.frame}] Columns = latitude, longitude, elevation and
        time
}
\description{
Get longitude and latitude from GPGGA sentence information (NMEA)
}
