% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS.R
\name{getUTMzone}
\alias{getUTMzone}
\title{Get UTM zone from lattidue and longitude}
\usage{
getUTMzone(lat, lon)
}
\arguments{
\item{lat}{[\code{numeric}] Latitude}

\item{lon}{[\code{numeric}] Longitude.}
}
\value{
[\code{integer(1)}] The UTM zone.
}
\description{
Get UTM zone from lattidue and longitude
}
