% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVel.R
\name{getVel}
\alias{getVel}
\alias{getVel,GPR-method}
\title{Get velocity model as GPR object}
\usage{
getVel(obj, type = c("vint", "vrms"))

\S4method{getVel}{GPR}(obj, type = c("vint", "vrms"))
}
\arguments{
\item{obj}{(\verb{GPR class}) An object of the class \code{GPR}}

\item{type}{(\code{vrm|vint}) Set the velocity you want (either root-mean square or internal velocity)}
}
\value{
(\verb{GPR class}) An object of the class \code{GPR} containing the velocity model.
}
\description{
Return the velocity model (either the root-mean square or internal velocity).
}
\concept{velocity model}
