% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVel2.R
\name{getVel2}
\alias{getVel2}
\alias{getVel2,GPR-method}
\title{Get velocity model}
\usage{
getVel2(x, type = c("vrms", "vint", "v"), strict = FALSE)

\S4method{getVel2}{GPR}(x, type = c("vrms", "vint", "v"), strict = FALSE)
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}

\item{type}{[\code{vrm|vint}] Set the velocity you want (either root-mean square or internal velocity)}
}
\value{
[\code{GPR class}] An object of the class \code{GPR} containing the velocity model.
}
\description{
Return the velocity model (either the root-mean square or internal velocity).
}
