% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotGPR.R
\name{ggplotGPR}
\alias{ggplotGPR}
\title{GG-plot of GPR data}
\usage{
ggplotGPR(x, col = palGPR())
}
\arguments{
\item{x}{GPR object}

\item{col}{Color palette}
}
\description{
Grammar of graphic (GG) plot of GPR data.
}
