% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{GPR-subset}
\alias{GPR-subset}
\alias{[<-}
\title{extract parts of GPR}
\usage{
\S4method{[}{GPR,ANY,ANY,ANY}(x, i, j, drop)

\S4method{[}{GPR,ANY,ANY,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{Object of class GPR}

\item{i}{integer}

\item{j}{integer}
}
\description{
Object of the class GPR can be manipulated as matrix
}
