% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{as.matrix}
\alias{as.GPR.list}
\alias{as.GPR.matrix}
\alias{as.SpatialLines}
\alias{as.SpatialPoints}
\alias{as.matrix}
\alias{as.numeric}
\alias{as.vector}
\title{Coercion to matrix}
\usage{
\S4method{as.matrix}{GPR}(x)

\S4method{as.vector}{GPR}(x, mode = "any")

as.SpatialLines(x, ...)

as.SpatialPoints(x, ...)

\S4method{as.numeric}{GPR}(x, ...)

as.GPR.matrix(x, ...)

as.GPR.list(x, ...)
}
\description{
Coercion to matrix

Coercion to vector

Coercion to SpatialLines

Coercion to SpatialPoints

Coercion to numeric

Coercion from matrix to GPR

Coercion from list to GPR
}

