% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{as.SpatialLines,GPR-method}
\alias{as.SpatialLines,GPR-method}
\alias{as.SpatialPoints,GPR-method}
\title{Coercion function from \code{GPR} to \code{SpatialLines} and 
\code{SpatialPoints}}
\usage{
\S4method{as.SpatialLines}{GPR}(x)

\S4method{as.SpatialPoints}{GPR}(x)
}
\description{
Coerce the trace positions into an object of the class \code{SpatialPoints}
or \code{SpatialPoints}.
}
\examples{
data(frenkeLine00)
x <- frenkeLine00
coord(x) <- cbind(seq_len(ncol(x)),
                  seq_len(ncol(x)),
                  rnorm(ncol(x)))
y <- as.SpatialLines(x)
y <- as.SpatialPoints(x)
}
