% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRcube.R
\docType{methods}
\name{GPRcube-subset}
\alias{GPRcube-subset}
\title{extract parts of GPRsurvey}
\usage{
\S4method{[}{GPRcube,ANY,ANY,ANY}(x, i, j, k, drop = TRUE)
}
\description{
Return an object of class GPR slice
}
