% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRset.R
\docType{methods}
\name{GPRset-subset}
\alias{GPRset-subset}
\title{extract parts of GPRset}
\usage{
\S4method{[}{GPRset,ANY,ANY,ANY}(x, i, j, k, drop = TRUE)
}
\description{
Return an object of class GPR slice
}
