% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRsurvey.R
\docType{methods}
\name{GPRsurvey-subset}
\alias{GPRsurvey-subset}
\title{extract parts of GPRsurvey}
\usage{
\S4method{[}{GPRsurvey,ANY,ANY,ANY}(x, i, j, drop)
}
\description{
Return an object of class GPRsurvey
}
