% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRsurvey.R
\docType{methods}
\name{[[}
\alias{[[}
\alias{[[<-,GPRsurvey,ANY,ANY-method}
\title{extract a GPR object from a GPRsurvey object}
\usage{
\S4method{[[}{GPRsurvey,ANY,ANY}(x, i, j, ...)

\S4method{[[}{GPRsurvey,ANY,ANY}(x, i) <- value
}
\description{
Return an object of class GPR
}
