% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRvirtual.R, R/z_string.R
\docType{class}
\name{GPRvirtual-class}
\alias{GPRvirtual-class}
\alias{extractPattern}
\title{Class GPR}
\usage{
extractPattern(x, pattern, start = 0, stop = 0)
}
\arguments{
\item{x}{[\code{character}] A character vector where matches are sought, 
or an object which can be coerced by \code{\link{as.character}} 
to a character vector.}

\item{pattern}{[\code{character}] String containing a regular expression 
to be matched (see \code{\link{grep}}).}

\item{start}{[\code{integer}] The first element of the matched string 
to be extracted (see \code{\link{substr}}).}

\item{stop}{[\code{integer}] The last element of the matched string 
to be extracted (see \code{\link{substr}}).}
}
\value{
[\code{character}] The matched string
}
\description{
A virtual S4 class to represent a ground-penetrating radar (GPR) data.

Get the match in a string given a regular expression and extract 
substrings.
}
\section{Slots}{

\describe{
\item{\code{version}}{[\code{character(1)}] Version of RGPR.}

\item{\code{name}}{[\code{character(1)}] Name of the GPR data.}

\item{\code{path}}{[\code{character(1)}] File path of the original GPR data.}

\item{\code{desc}}{[\code{character(1)}] Description of the GPR data.}

\item{\code{mode}}{[\code{character(1)}] Survey mode 
(e.g., \code{"CO"}, \code{"CMP"}).}

\item{\code{date}}{[\code{Date(1)}] Date of the survey (class \code{Date},
e.g., \code{Sys.Date()}).}

\item{\code{freq}}{[\code{numeric(1)}] GPR antennae frequency (in MHz).}

\item{\code{data}}{[\code{array}] GPR data.}

\item{\code{dunit}}{[\code{character(1)}] GPR data unit.}

\item{\code{dlab}}{[\code{character(1)}] GPR data label.}

\item{\code{crs}}{[\code{character(1)}] Coordinate reference system following the 
R notation of proj4string from the PROJ.4 library.}

\item{\code{spunit}}{[\code{character(1)}] Spatial units.}

\item{\code{xunit}}{[\code{character(1)}] Unit of \code{x}.}

\item{\code{xlab}}{[\code{character(1)}] Label of \code{x}.}

\item{\code{zunit}}{[\code{character(1)}] Unit of \code{z}.}

\item{\code{zlab}}{[\code{character(1)}] Label of \code{z}.}

\item{\code{vel}}{[\code{list}] Velocity model.}

\item{\code{proc}}{[\code{list}] Each element corresponds
to a processing step applied to the data.}

\item{\code{delineations}}{[\code{list}] Hand-delineated structures.}

\item{\code{md}}{[\code{list}] Additional meta data from GPR file.}
}}

\section{\code{mode}}{
 survey mode, difference between reflection and CMP and WARR
\describe{
  \item{CO}{Common-offset: The trace positions increase from trace to trace. The
                    antenna separation distance is constant.
                    Amplitude = f(depth, position)}
  \item{CMP}{The trace positions, \code{x@pos}, are constant and equal to 
             zero. The antenna separation distance increases increase from
             trace to trace.
             Amplitude = f(depth, antsep)}
  \item{WARR}{The trace positions, \code{x@pos}, are not constant because
             while the signal is emitted from the same position, the signal
             the signal is recorded from different positions. 
             The antenna separation distance increases increase from
             trace to trace.
             Is "CMP" a special case of "WARR"?
             Amplitude = f(depth, antsep)}
  \item{CMPAnalysis}{No trace positions, \code{x@pos = numeric(0)},
                     no antenna separation, \code{x@antsep = numeric(0)},
                     coherence/semblance = f(depth, velocity)}
}
}

\section{\code{data}}{

\describe{
  \item{class \code{GPR}}{TA \eqn{m \times n} numeric matrix consiting of a 
           cross-section of signal amplitudes as a function of the GPR 
           position. The columns of \code{data} correspond to the GPR 
           traces and the row of \code{data} to the time/depth samples.}
  \item{class \code{GPRset}}{3D data}
  \item{class \code{GPRcube}}{3D data}
}
}

\seealso{
\code{\link{substr}} and \code{\link{regexpr}}
}
