% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperbola.R
\name{hyperbolaFit}
\alias{hyperbolaFit}
\title{Fit hyperbola to data points}
\usage{
hyperbolaFit(x, y = NULL)
}
\arguments{
\item{x}{[\code{numeric}|\code{list}] Either the x-values of the data
points or a list with elements x and t (in this case, leave 
\code{y = NULL})}

\item{y}{[\code{numeric}] The time values of the data (if \code{x} is not
a list).}
}
\value{
[\code{list}] A list with class \code{hyperbola} containing the 
following elements:
\describe{
  \item{reg}{The regression output of the function \code{lm}}
  \item{vrms}{The estimated root-mean-square velocity}
  \item{x0}{The horizontal position of the hyperbola vertex}
  \item{t0}{The vertical position  (time) of the hyperbola vertex}
  \item{x}{The input point positions x}
  \item{y}{The input point positions y}
}
}
\description{
Fit hyperbola to data points \eqn{\mathbf{x}, \mathbf{t}}, where
\eqn{\mathbf{x} = (x_1, x_2, \ldots)} is the horizontal point position
and \eqn{\mathbf{t} = (t_1, t_2, \ldots)} is the vertical (time) point 
position.
The parameters \eqn{a} and \eqn{b} in 
\eqn{\frac{\mathbf{t}^2}{4} = a \mathbf{x}^2 + b \mathbf{x} + c} are 
estimated with the \code{lm} function.
The position of the vertex is given by:
\deqn{x_0 = -\frac{b}{2} v_{RMS}}
and
\deqn{t_0 = 2 \sqrt{c - (\frac{x_0}{v_{RMS}})^2}}
where \eqn{v_{RMS} = \sqrt{\frac{1}{a}}}
Note that the antenna separation distance is not accounted for.
}
\examples{
data("frenkeLine00")
x <- frenkeLine00
x <- estimateTime0(x, w = 5, method = "MER", FUN = mean)
x <- time0Cor(x, method = "pchip")
x <- gainAGC(fFilter(x, f = c(180, 250), type = "low"), w = 20)
plot(x)

# xy <- locator(type = "l")
xy <- list( x = c( 11.8,  15.0,  17.7, 20.3, 24.4,  27.4,  30.9,  35.2),
            y = c(142.2, 119.8, 107.7, 99.5, 97.5, 105.6, 120.9, 138.1))
hyp <- hyperbolaFit(xy)
hyperbolaPlot(hyp, x = seq(5, 50, by = 0.01), col = "green", 
              lwd = 4, ann = TRUE)
points(xy, pch = 20, col = "blue")
}
\seealso{
\code{\link{hyperbolaPlot}}, \code{\link{hyperbolaSim}}
}
