% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperbolaFromVrms.R
\name{hyperbolaFromVrms}
\alias{hyperbolaFromVrms}
\title{Returns hyperbola associated with Vrms velociies}
\usage{
hyperbolaFromVrms(x)
}
\description{
Returns hyperbola associated with Vrms velociies
}
