% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperbola.R
\name{hyperbolaPlot}
\alias{hyperbolaPlot}
\title{Plot a hyperbola}
\usage{
hyperbolaPlot(
  hyp,
  x = NULL,
  add = TRUE,
  ann = FALSE,
  ann.pos = 3,
  ann.offset = 0.5,
  ann.col = "green",
  ann.font = 2,
  n = 100,
  ...
)
}
\arguments{
\item{hyp}{[\code{list}|\code{class hyperbola}] 
  Either a list with elements \code{x0}, \code{t0}, and \code{vrms}
  corresponding to the coordinates of the hyperbola vertex and
  the root-mean-square velocity, or a list of class 
  \code{hyperbola} (i.e., the output of the function 
  \code{\link{hyperbolaFit}}). 
\code{y = NULL})}

\item{x}{[\code{numeric}] 
The horitzontal positions at which to plot the hyperbola (optional,
see details).}

\item{add}{[\code{logical(1)}]
If \code{TRUE} (defaults), add the hyperbola to current 
plot. If \code{FALSE}, display the hyperbola in a new plot.
In this case you must specify \code{x}.}

\item{ann}{[\code{logical(1)}]
If \code{TRUE}, add the following annotation above the 
hyperbola vertex: root-mean-square velocity and vertex depth.}

\item{ann.pos}{[\code{numeric(1)}]
A position specifier for the annotation text. 
See argument \code{pos}
of the function \code{\link[graphics]{text}}.}

\item{ann.offset}{[\code{numeric(1)}]
when \code{pos} is specified, this value controls the 
distance ('offset') of the annotation text label from the 
specified coordinate in fractions of a character width. 
See argument \code{offset}
of the function \code{\link[graphics]{text}}.}

\item{ann.col}{[\code{character(1)}]
Color of the annotation text. 
See argument \code{col}
of the function \code{\link[graphics]{text}}.}

\item{ann.font}{[\code{integer(1)}]
Number of points used to plot the hyperbola when
\code{x} is not specified.}

\item{...}{Additional arguments passed to 
\code{\link[graphics]{lines}} (if \code{add = TRUE}) or to
\code{\link[graphics]{plot}} (if \code{add = FALSE})    .}
}
\description{
Plot a hyperbola
}
\details{
If \code{x = NULL} and \code{add = TRUE}, the hyperbola is per default
displayed over the entire plot. You can define the extent of the hyperbola
by setting \code{xlim} as in the function \code{\link[graphics]{plot}}.
If \code{add = FALSE} \code{x} must be defined otherwise
an error is raised.
}
\examples{
data("frenkeLine00")
x <- frenkeLine00
x <- estimateTime0(x, w = 5, method = "MER", FUN = mean)
x <- time0Cor(x, method = "pchip")
x <- gainAGC(fFilter(x, f = c(180, 250), type = "low"), w = 20)
plot(x)

# xy <- locator(type = "l")
xy <- list( x = c( 11.8,  15.0,  17.7, 20.3, 24.4,  27.4,  30.9,  35.2),
            y = c(142.2, 119.8, 107.7, 99.5, 97.5, 105.6, 120.9, 138.1))
hyp <- hyperbolaFit(xy)
hyperbolaPlot(hyp, x = seq(5, 50, by = 0.01), col = "green", 
              lwd = 4, ann = TRUE)
points(xy, pch = 20, col = "blue")

plot(x)
hyp2 <- list(x0 = hyp$x0, t0 = hyp$t0, vrms = hyp$vrms)
hyperbolaPlot(hyp2, col = "blue", lwd = 2, ann = TRUE, xlim = c(10, 40))
points(hyp$x0, hyp$t0, pch = 20, col = "red", cex = 1.3)

# with 'add = FALSE'
plot(x)
hyperbolaPlot(hyp, x = seq(10, 40, by = 0.1), col = "red", 
lwd = 2,  add = FALSE, type = "l")
hyperbolaPlot(hyp, x = seq(10, 40, by = 0.1), col = "red", 
             lwd = 2,  add = FALSE, type = "l", ylim = c(175, 95))

}
\seealso{
\code{\link{hyperbolaFit}}, \code{\link{hyperbolaSim}}
}
