% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect GPR data plot}
\usage{
\S4method{inspect}{GPR}(x, ...)
}
\description{
Reads the position of the graphics cursor when the (first) mouse button is 
pressed and return position as well as amplitude.
}
