% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complextrace.R
\name{instAmpl}
\alias{instAmpl}
\alias{instAmpl,GPR-method}
\title{Instantaneous amplitude}
\usage{
instAmpl(x, npad = 100)

\S4method{instAmpl}{GPR}(x, npad = 100)
}
\description{
Instantaneous amplitude
}
