% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complextrace.R
\name{instPhase}
\alias{instPhase}
\alias{instPhase,GPR-method}
\title{Instantaneous phase}
\usage{
instPhase(x, npad = 100, unwrapPhase = TRUE)

\S4method{instPhase}{GPR}(x, npad = 100, unwrapPhase = TRUE)
}
\description{
Instantaneous phase
}
