% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_con.R
\name{int32touint32}
\alias{int32touint32}
\title{Fix a 32 bit unsigned integer that has been read as signed}
\usage{
int32touint32(x, nbits = 32)
}
\description{
Fix a 32 bit unsigned integer that has been read as signed
}
