% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp2DDelineations.R
\docType{methods}
\name{interp2DDelineations}
\alias{interp2DDelineations}
\alias{interp2DDelineations,GPRsurvey-method}
\title{2D interpolation of the delineations and surface topography}
\usage{
interp2DDelineations(x, dx = NULL, dy = NULL, h = 6,
  extend = c("bbox", "chull", "buffer"), buffer = NULL, shp = NULL)

\S4method{interp2DDelineations}{GPRsurvey}(x, dx = NULL, dy = NULL,
  h = 6, extend = c("bbox", "chull", "buffer"), buffer = NULL,
  shp = NULL)
}
\arguments{
\item{x}{[\code{GPRsurvey class}] An object of the class \code{GPRsurvey}}

\item{dx}{[\code{numeric(1)}] x-resolution}

\item{dy}{[\code{numeric(1)}] y-resolution}

\item{h}{[\code{integer(1)}] Number of levels in the hierarchical 
construction (see \code{\link[MBA]{mba.surf}}).}

\item{extend}{[\code{character(1)}] Extend type for the interpolation based 
on the GPR data (if \code{shp = NULL} or on the object passed
to \code{shp}), see details.}

\item{buffer}{[\code{character(1)}] A buffer can be applied to increase the 
extend by the value defined by \code{buffer}.}

\item{shp}{[\code{matrix|list|sf|sp}] An object to define the interpolation
extend.}
}
\value{
[\code{GPRcube}] A \code{GPRcube} object consisting of \eqn{k + 1} 
        surfaces of dimension \eqn{m \times n}. \eqn{k} is the number of
        delineations. The first surface corresponds to the topography while
        the others to the interpolated delineations.
}
\description{
This function groups the delineations by their indexes or names and 
interpolates each group separately. You need to check, that all your 
delineations are either in the same order (e.g., for every data, your first 
delineated the lake bottom, than the sediment below) or the delineations 
have consistent names across all the data set 
(e.g., "lake_bottom", "lake_sediments"). 
It interpolates also the topography (surface elevation given by the 
coordinates).
}
\details{
\describe{
  \item{\code{bbox}}{axis-aligned bounding box}
  \item{\code{chull}}{convex hull}
  \item{\code{buffer}}{BUffer polygon at a specified distance around 
                       the GPR lines. The argument \code{buffer} must be 
                       specified (i.e., not \code{NULL})}
}
}
