% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateTimeToDepth.R
\name{interpolateTimeToDepth}
\alias{interpolateTimeToDepth}
\title{Time-to-depth conversion based on depth information}
\usage{
interpolateTimeToDepth(
  x,
  z = NULL,
  twt = NULL,
  pos = NULL,
  method = "linear",
  h = 2
)
}
\description{
1D and 2D interpolation
}
