% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R, R/ClassGPRsurvey.R
\docType{methods}
\name{interpPos<-}
\alias{interpPos<-}
\title{Interpolate the trace position.}
\usage{
\S4method{interpPos}{GPR}(x, topo, ...) <- value

\S4method{interpPos}{GPRsurvey}(x, topo, ...) <- value
}
\description{
Interpolate the trace position.
}

