% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\name{interpPosFromGeoJSON}
\alias{interpPosFromGeoJSON}
\title{Interpolate GPR coordinates from geoJSON data}
\usage{
interpPosFromGeoJSON(x, geojson, tol = NULL, backproject = TRUE)
}
\arguments{
\item{x}{Object of the class GPR}

\item{geojson}{Either a geojson string or a filepath pointing to a 
geojson file}
}
\description{
Interpolate GPR coordinates from geoJSON data
}
