% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\name{interpPosFromGPGGA}
\alias{interpPosFromGPGGA}
\title{Interpolate GPR coordinates from GPS data (GPGGA string) data}
\usage{
interpPosFromGPGGA(ntr, GPGGA, tol = NULL, backproject = TRUE)
}
\arguments{
\item{x}{Object of the class GPR}

\item{gpgga}{output of the function "readGPGGA()"}
}
\description{
Interpolate GPR coordinates from GPS data (GPGGA string) data
}
