% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\name{interpPosFromXYZ}
\alias{interpPosFromXYZ}
\title{Interpolate GPR coordinates from xyz coordinates}
\usage{
interpPosFromXYZ(x, xyz, tol = NULL)
}
\arguments{
\item{x}{Object of the class GPR}

\item{xyz}{[\code{matrix(n, 2|3)}] A two-columns (or three-column matrix) 
containing the x and y (or x, y, and z) coordinates.}

\item{tol}{[\numeric(1)] A tolerance value used to remove duplicated
coordinates in \code{xyz}. When left equal to \code{NULL}
\code{tol} is internally set equal to 
\code{sqrt(.Machine$double.eps)}.}
}
\value{
[\code{GPR}] object.
}
\description{
This function interpolate the trace positions from a series of coordinates
assuming no link between the traces and the points (e.g., it is not know
which trace belong to which point). Therefore, it is assumed, that the
trace spacing is constant over the GPR line.
}
