% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{interpRegRaster}
\alias{interpRegRaster}
\title{Interpolate raster to regular spacing}
\usage{
interpRegRaster(
  vx,
  vy,
  z,
  nx,
  ny,
  method = c("linear", "nearest", "pchip", "cubic", "spline"),
  xy = c(TRUE, TRUE)
)
}
\description{
Interpolate raster to regular spacing
}
