% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpVel.R
\name{interpVel}
\alias{interpVel}
\title{Interpolate velocity model}
\usage{
interpVel(
  x,
  type = c("vrms", "vint"),
  method = c("stairs", "linear", "nearest", "pchip", "cubic", "spline")
)
}
\description{
Interpolate velocity model
}
