% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRsurvey.R
\name{intersections}
\alias{intersections}
\alias{intersections,GPRsurvey-method}
\title{Return intersection from GPRsurvey}
\usage{
\S4method{intersections}{GPRsurvey}(x)
}
\description{
Return intersection from GPRsurvey
}
