% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\name{is.finite}
\alias{is.finite}
\alias{is.finite,GPRvirtual-method}
\alias{is.infinite}
\alias{is.infinite,GPRvirtual-method}
\alias{is.nan}
\alias{is.nan,GPRvirtual-method}
\title{Finite, Infinite and NAN Numbers}
\usage{
\S4method{is.finite}{GPRvirtual}(x)

\S4method{is.infinite}{GPRvirtual}(x)

\S4method{is.nan}{GPRvirtual}(x)
}
\arguments{
\item{x}{(\verb{GPR*})}
}
\value{
(\verb{GPR*}) With logical values (\code{TRUE} is the value is
finite, \code{FALSE} if not.)
}
\description{
is.finite and is.infinite return an object of the same dimension as x,
indicating which elements are finite (not infinite and not missing) or
infinite.
}
