% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\name{is.na,GPRvirtual-method}
\alias{is.na,GPRvirtual-method}
\title{Logical negation}
\usage{
\S4method{is.na}{GPRvirtual}(x)
}
\arguments{
\item{x}{(\verb{GPR*})}
}
\value{
(\verb{GPR*}) With logical values (\code{TRUE} is the value is
\code{NA}, \code{FALSE} if not.)
}
\description{
Indicates which elements are missing.
}
