% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCMP.R
\name{isCMP}
\alias{isCMP}
\alias{isCMP,GPR-method}
\alias{isCommonOffset}
\alias{isCommonOffset,GPR-method}
\alias{isVelSpectrum}
\alias{isVelSpectrum,GPR-method}
\alias{isVelModel}
\alias{isVelModel,GPR-method}
\title{Return TRUE if survey mode is CMP}
\usage{
isCMP(x)

\S4method{isCMP}{GPR}(x)

isCommonOffset(x)

\S4method{isCommonOffset}{GPR}(x)

isVelSpectrum(x)

\S4method{isVelSpectrum}{GPR}(x)

isVelModel(x)

\S4method{isVelModel}{GPR}(x)
}
\arguments{
\item{x}{[\code{GPR}] An object of the class GPR.}
}
\value{
[\code{logical()}] \code{TRUE} if \code{x} is of type CMP/WARR
}
\description{
Return TRUE if survey mode is CMP/WARR
}
