% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCRSLonLat.R
\docType{methods}
\name{isCRSLonLat}
\alias{isCRSLonLat}
\alias{isCRSLonLat,numeric-method}
\alias{isCRSLonLat,integer-method}
\alias{isCRSLonLat,character-method}
\alias{isCRSLonLat,GPR-method}
\title{Check wether the Coordinate Reference Systems is lon/lat}
\usage{
isCRSLonLat(x)

\S4method{isCRSLonLat}{numeric}(x)

\S4method{isCRSLonLat}{integer}(x)

\S4method{isCRSLonLat}{character}(x)

\S4method{isCRSLonLat}{GPR}(x)
}
\arguments{
\item{x}{[\code{character}] CRS (one or more)}
}
\value{
[\code{logical}] \code{TRUE} if lon/lat else \code{FALSE}
}
\description{
Returns \code{TRUE} or \code{FALSE}.
}
