% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDepthUnit.R
\name{isDepthTime}
\alias{isDepthTime}
\alias{isDepthTime,GPR-method}
\alias{isDepthTime,GPRsurvey-method}
\alias{isDepthDepth}
\alias{isDepthDepth,GPRsurvey-method}
\alias{isDepthDepth,GPR-method}
\title{Is the z-dimension time?}
\usage{
isDepthTime(x)

\S4method{isDepthTime}{GPR}(x)

\S4method{isDepthTime}{GPRsurvey}(x)

isDepthDepth(x)

\S4method{isDepthDepth}{GPRsurvey}(x)

\S4method{isDepthDepth}{GPR}(x)
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}
}
\value{
[\code{logical(1)}] \code{TRUE} if the z-dimension is in unit of
          time, \code{FALSE} if the z-dimension is in unit of length (depth)
}
\description{
Is the z-dimension time?
}
