% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R, R/ClassGPRsurvey.R
\docType{methods}
\name{isLengthUnit}
\alias{isLengthUnit}
\title{Return TRUE is the data are a function of length}
\usage{
\S4method{isLengthUnit}{GPR}(x)

\S4method{isLengthUnit}{GPRsurvey}(x)
}
\description{
Return TRUE is the data are a function of length

Return TRUE is the data are a function of length
}
