% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSamplingRegular.R
\name{isSamplingRegular}
\alias{isSamplingRegular}
\alias{isSamplingRegular,GPR-method}
\title{Is the sampling along all axis regular}
\usage{
isSamplingRegular(x, axes = NULL)

\S4method{isSamplingRegular}{GPR}(x, axes = NULL)
}
\arguments{
\item{axes}{(\code{numeric[n]}) Along x-axis = 1, along y-axis = 2}

\item{obj}{(\verb{GPR*} object)}
}
\value{
(\code{logical[1]})
}
\description{
Is the sampling along all axis regular
}
