% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R, R/ClassGPRsurvey.R,
%   R/ClassGPRcube.R
\docType{methods}
\name{isTimeUnit}
\alias{isTimeUnit}
\title{Return TRUE if the data are a function of time}
\usage{
\S4method{isTimeUnit}{GPR}(x)

\S4method{isTimeUnit}{GPRsurvey}(x)

\S4method{isTimeUnit}{GPRslice}(x)
}
\description{
Return TRUE if the data are a function of time

Return TRUE is the data are a function of time

Return TRUE if the data are a function of time
}
