% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isZTime.R
\name{isZTime}
\alias{isZTime}
\alias{isZTime,GPRvirtual-method}
\alias{isZTime,GPRsurvey-method}
\alias{isZDepth}
\alias{isZDepth,GPRsurvey-method}
\alias{isZDepth,GPRvirtual-method}
\title{Is the z-dimension time?}
\usage{
isZTime(x)

\S4method{isZTime}{GPRvirtual}(x)

\S4method{isZTime}{GPRsurvey}(x)

isZDepth(x)

\S4method{isZDepth}{GPRsurvey}(x)

\S4method{isZDepth}{GPRvirtual}(x)
}
\arguments{
\item{x}{(\code{GPR}) An object of the class \code{GPR}}
}
\value{
(\code{logical[1]}) \code{TRUE} if the z-dimension is in unit of
time, \code{FALSE} if the z-dimension is in unit of length (depth)
}
\description{
Is the z-dimension time?
}
