% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\docType{methods}
\name{length,GPRcube-method}
\alias{length,GPRcube-method}
\title{Return length of a GPRcube/GPRslice object (number of slices)}
\usage{
\S4method{length}{GPRcube}(x)
}
\arguments{
\item{x}{[\code{GPR}]}
}
\description{
Return length of a GPRcube/GPRslice object (number of slices)
}
