% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRstack.R
\docType{methods}
\name{length,GPRstack-method}
\alias{length,GPRstack-method}
\title{Length of an object of the class \code{GPR}}
\usage{
\S4method{length}{GPRstack}(x)
}
\description{
The length of an object of the class \code{GPR} is equal to the number of
traces.
}
\examples{
data(frenkeLine00)
length(frenkeLine00)
}
