% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPRsurvey.R
\docType{methods}
\name{length,GPRsurvey-method}
\alias{length,GPRsurvey-method}
\title{Return length of a GPRsurvey object (number of GPR lines)}
\usage{
\S4method{length}{GPRsurvey}(x)
}
\arguments{
\item{x}{[\code{GPRsurvey}]}
}
\description{
Return length of a GPRsurvey object (number of GPR lines)
}
