% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GPR.R
\name{lines.GPR}
\alias{lines.GPR}
\title{Add a GPR trace on a plot}
\usage{
\method{lines}{GPR}(x, relTime0 = FALSE, ...)
}
\arguments{
\item{x}{[\code{GPR class}]}

\item{relTime0}{[\code{logical(1)}] If \code{TRUE}, shift \code{x} to time-
zero.}

\item{...}{Additional parameters to be passed to \code{\link{lines}}.}
}
\description{
Add a GPR trace on a plot
}
