% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{linkCoordFid}
\alias{linkCoordFid}
\title{Link coordinates to fiducial marker}
\usage{
linkCoordFid(y, xyz, pcode)
}
\arguments{
\item{y}{object of class GPSsurvey}

\item{xyz}{matrix of coordinates}

\item{pcode}{character vector (length(pcode) = nrow(xyz)) indicating which
coordinates to which GPR data belongs}
}
\description{
To interpolate topo
}
