% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupLogic_GPRvirtual.R
\name{Logic,GPRvirtual,ANY-method}
\alias{Logic,GPRvirtual,ANY-method}
\alias{Logic}
\alias{Logic,GPRvirtual,GPRvirtual-method}
\alias{Logic,ANY,GPRvirtual-method}
\title{Logic operations}
\usage{
\S4method{Logic}{GPRvirtual,ANY}(e1, e2)

\S4method{Logic}{GPRvirtual,GPRvirtual}(e1, e2)

\S4method{Logic}{ANY,GPRvirtual}(e1, e2)
}
\arguments{
\item{e1}{An object of the class GPRvirtual}

\item{e2}{An object of the class GPRvirtual}
}
\description{
Methods for the base Logic methods \link[methods]{S4groupGeneric}: "==", ">", "<", "!=", "<=", ">="
}
