% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicalNegation.R
\docType{methods}
\name{logicalNegation}
\alias{logicalNegation}
\alias{!,GPRvirtual-method}
\title{Logical negation}
\usage{
\S4method{!}{GPRvirtual}(x)
}
\arguments{
\item{x}{[\code{GPRvirtual}]}
}
\description{
Apply logical negation to GPR data
}
