% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS.R
\name{lonlatToUTM}
\alias{lonlatToUTM}
\title{Convert longitude, latitude to UTM}
\usage{
lonlatToUTM(lon, lat, zone = NULL, south = NULL)
}
\arguments{
\item{lon}{[\code{numeric}] Longitude.}

\item{lat}{[\code{numeric}] Latitude}

\item{zone}{[\code{integer(1)}] UMT zone (optional).}

\item{south}{[\code{logical(1)}] \code{TRUE} if the coordinates are in the
southern hemisphere, else \code{FALSE}.}
}
\value{
[\code{list(2)}] \code{xy} the coordinates in UTM, 
        \code{crs} the UTM coordinate reference system (proj4string).
}
\description{
see https://stackoverflow.com/a/30225804
https://stackoverflow.com/questions/18639967/converting-latitude-and-longitude-points-to-utm
}
