% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{Math,GPR-method}
\alias{Math,GPR-method}
\alias{Math-GPR-method}
\title{Basic mathematical functions}
\usage{
\S4method{Math}{GPR}(x)
}
\arguments{
\item{x}{An object of the class RGPR.}
}
\description{
Methods for the base Math methods \link[methods]{S4groupGeneric}
}
\details{
Currently implemented methods include:
\itemize{
 \item{"abs", "sign", "sqrt", "ceiling", "floor", "trunc",
       "exp", "expm1", "log", "log10", "log2", "log1p", "cos",
       "cosh", "sin", "sinh", "tan", "tanh"}
 }
}
\examples{
data(frenkeLine00)
A <- exp(frenkeLine00)
}
