% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMath2_GPRvirtual.R
\docType{methods}
\name{Math2,GPRvirtual-method}
\alias{Math2,GPRvirtual-method}
\alias{Math2-GPRvirtual-method}
\title{Round and signif}
\usage{
\S4method{Math2}{GPRvirtual}(x, digits)
}
\arguments{
\item{x}{An object of the class GPRvirtual}

\item{digits}{number of digits to be used in round or signif.}
}
\description{
Methods for the base Math2 methods \link[methods]{S4groupGeneric}: "round", "signif"
}
