% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{max,GPR-method}
\alias{max,GPR-method}
\title{Maximum value of GPR data}
\usage{
\S4method{max}{GPR}(x, ..., na.rm = FALSE)
}
\description{
Maximum value of GPR data
}
\examples{
data(frenkeLine00)
min(frenkeLine00)
}
