% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\docType{methods}
\name{mean,GPR-method}
\alias{mean,GPR-method}
\title{Mean of a GPR object}
\usage{
\S4method{mean}{GPR}(x, ...)
}
\arguments{
\item{x}{[\code{GPR}]}

\item{...}{Additional parameters to be passed (see \code{\link{mean}}).}
}
\description{
Mean of all the values contained in \code{x}.
}
