% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\name{median,GPRvirtual-method}
\alias{median,GPRvirtual-method}
\title{Median of a GPR object}
\usage{
\S4method{median}{GPRvirtual}(x, na.rm = FALSE)
}
\arguments{
\item{x}{[\code{GPR}]}

\item{na.rm}{[\code{logical}]  If \code{TRUE} NA values are stripped before 
the computation proceeds.}
}
\description{
Median of all the values contained in \code{x}.
}
