% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{metadata,GPR-method}
\title{Meta-data extracted from the GPR data}
\usage{
metadata(x)

\S4method{metadata}{GPR}(x)
}
\arguments{
\item{x}{[\code{GPR}] An object of the class GPR.}
}
\value{
[\code{list}]
}
\description{
Meta-data extracted from the GPR data that depends on the manufacturer
file format. All the information that are not used to construct the RGPR
objects is still stored as meta-data and can be retrieved with \code{md()}.
}
\details{
The element names of the list are given below
\describe{
  \item{hd}{Here comes all the information from the raw GPR data file that
            is not used to construct a RGPR object.
            It can be retrieved with \code{metadata(x)$hd}.}
  \item{GPS}{When reading GPR data, the optional GPS data file is red if 
             available and red even if \code{interpGPS = FALSE} 
             (see \code{\link{readGPR}}). If \code{interpGPS = FALSE}, 
             the (formated) content of the GPS data file is then stored as 
             meta-data and can be retrieved with \code{metadata(x)$GPS}.}
  \item{clip}{When reading GPR data, the clipped signal values are directly
              estimated from the bit values and stored as metadata.
              They can be retrieved with \code{metadata(x)$clip}.
              These clipped values can be used directly
              .}
}
}
\seealso{
\code{\link{readGPR}}
}
