% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migration.R
\docType{methods}
\name{migration}
\alias{migration}
\title{Migration of the GPR data}
\usage{
\S4method{migration}{GPR}(x, type = c("static", "kirchhoff"), ...)
}
\description{
Fresnel zone defined according to 
Perez-Gracia et al. (2008) Horizontal resolution in a non-destructive
shallow GPR survey: An experimental evaluation. NDT & E International,
41(8): 611-620.
doi:10.1016/j.ndteint.2008.06.002
}
