% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRvirtual.R
\docType{methods}
\name{min,GPRvirtual-method}
\alias{min,GPRvirtual-method}
\title{Mininum value of GPR data}
\usage{
\S4method{min}{GPRvirtual}(x, ..., na.rm = FALSE)
}
\description{
Mininum value of GPR data
}
\examples{
data(frenkeLine00)
min(frenkeLine00)
}
