% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getterSetter.R
\name{name}
\alias{name}
\alias{name<-}
\alias{setName}
\title{Name of the GPR data}
\usage{
\S4method{name}{GPR}(x)

\S4method{name}{GPR}(x) <- value

\S4method{setName}{GPR}(x, value)
}
\description{
Name of the GPR data
}
