% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\docType{methods}
\name{ncol,GPRvirtual-method}
\alias{ncol,GPRvirtual-method}
\title{Number of columns (samples per trace)}
\usage{
\S4method{ncol}{GPRvirtual}(x)
}
\arguments{
\item{x}{[\code{GPR}]}
}
\description{
Number of columns (samples per trace)
}
