% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMO.R
\docType{methods}
\name{NMO}
\alias{NMO}
\alias{NMO,GPR-method}
\title{Normal Move-Out}
\usage{
NMO(x, v = NULL)

\S4method{NMO}{GPR}(x, v = NULL)
}
\arguments{
\item{x}{An object of the class \code{GPR}}

\item{v}{A length-one numeric vector defining the radar wave velocity in 
the ground}
}
\description{
Compute the Normal Move-Out (NMO) for a data set given a constant velocity: 
The Normal Move-out is defined as the
difference between the two-way time at a given offset and the two-way 
zero-offset time.
 We write the vertical
two-way traveltime at zero offset 
\deqn{t_0 = t_{TWT}(x = 0) = \frac{2z}{v}}
Therefore, the NMO-correction \eqn{\Delta_{NMO}} is
\deqn{\Delta_{NMO} = t_{TWT}(x) - t_0}  
\deqn{\Delta_{NMO} = t_0 (\sqrt{1 + \frac{x^2}{v^2 t_0^2}} - 1)}
}
