% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{NMOCor}
\alias{NMOCor}
\alias{NMOCor,GPR-method}
\title{Normal Move-Out correction}
\usage{
\S4method{NMOCor}{GPR}(x, v = NULL)
}
\arguments{
\item{x}{An object of the class \code{GPR}}

\item{v}{A length-one numeric vector defining the radar wave velocity in 
the ground}

\item{asep}{A length-n numeric vector defining the antenna separation for
each trace (n = number of traces; for example:
\code{seq(x@antsep, by = x@dx, length.out = length(x))}). 
If \code{NULL}, the
slots \code{rec} and \code{trans} of \code{x} are used to 
compute the distance between the antennas}
}
\description{
Remove the Normal Move-Out (NMO) from the trace given a constant velocity: 
this is a non-linear 
correction of the time axis that require interpolation. Note that
only the conventional NMO correction is currently implemented. The 
conventional NMO introduces a streching effect. A nonstretch NMO will
be implemented in a near future.
}
\details{
Assuming a horizontal reflecting plane and homogeneous medium, the two-way
bistatic travel time of the reflected wave 
for an antenna separation \eqn{x} follows directly from the Pythagorean 
theorem:
\deqn{t_{TWT}(x,z) = \sqrt{\frac{x^2}{v^2} + \frac{4z^2}{v^2}}}
where \eqn{t_{TWT}(x)} is the two-way travel time at antenna
separation \eqn{x} of the wave reflected at depth \eqn{z} with propagation
velocity \eqn{v}. This equation defines an hyperbola (keep \eqn{z} constant,
increase the antenna separation \eqn{x} and you obtain a hyperbola similar
to the reflection signals you obtain with common-mid point survey).
The idea behind NMO-correction is to correct the signal for the antenna 
separation (offset) and therefore to transform the signal to the signal we 
would have recorded with zero offset (\eqn{x = 0}). We write the vertical
two-way traveltime at zero offset 
\deqn{t_0 = t_{TWT}(x = 0) = \frac{2z}{v}}
Therefore, the NMO-correction \eqn{\Delta_{NMO}} is
\deqn{\Delta_{NMO} = t_{TWT}(x) - t_0}  
\deqn{\Delta_{NMO} = t_0 (\sqrt{1 + \frac{x^2}{v^2 t_0^2}} - 1)}
}
\references{
\itemize{
  \item{Tillard and Dubois (1995) Analysis of GPR data: wave propagation
        velocity determination. Journal of Applied Geophysics, 33:77-91}
  \item{Shatilo and Aminzadeh (2000) Constant normal-moveout (CNMO) 
        correction: a technique and test results. Geophysical Prospecting,
        473-488}
}
}
