% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{NMOCor}
\alias{NMOCor}
\title{Normal Move-Out correction}
\usage{
\S4method{NMOCor}{GPR}(x, v = NULL, asep = NULL)
}
\arguments{
\item{x}{An object of the class \code{GPR}}

\item{v}{A length-one numeric vector defining the radar wave velocity in 
the ground}

\item{asep}{A length-n numeric vector defining the antenna separation for
each trace (n = number of traces)}
}
\description{
either use 'rec' and 'trans' to compute the distance between the antennas
or give the distance between the antennas (asep)
or seq(x@antsep, by = x@dx, length.out = length(x))
}
