% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMOstreching.R
\name{NMOstreching}
\alias{NMOstreching}
\alias{NMOstreching,GPR-method}
\title{Normal Move-Out streching}
\usage{
NMOstreching(x, v = NULL)

\S4method{NMOstreching}{GPR}(x, v = NULL)
}
\arguments{
\item{x}{An object of the class \code{GPR}}

\item{v}{A length-one numeric vector defining the radar wave velocity in 
the ground}
}
\description{
Compute the Normal Move-Out (NMO) streching for a data set given a constant velocity.
The NMO streching is defined by \deqn{S_{NMO} = \frac{\Delta_{NMO}}{t_0}}, where
\deqn{t_0 = t_{TWT}(x = 0) = \frac{2z}{v}} is the vertical two-way traveltime at zero offset.
}
