% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\docType{methods}
\name{nrow,GPRvirtual-method}
\alias{nrow,GPRvirtual-method}
\title{Number of rows (samples per trace)}
\usage{
\S4method{nrow}{GPRvirtual}(x)
}
\arguments{
\item{x}{[\code{GPR}]}
}
\description{
Number of rows (samples per trace)
}
