% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{palCol}
\alias{palCol}
\title{Return color from palette}
\usage{
palCol(x, col = palGPR(n = 101), sym = TRUE, clim = NULL)
}
\arguments{
\item{x}{Values}

\item{col}{[\code{character}] Colors to be used.}

\item{sym}{[\code{logical(1)}] Should the color palette be symmetric?}

\item{clim}{[\code{numeric(2)}] The range of the color values, used in the 
color palette.}
}
\description{
Return color from palette
}
\concept{colors}
