% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{palGPR}
\alias{displayPalGPR}
\alias{palGPR}
\alias{plotPal}
\title{Colour palette}
\usage{
palGPR(colPal = "default", n = 101, power = 1, returnNames = FALSE)

plotPal(col, border = NA)

displayPalGPR()
}
\description{
Colour palette
}
\examples{
displayPalGPR()
plotPal(palGPR("hcl_5"))
}

