% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_path.R
\name{pathLength}
\alias{pathLength}
\title{Path length}
\usage{
pathLength(xy)
}
\arguments{
\item{xy}{[\code{matrix}] Each column corresponding to a coordinate}
}
\value{
Path length.
}
\description{
Path length
}
