% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{perpPoints}
\alias{perpPoints}
\title{Points perdicular to a polyline}
\usage{
perpPoints(xy, d)
}
\arguments{
\item{xy}{matrix n row, 2 column (x and y positions)}

\item{d}{numeric vector of length m defining the distance of the transect from the
polyline points. If length m > 1 several transects are returned.}
}
\value{
a list with elements x and y of dimension (n, m).
}
\description{
Compute oriented transects along a polyline (one transect per points)
}
