% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{phaseRotation}
\alias{phaseRotation}
\title{Phase rotation
      
shift the phase of signal by phi (in radian)}
\usage{
phaseRotation(x, phi, npad = 10)
}
\description{
Phase rotation
      
shift the phase of signal by phi (in radian)
}
