% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2PDF.R
\name{plot2PDF}
\alias{plot2PDF}
\title{Export a PDF showing the GPR profile.}
\usage{
\S4method{plot2PDF}{GPR}(
  x,
  fPath = NULL,
  addTopo = FALSE,
  clip = NULL,
  normalize = NULL,
  nupspl = NULL,
  type = "raster",
  ...
)
}
\description{
Export a PDF showing the GPR profile.
}
