% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2PNG.R
\name{plot2PNG}
\alias{plot2PNG}
\title{Export a PNG showing the GPR profile.}
\usage{
\S4method{plot2PNG}{GPR}(
  x,
  fPath = NULL,
  addTopo = FALSE,
  clip = NULL,
  normalize = NULL,
  nupspl = NULL,
  type = "raster",
  ...
)
}
\description{
Export a PNG showing the GPR profile.
}
