% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly}
\alias{plotly}
\title{Interactive plot of GPR objects with plotly.js}
\usage{
plotly(
  x,
  col = palGPR(),
  add = FALSE,
  relTime0 = FALSE,
  main = NULL,
  addTopo = FALSE,
  elev = FALSE,
  clip = NULL,
  ...
)
}
\arguments{
\item{x}{Object of class \code{GPR}}

\item{relTime0}{logical. If \code{TRUE}, adjust vertical axis to time-zero.
If time-zero varies from trace to trace, the vertical axis
is adjusted to the mean time-zero. Apply first the function
\code{time0Cor()} to shift the traces to their time-zero.}

\item{addTopo}{logical. For 2D plot, add topography (if the data are sampled
in time unit, the data are migrated with a static
migration)}

\item{elev}{logical. If \code{TRUE} the elevation instead of depth is
displayed on y-axis (and therefore, time-to-depth
conversion is applied).}

\item{clip}{numeric. If length-one numeric vector, clip the amplitudes 
larger than \code{clip} and smaller than \code{-clip}.
If length-two numeric vector, clip the amplitudes
smaller than \code{clip[1]} and larger than 
\code{clip[2]}. Per default, values below the 
0.01-quantile and above the 0.99-quantile are clipped.
If \code{clip = FALSE} the data are not clipped.}

\item{...}{additional arguments passed to the plot_ly function, see 
\code{\link[plotly]{plot_ly}}.}
}
\description{
A wrapper for the function \code{\link[plotly]{plot_ly}}. Still experimental.
}
