% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{plotRaster}
\alias{plotRaster}
\title{Plot GPR as image (raster)}
\usage{
plotRaster(z, x = NULL, y = NULL, main = "", xlim = NULL, note = NULL,
  ratio = 1, time_0 = 0, antsep = 1, v = 0.1, surveymode = NULL,
  addFid = TRUE, fid = NULL, ylim = NULL, addAnn = TRUE,
  annotations = NULL, depthunit = "ns", posunit = "m",
  rasterImage = FALSE, resfac = 1, clab = "mV", add = FALSE,
  barscale = TRUE, addGrid = FALSE, col = palGPR(n = 101), yaxt = "s",
  bty = "o", relTime0 = FALSE, clim = NULL, pdfName = NULL, ...)
}
\description{
Plot GPR as image (raster)
}
