% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTr.R
\name{plotTr}
\alias{plotTr}
\alias{plotTr,GPRvirtual-method}
\title{Plot all the traces in one 1D plot}
\usage{
plotTr(x, add = FALSE, ...)

\S4method{plotTr}{GPRvirtual}(x, add = FALSE, ...)
}
\arguments{
\item{x}{Object of the class GPR}

\item{add}{(\code{logical[1]}) If \code{TRUE}, add to current plot.}

\item{...}{Arguments to be passed to \code{plot}/\code{line}}
}
\description{
Plot all the traces in one 1D plot
}
