% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVelLayers.R
\name{plotVelLayers}
\alias{plotVelLayers}
\alias{plotVelLayers,GPR-method}
\title{Plot the velocity layers on a 2D plot}
\usage{
plotVelLayers(
  obj,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  ...
)

\S4method{plotVelLayers}{GPR}(
  obj,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  ...
)
}
\arguments{
\item{obj}{(\code{GPR* object}) An object of the class \code{GPR}}

\item{method}{(\code{character[1]}) Interpolation method.}

\item{col}{(\code{character[1|n]}) Color.}

\item{...}{Additional arguments for the function \code{plotDelineations()}}
}
\description{
Plot the velocity layers on a 2D plot
}
