% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GPR.R
\name{points}
\alias{points}
\alias{points.GPR}
\title{Add a GPR trace points on a plot}
\usage{
\method{points}{GPR}(x, relTime0 = FALSE, ...)
}
\arguments{
\item{x}{[\code{GPR class}]}

\item{relTime0}{[\code{logical(1)}] If \code{TRUE}, shift \code{x} to time-
zero.}

\item{...}{Additional parameters to be passed to \code{\link{points}}.}
}
\description{
Add a GPR trace points on a plot
}
