% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/getterSetter.R
\docType{methods}
\name{pos}
\alias{pos}
\alias{pos<-}
\title{Position of the GPR traces}
\usage{
\S4method{pos}{GPR}(x)

\S4method{pos}{GPR}(x) <- value
}
\description{
Position of the GPR traces
}
