% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{posunit}
\alias{posunit}
\alias{posunit<-}
\alias{posunit}
\alias{posunit<-}
\title{Position unit of the GPR data}
\usage{
\S4method{posunit}{GPR}(x)

\S4method{posunit}{GPR}(x) <- value
}
\description{
Position unit of the GPR data
}
