% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_GPRcube.R
\name{print.GPRcube}
\alias{print.GPRcube}
\title{Print GPRcube}
\usage{
\method{print}{GPRcube}(x, ...)
}
\arguments{
\item{x}{[\code{GPRcube object}]}

\item{...}{Not used.}
}
\description{
Print GPRcube
}
