% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_GPRslice.R
\name{print.GPRslice}
\alias{print.GPRslice}
\title{Print GPRslice}
\usage{
\method{print}{GPRslice}(x, ...)
}
\arguments{
\item{x}{(\verb{GPRslice object})}

\item{...}{Not used.}
}
\description{
Print GPRslice
}
