% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_GPRsurvey.R
\name{print.GPRsurvey}
\alias{print.GPRsurvey}
\title{Print GPRsurvey}
\usage{
\method{print}{GPRsurvey}(x, ...)
}
\arguments{
\item{x}{[\code{GPRsurvey object}]}

\item{...}{Not used.}
}
\description{
Print GPRsurvey
}
