% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc.R
\docType{methods}
\name{proc}
\alias{proc}
\alias{proc<-}
\alias{proc,GPR-method}
\alias{proc<-,GPR-method}
\title{Get processing steps applied to the data}
\usage{
proc(x)

proc(x) <- value

\S4method{proc}{GPR}(x)

\S4method{proc}{GPR}(x) <- value
}
\arguments{
\item{x}{[\code{GPR}] An object of the class GPR.}

\item{value}{[\code{character}]}
}
\value{
A character vector whose elements contain the name of the 
processing functions with their arguments applied previously on the
GPR data.
}
\description{
\code{processing} returns all the processing steps applied to the data.
}
