% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\alias{project,GPR-method}
\alias{project,GPRsurvey-method}
\title{Project trace coordinates}
\usage{
project(x, CRSobj)

\S4method{project}{GPR}(x, CRSobj)

\S4method{project}{GPRsurvey}(x, CRSobj)
}
\arguments{
\item{x}{Object of the class GPR}

\item{CRSobj}{(\code{character[1]}) A string accepted by GDAL
(e.g., \code{"EPSG:2056"}, WKT-string).}
}
\description{
Project the trace coordinates give a coordinate reference system.
}
\details{
Modified slots
\itemize{
\item \code{coord} the trace coordinates
\item \code{x} the local trace position (along profile)
\item \code{crs} the coordinate reference system.
\item \code{spunit} the spatial units are updated accroding to the new
coordinate reference system.
}
}
\concept{spatial computation}
