% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{quiet}
\alias{quiet}
\title{Suppressing output from cat() or print()}
\usage{
quiet(x)
}
\description{
This function suppresses the output from cat() or print() in a function. 
It was proposed by Hadley Wickham 
https://r.789695.n4.nabble.com/Suppressing-output-e-g-from-cat-td859876.html
}
