% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{range,GPR-method}
\alias{range,GPR-method}
\title{Range of GPR data}
\usage{
\S4method{range}{GPR}(x, ..., na.rm = FALSE)
}
\description{
Range of GPR data
}
\examples{
data(frenkeLine00)
range(frenkeLine00)
}
