% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raytracing.R
\name{raytrace}
\alias{raytrace}
\title{Ray tracing}
\usage{
raytrace(h, v, theta, x = 0)
}
\arguments{
\item{h}{(\code{numeric[n]}) Layer heights}

\item{v}{(\code{numeric[n]}) Layer velocity}

\item{theta}{(\code{numeric[n]}) Incidence angle above first layer}
}
\description{
Ray tracing
}
