% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRDX.R
\name{readCOR}
\alias{readCOR}
\title{Read MALA coordinate files}
\usage{
readCOR(dsn, UTM = FALSE)
}
\arguments{
\item{dsn}{(\code{*}) Data source name. Can be a filename or a connection to a file.}

\item{UTM}{(\code{logical[1]}) Project to UTM CRS?}
}
\value{
(\code{sf}) A \code{sf} point object.
}
\description{
Read MALA coordinate files
}
