% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDT1.R
\name{readDT1}
\alias{readDT1}
\title{Read Sensors and Software GPR data}
\usage{
readDT1(dsn, ntr, npt)
}
\arguments{
\item{dsn}{(\verb{character(1)|connection object}) data source name:
either the filepath to the GPR data (character),
or an open file connection.}

\item{ntr}{(\code{integer(1)}) Number of traces (given by .hd file)}

\item{npt}{(\code{integer(1)}) Number of samples per traces (given by .hd file)}
}
\value{
(\code{list(2)}) Two-elements list: \code{dt1hd} with trace header and
\code{data} with the traces.
}
\description{
Read Sensors and Software GPR data
}
\seealso{
\code{\link[=readHD]{readHD()}}, \code{\link[=readGPS]{readGPS()}}
}
\concept{I/O}
