% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFID.R
\name{readFID}
\alias{readFID}
\title{read fiducial marker files}
\usage{
readFID(FID, sep = NULL, verbose = TRUE)
}
\arguments{
\item{FID}{list}

\item{sep}{separator}

\item{verbose}{logical}
}
\value{
list
}
\description{
read fiducial marker files
}
