% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS.R
\name{readGPGGA}
\alias{readGPGGA}
\title{Read GPS file with GPGGA string}
\usage{
readGPGGA(dsn, sep = ",")
}
\arguments{
\item{dsn}{[\code{character(1)|connection object}] Data source name: 
either the filepath to the GPR data (character),
or an open file connection.}

\item{sep}{[\code{character(1)}] The field separator character
(see\code{\link{read.table}}).}
}
\value{
[\code{SpatialPoints}]
}
\description{
Read GPS file with GPGGA string
}
