% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDT1.R
\name{readHD}
\alias{readHD}
\title{Read Sensors and Software .HD file}
\usage{
readHD(dsn)
}
\arguments{
\item{dsn}{[\code{character(1)|connection object}] data source name: 
either the filepath to the GPR data (character),
or an open file connection.}
}
\value{
[\code{list(3)}]  Three-elements list: \code{HD} containing the 
        header info, \code{ntr} the number of trace and \code{npt} the
        number of points per trace.
}
\description{
Read Sensors and Software .HD file
}
\seealso{
\code{\link{readDT1}}, \code{\link{readGPS}}
}
