% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readRD3}
\alias{readRD3}
\title{Read ".rd3" GPR files}
\usage{
readRD3(fPath)
}
\arguments{
\item{fPath}{A length-one character vector: the GPR data file path}
}
\value{
A list with all the elements contained in the GPR data file.
}
\description{
This function read *.rd3 and associated *.rad files from Mala
}
