% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readSEGY}
\alias{readSEGY}
\title{Read ".segy" GPR files}
\usage{
readSEGY(fPath)
}
\arguments{
\item{fPath}{A length-one character vector: the GPR data file path}
}
\value{
A list with all the elements contained in the GPR data file.
}
\description{
This function read *.segy (RadSys Zond GPR device)
}
