% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTopo.R
\name{readTopo}
\alias{readTopo}
\title{read topo file}
\usage{
readTopo(TOPO, sep = NULL, verbose = TRUE)
}
\arguments{
\item{TOPO}{file}

\item{sep}{separatro}

\item{verbose}{logical(1)}
}
\value{
list
}
\description{
read topo file
}
