% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%T>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}, \code{\link[magrittr:\%T>\%]{\%T>\%}}}
}}

