% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raytracing.R
\name{refracAng}
\alias{refracAng}
\title{Refraction angle}
\usage{
refracAng(theta1, v1, v2)
}
\arguments{
\item{theta1}{Incidence angle (media 1)}

\item{v1}{Electromagnetic wave velocity of media 1}

\item{v2}{Electromagnetic wave velocity of media 2}
}
\description{
Compute the refraction angle based on the incidence angle and the
refractive indices of the two media (applies Snell's Law, also known as the
law of refraction)
}
