% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{regInterpPos}
\alias{regInterpPos}
\title{Trace interpolation at regularly spaced positions}
\usage{
\S4method{regInterpPos}{GPR}(x, type = c("linear", "cosine"),
  dx = NULL)
}
\description{
Trace interpolation at regularly spaced positions
}
