% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relPos.R
\name{relPos}
\alias{relPos}
\alias{relPos,GPR-method}
\title{Relative positions along path}
\usage{
relPos(x)

\S4method{relPos}{GPR}(x)
}
\arguments{
\item{x}{[\code{GPR|GPRsurvey}] An object of the class \code{GPR} or
\code{GPRsurvey}}
}
\value{
[\code{numeric{m}}] Relative trace position along GPR line.
}
\description{
Relative positions along path
}
