% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\name{relTrPos}
\alias{relTrPos}
\alias{relTrPos3D}
\title{Relative trace position on the GPR profile.}
\usage{
\S4method{relTrPos}{GPR}(x, last = FALSE)

\S4method{relTrPos3D}{GPR}(x, last = FALSE)
}
\description{
Trace position computed from (x, y)

Trace position computed from (x, y, z)
}
