% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{repmat}
\alias{repmat}
\title{Repeat matrix}
\usage{
repmat(A, n, m)
}
\description{
Repeat a matrix row-wise n times and column-wise m times.
}
\details{
Source
A replication of MatLab repmat function!
R FOR OCTAVE USERS
version 0.4, Copyright (C) 2001 Robin Hankin
http://cran.r-project.org/doc/contrib/R-and-octave.txt
}

