% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustGamma.R
\name{gammaCorrection}
\alias{gammaCorrection}
\title{Gamma correction of the amplitude}
\usage{
\S4method{gammaCorrection}{GPR}(x, a = 1, b = 1, track = TRUE)
}
\description{
This function computes the gamma correction.
}
