% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGPR.R
\docType{package}
\name{RGPR-package}
\alias{RGPR}
\alias{RGPR-package}
\title{RGPR: A package for processing and visualising ground-penetrating data
radar (GPR) data.}
\description{
The RGPR package provides two classes GPR and GPRsurvey
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emanuelhuber/RGPR}
  \item Report bugs at \url{https://github.com/emanuelhuber/RGPR/issues}
}

}
\author{
\strong{Maintainer}: Emanuel Huber \email{emanuel.huber@pm.me}

Authors:
\itemize{
  \item Emanuel Huber \email{emanuel.huber@pm.me}
  \item Guillaume Hans \email{guillaume.hans@fpinnovations.ca}
}

}
