% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{RGPR}
\alias{RGPR}
\title{RGPR: A package for processing and visualising ground-penetrating data 
(GPR).}
\description{
The RGPR package provides two classes GPR and GPRsurvey
}
\section{Reading/writing/export functions}{

\itemize{
  \item \code{readGPR()}: format DT1 (Sensors&Software), rds (R-format)
  \item \code{writeGPR()}: format DT1 (Sensors&Software), rds (R-format)
  \item \code{exportPDF()}
  \item \code{exportDelineations()}
  \item \code{exportFid()}: ASCII-file
  \item \code{exportCoord()}: output either SpatialLines or SpatialPoints
  \item \code{exportProc()}: ASCII-file
}
}

\section{Plot functions}{

\itemize{
  \item \code{plot()}: raster or wiggles.
  \item \code{plot3D()}:
  \item \code{plotAmpl()}
  \item \code{plotDelineations()}
}
}

