% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRsurvey.R
\docType{methods}
\name{rotate,GPRsurvey-method}
\alias{rotate,GPRsurvey-method}
\title{Rotate coordinates of the GPR traces}
\usage{
\S4method{rotate}{GPRsurvey}(x, alpha, center = NULL, center2 = NULL)
}
\description{
Rotate coordinates of the GPR traces
}

