% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale1D.R
\name{scale1D}
\alias{scale1D}
\title{Scale the traces}
\usage{
scale1D(
  x,
  type = c("stat", "min-max", "95", "eq", "sum", "rms", "mad", "invNormal"),
  track = TRUE
)

\S4method{scale1D}{GPR}(
  x,
  type = c("stat", "min-max", "95", "eq", "sum", "rms", "mad", "invNormal"),
  track = TRUE
)
}
\description{
Scale the traces
}
