% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRsurvey.R
\name{setCoordref}
\alias{setCoordref}
\alias{setCoordref,GPRsurvey-method}
\title{Define a local reference coordinate}
\usage{
\S4method{setCoordref}{GPRsurvey}(x)
}
\description{
Define a local reference coordinate
}
