% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{setDefaultListValues}
\alias{setDefaultListValues}
\title{set default values to the element list.}
\usage{
setDefaultListValues(x, xdefault)
}
\arguments{
\item{x}{[\code{list}] A list.}

\item{xdefault}{[\code{list}] The list with default values.}
}
\description{
Set default values to elements of a list.
}
