% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{setDots}
\alias{setDots}
\title{set default values}
\usage{
setDots(dots, defaults)
}
\arguments{
\item{dots}{[\code{list}] A list.}

\item{defaults}{[\code{list}] The list with default values.}
}
\description{
Set default values to missing elements of a list.
}
