% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRsurvey.R
\docType{methods}
\name{setGridCoord}
\alias{setGridCoord}
\alias{setGridCoord<-,GPRsurvey-method}
\title{Set grid coordinates the trace position.}
\usage{
setGridCoord(x) <- value

\S4method{setGridCoord}{GPRsurvey}(x) <- value
}
\arguments{
\item{x}{An object of the class GPRsurvey}

\item{value}{A list with following elements: \code{xlines} (number or id of 
the GPR data along the x-coordinates), \code{ylines} (number or 
id of the GPR data along the y-coordinates), \code{xpos} 
(position of the x-GPR data on the x-axis),
\code{xpos} (position of the y-GPR data on the y-axis)}
}
\description{
Set grid coordinates to a survey
}
