% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLayerVelocities.R
\docType{methods}
\name{setLayerVelocites}
\alias{setLayerVelocites}
\title{Set layer velocities}
\usage{
setLayerVelocites(x, v, twt = NULL, method = "pchip", clean = TRUE,
  extrap = NA)

\S4method{setLayerVelocites}{GPR}(x, v, twt = NULL, method = "pchip",
  clean = TRUE, extrap = NA)
}
\arguments{
\item{x}{GPR object}

\item{v}{velocities (length(v) = nrow(twt_int + 1))  (+1 because of the last layers)}

\item{twt}{vertical resolution of the migrated data}

\item{method}{interpolation method}
}
\description{
Set layer velocities
}
