% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLayerVelocities.R
\docType{methods}
\name{setLayerVelocities}
\alias{setLayerVelocities}
\title{Set layer velocities}
\usage{
setLayerVelocities(x, v, twt = NULL, method = "pchip", clean = TRUE,
  extrap = NA)

\S4method{setLayerVelocities}{GPR}(x, v, twt = NULL, method = "pchip",
  clean = TRUE, extrap = NA)
}
\arguments{
\item{x}{GPR object}

\item{v}{velocities (\code{length(v) = nrow(twt_int + 1)}, (\code{+1} because of the last layers)}

\item{twt}{vertical resolution of the migrated data}

\item{method}{interpolation method}
}
\description{
Set layer velocities
}
