% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVel2.R
\name{setVel2}
\alias{setVel2}
\alias{setVel2,GPR-method}
\title{Set velocities}
\usage{
setVel2(x, v, twt, type = c("vrms", "vint"))

\S4method{setVel2}{GPR}(x, v, twt, type = c("vrms", "vint"))
}
\arguments{
\item{x}{[\code{GPR}] GPR object}

\item{v}{[\code{numeric(n)}] \code{n} velocity values}

\item{twt}{[\code{numeric(n)}] \code{n} two-way travel time values associated
with the velocities \code{v}. Optional.}

\item{type}{[\code{"vrms"}|\code{vint}] Kind of velocities \code{v}: 
root-mean-square velocities \code{"vrms"} (genrally inferred
from CMP data) or internal velocities \code{"vint"} (layer 
velocities).}
}
\value{
[\code{GPR}]
}
\description{
set velocities.
}
