% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\name{shiftEst}
\alias{shiftEst}
\title{Shift estimation between two GPR profiles.}
\usage{
\S4method{shiftEst}{GPR}(x, y = NULL, method = c("phase", "WSSD"), dxy = NULL, ...)
}
\description{
Shift estimation between two GPR profiles.
}
