% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftGridLine.R
\name{shiftGridLine}
\alias{shiftGridLine}
\alias{shiftGridLine,GPRsurvey-method}
\title{Shift trace positions of one GPR data}
\usage{
shiftGridLine(x, i, dx = 0, dy = 0, dz = 0)

\S4method{shiftGridLine}{GPRsurvey}(x, i, dx = 0, dy = 0, dz = 0)
}
\arguments{
\item{x}{(\code{GPRsurvey}) Index of the line to be shifted}

\item{i}{(\code{numeric[1]}) Index of the line to be shifted}

\item{dx}{(\code{numeric[1]}) How much should the line be shifted in the x-direction (in spatial unit).}

\item{dy}{(\code{numeric[1]}) How much should the line be shifted in the y-direction (in spatial unit).}

\item{dz}{(\code{numeric[1]}) How much should the line be shifted in the z-direction (in spatial unit).}
}
\description{
Shift trace positions of GPR data \code{i} by \code{dx} along x-axis and
by \code{dy} along y-axis.
}
