% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftTopo.R
\name{shiftTopo}
\alias{shiftTopo}
\title{Shift trace vertically according to topography}
\usage{
shiftTopo(
  x,
  method = c("pchip", "linear", "nearest", "spline", "cubic", "none"),
  crop = FALSE,
  track = TRUE
)
}
\arguments{
\item{x}{(\code{GPR}) An object of the class \code{GPR}}

\item{method}{(\code{character[1]}) Interpolation method to be applied:
one of \code{pchip}, \code{linear}, \code{nearest},
\code{spline}, \code{cubic}, \code{none}
(see also \code{\link[signal:interp1]{signal::interp1()}}).
\code{"none"} means that the trace is shifted by the
amount of trace samples the closest to \code{z} without
interpolation.}

\item{crop}{(\code{logical[1]})
If \code{TRUE} (default), remove the rows containing only
zero's (no data).}

\item{track}{(\code{logical[1]}) If \code{TRUE}, processing will be tracked.}
}
\value{
(\verb{GPR class}) An object of the class GPR.
}
\description{
Shift traces vertically according to topography. New traces
are interpolated.
}
\details{
Modified slots
\itemize{
\item \code{data}: trace shifted. The number of rows of data may
be smaller if \code{crop = TRUE}.
\item \code{proc}: updated with function name and arguments.
}
}
