% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_GPR.R
\docType{methods}
\name{show,GPR-method}
\alias{show,GPR-method}
\title{Show some information on the GPR object}
\usage{
\S4method{show}{GPR}(object)
}
\arguments{
\item{object}{[\code{GPR}]}
}
\description{
Identical to print().
}
