% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_GPRslice.R
\docType{methods}
\name{show,GPRslice-method}
\alias{show,GPRslice-method}
\title{Show some information on the GPRslice object}
\usage{
\S4method{show}{GPRslice}(object)
}
\arguments{
\item{object}{[\code{GPRslice object}]}
}
\description{
Identical to print().
}
