% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_GPRsurvey.R
\name{show,GPRsurvey-method}
\alias{show,GPRsurvey-method}
\title{Show some information on the GPR object}
\usage{
\S4method{show}{GPRsurvey}(object)
}
\arguments{
\item{object}{[\code{GPRsurvey object}]}
}
\description{
Identical to print().
}
