% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCMP.R
\name{simCMP}
\alias{simCMP}
\title{Simulates CMP data}
\usage{
simCMP(
  vint = c(0.1, 0.095, 0.08, 0.09, 0.105, 0.09, 0.095),
  d = c(0.5, 0.75, 1.01, 1.4, 1.9, 2.2, 2.6),
  antsep = seq(0, to = 20, by = 0.25),
  dz = 0.25,
  zmax = 250,
  fc = 100,
  lw = 15,
  qw = 0.9
)
}
\arguments{
\item{vint}{(\code{numeric[n]}) internal velocity of the layers}

\item{d}{(\code{numeric[n]}) thickness of the layers}

\item{antsep}{(\code{numeric[m]}) antenna separations use for the data acquisition}

\item{dz}{(\code{numeric[1]}) time sampling (ns)}

\item{zmax}{(\code{numeric[1]}) maximum time (ns)}

\item{fc}{(\code{numeric[1]}) Center frequency in MHz}

\item{lw}{(\code{numeric[1]}) wavelet duration (ns)}

\item{qw}{(\code{numeric[1]}) Damping factor for the wavelet, where 0 < q < 1}
}
\value{
(\code{GPR}) CMP data
}
\description{
Simulates CMP data.
}
