% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spAngle.R
\name{spAngle}
\alias{spAngle}
\alias{spAngle,GPR-method}
\alias{spAngle,GPRsurvey-method}
\title{Angle of the GPR data}
\usage{
spAngle(x)

\S4method{spAngle}{GPR}(x)

\S4method{spAngle}{GPRsurvey}(x)
}
\arguments{
\item{x}{[\code{GPR|GPRsurvey}]}
}
\value{
[\code{numeric(1)}] The angle of the oriented bounding box.
}
\description{
The angle is computed based on the orientation of the oriented bounding
box (\code{\link{spOBB}}).
}
