% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spBuffer.R
\docType{methods}
\name{spBuffer}
\alias{spBuffer}
\alias{spBuffer,GPR-method}
\alias{spBuffer,GPRsurvey-method}
\alias{spBuffer,sfc-method}
\alias{spBuffer,sf-method}
\title{Buffer the GPR lines}
\usage{
spBuffer(x, d, combine = TRUE)

\S4method{spBuffer}{GPR}(x, d)

\S4method{spBuffer}{GPRsurvey}(x, d, combine = TRUE)

\S4method{spBuffer}{sfc}(x, d, combine = TRUE)

\S4method{spBuffer}{sf}(x, d, combine = TRUE)
}
\arguments{
\item{x}{[\code{GPR|GPRsurvey}] An object of the class \code{GPR} or
\code{GPRsurvey}}

\item{d}{[\code{numeric(1)}] buffer distance}

\item{combine}{[\code{logical(1)}] If \code{TRUE}, returns the buffer for
all the GPR lines together. Otherwise it returns the buffer
for each line.}
}
\value{
[\code{sfc}] Polygon as a simple feature geometry list-column.
}
\description{
Returns the buffered lines as polygon(s)
}
