% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spConvexHull.R
\docType{methods}
\name{spConvexHull}
\alias{spConvexHull}
\alias{spConvexHull,GPRsurvey-method}
\alias{spConvexHull,GPR-method}
\alias{spConvexHull,matrix-method}
\title{Convex hull of GPR data}
\usage{
spConvexHull(x, verbose = FALSE)

\S4method{spConvexHull}{GPRsurvey}(x, verbose = FALSE)

\S4method{spConvexHull}{GPR}(x, verbose = FALSE)

\S4method{spConvexHull}{matrix}(x, verbose = FALSE)
}
\arguments{
\item{x}{[\code{GPR|GPRsurvey}]}

\item{verbose}{[\code{logical(1)}] If \code{FALSE}, all messages and warnings 
are suppressed (use with care).}
}
\description{
Return the convex hull.
}
