% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{spec}
\alias{spec}
\title{Return the amplitude spectrum of the GPR object.}
\usage{
\S4method{spec}{GPR}(x, type = c("f-x", "f-k"), plotSpec = TRUE,
  unwrapPhase = TRUE, ...)
}
\description{
Return the amplitude spectrum of the GPR object.
}

