% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec1D.R
\name{spec1D,GPR-method}
\alias{spec1D,GPR-method}
\title{Frequency spectrum the traces (1D)}
\usage{
\S4method{spec1D}{GPR}(x, MARGIN = 2, plotSpec = TRUE, unwrapPhase = TRUE, ...)
}
\description{
Based on the function `fft()` of R base
}
