% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec2D.R
\docType{methods}
\name{spec2D,GPR-method}
\alias{spec2D,GPR-method}
\title{Frequency spectrum the traces (1D)}
\usage{
\S4method{spec2D}{GPR}(x, plotSpec = TRUE, unwrapPhase = TRUE, ...)
}
\description{
Based on the function `fft()` of R base
}
