% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spExtractFromRaster.R
\name{spExtractFromRaster}
\alias{spExtractFromRaster}
\title{Extract elevation from raster data}
\usage{
\S4method{spExtractFromRaster}{GPRsurvey}(x, r, method = "bilinear", ...)
}
\description{
Based on the function \code{extract} from \code{raster} package.
Check arguments and help for this function
}
