% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spGeoref.R
\docType{methods}
\name{spGeoref}
\alias{spGeoref}
\alias{spGeoref,GPRsurvey-method}
\title{Rotate coordinates of the GPR traces}
\usage{
spGeoref(x, alpha = NULL, cloc = NULL, creg = NULL, ploc = NULL,
  preg = NULL, FUN = mean)

\S4method{spGeoref}{GPRsurvey}(x, alpha = NULL, cloc = NULL,
  creg = NULL, ploc = NULL, preg = NULL, FUN = mean)
}
\arguments{
\item{x}{[\code{GPR|GPRsurvey}]}

\item{alpha}{[\code{numeric(1)}] The rotation angle in radians. 
If \code{alpha = NULL}, \code{alpha} is estimated from the 
pairs of points in the local reference system (\code{ploc}) 
and in the regional reference system (\code{preg}).}

\item{cloc}{[\code{numeric(2)}] the coordinate
center of the local reference system}

\item{creg}{[\code{numeric(2|3)}] the coordinate
center of the regional reference system. Setting 
\code{creg = NULL} (default) is equivalent to apply a rotation
of angle \code{alpha} and center \code{cloc}.}

\item{ploc}{[\code{numeric(2|3)}] the first two columns corresponding}

\item{preg}{[\code{numeric(2|3)}]}

\item{FUN}{[\code{function}]}
}
\description{
Rotate coordinates of the GPR traces...
}
