% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spInterp.R
\docType{methods}
\name{spInterp}
\alias{spInterp}
\alias{spInterp,GPR-method}
\alias{spInterp,GPRsurvey-method}
\title{Interpolate trace positions from measurement (e.g., GPS).}
\usage{
spInterp(x, topo, tt = NULL, r = NULL, lonlat = FALSE,
  projToUTM = FALSE, CRSin = NULL, CRSout = NULL, odometer = FALSE,
  tol = NULL, verbose = TRUE, plot = TRUE, method = c("linear",
  "linear", "linear"))

\S4method{spInterp}{GPR}(x, topo, tt = NULL, r = NULL,
  lonlat = FALSE, projToUTM = FALSE, CRSin = NULL, CRSout = NULL,
  odometer = FALSE, tol = NULL, verbose = TRUE, plot = TRUE,
  method = c("linear", "linear", "linear"))

\S4method{spInterp}{GPRsurvey}(x, topo, tt = NULL, r = NULL,
  lonlat = FALSE, projToUTM = FALSE, CRSin = NULL, CRSout = NULL,
  odometer = FALSE, tol = NULL, verbose = TRUE, plot = TRUE,
  method = c("linear", "linear", "linear"))
}
\arguments{
\item{x}{[\code{GPR|GRPsurvey}]}

\item{topo}{[\code{matrix|list}] A numeric matrix or a list of matrix 
(if \code{x} is a \code{GPRsurvey} object) with at least
3 columns (the (x,y,z)-coordinates) and optionally a fourth
column containing the trace number (id).}

\item{tt}{[\code{numeric|list}] A numeric vector or a list of vectors
(if \code{x} is a \code{GPRsurvey} object) corresponding to the
trace recording time (e.g., output GPS device). If provided,
it will be assumed that the traces were recorded at regular
time interval (the 4th column of \code{topo} will be ignored).}

\item{r}{[\code{raster::RasterLayer}]  A 'RasterLayer' 
object from the package 
'raster' from which trace elevation \code{z} will be extracted 
based on the trace position \code{(x, y)} on the raster. 
The extracted trace elevation will overwrite the values from the 
third column of \code{topo}.}

\item{lonlat}{[\code{logical(1)}] Are the coordinates in geographic WGS84 
lat/long coordinate reference system (e.g., GPS data)?}

\item{projToUTM}{[\code{logical(1)}] Should the geographic coordinates be
projected into the corresponding UTM WGS84 coordinate 
reference system (only if \code{lonlat = TRUE}).
RGPR will guess the correct UTM WGS84 project.}

\item{CRSin}{[\code{character(1)|integer(1)}] CRS of \code{topo}, only used
if \code{lonlat = FALSE}. Either a valid proj4string character
or a valid EPSG integer value 
(see \code{\link[rgdal]{make_EPSG}}).}

\item{CRSout}{[\code{character(1)|integer(1)}] CRS to which the trace 
coordinates must be projected. Either a valid proj4string 
character or a valid EPSG integer value 
(see \code{\link[rgdal]{make_EPSG}}).}

\item{odometer}{[\code{logical(1)}] Was the data collection triggered by
an odometer? If yes, then 
\code{odometer = TRUE}.}

\item{tol}{[\code{numeric(1)}]     Length-one numeric vector: if the horizontal distance between 
two consecutive trace positions is smaller than \code{tol}, 
then the traces in between as well as the second trace
position are removed.
If \code{tol = NULL}, \code{tol} is set equal to
\code{sqrt(.Machine$double.eps)}.}

\item{verbose}{[\code{logical(1)}] If \code{FALSE}, all messages and warnings are
suppressed (use with care).}

\item{plot}{[\code{logical(1)}] If \code{TRUE} some control/diagnostic 
plots are displayed.}

\item{method}{[\code{character(3)}] The interpolation
methods:
First element for the interpolation of the 
inter-trace distances, 
second element for the interpolation of the horizontal 
trace positions, and third element for the interpolation
of the vertical trace positions.
The methods that can be used are 
\code{linear}, \code{nearest}, \code{pchip}, \code{cubic}, 
and \code{spline}
 (same methods as in \code{\link{signal}{interp1}}.}
}
\value{
x with interpolated trace positions.
}
\description{
Interpolate trace positions from measurement (e.g., GPS).
}
