% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spInterpToRaster.R
\name{spInterpToRaster}
\alias{spInterpToRaster}
\title{Interpolate trace elevation to raster}
\usage{
\S4method{spInterpToRaster}{GPRsurvey}(x, nx = 100, ny = 100, h = 5, ...)

\S4method{spInterpToRaster}{matrix}(x, nx = 100, ny = 100, h = 5, CRSobj, ...)
}
\description{
Interpolate trace elevation to raster
}
