% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spIntersect.R
\docType{methods}
\name{spIntersect}
\alias{spIntersect}
\alias{spIntersect,GPRsurvey-method}
\title{Compute GPR profile intersections}
\usage{
spIntersect(x)

\S4method{spIntersect}{GPRsurvey}(x)
}
\arguments{
\item{x}{[\code{object class GPRsurvey}] An object of the class \code{GPRsurvey}}
}
\value{
[\code{object GPRsurvey}] An object of the class GPRsurvey.
}
\description{
Compute GPR profile intersections
}
\details{
Modified slots
\itemize{
  \item \code{intersects}: trace shifted. The number of rows of data may 
        be smaller if \code{crop = TRUE}.
}
}
