% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spIntersection.R
\name{spIntersection}
\alias{spIntersection}
\alias{spIntersection,GPRsurvey-method}
\title{Compute GPR profile intersections}
\usage{
spIntersection(x)

\S4method{spIntersection}{GPRsurvey}(x)
}
\arguments{
\item{x}{[\code{object class GPRsurvey}] An object of the class \code{GPRsurvey}}
}
\value{
[\code{object GPRsurvey}] An object of the class GPRsurvey.
}
\description{
Compute GPR profile intersections
}
\details{
Modified slots
\itemize{
  \item \code{intersects}: trace shifted. The number of rows of data may 
        be smaller if \code{crop = TRUE}.
}
}
