% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spPathRel.R
\name{spPathRel}
\alias{spPathRel}
\alias{spPathRel,GPR-method}
\title{Relative positions along path}
\usage{
spPathRel(x)

\S4method{spPathRel}{GPR}(x)
}
\arguments{
\item{x}{[\code{GPR|GPRsurvey}] An object of the class \code{GPR} or
\code{GPRsurvey}}
}
\value{
[\code{numeric{m}}] Relative trace position along GPR line.
}
\description{
Relative positions along path
}
