% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spProjectToCRS.R
\docType{methods}
\name{spProjectToCRS}
\alias{spProjectToCRS}
\alias{spProjectToCRS,GPR-method}
\alias{spProjectToCRS,GPRsurvey-method}
\title{Trace projection}
\usage{
spProjectToCRS(x, CRSobj)

\S4method{spProjectToCRS}{GPR}(x, CRSobj)

\S4method{spProjectToCRS}{GPRsurvey}(x, CRSobj)
}
\arguments{
\item{x}{Object of the class GPR}

\item{CRSobj}{object of class \link{CRS}, or of class \code{character} in
which case it is converted to \link{CRS}.}
}
\description{
Project the trace coordinates give a coordinate reference system.
}
\details{
Modified slots
\itemize{
  \item \code{coord} the trace coordinates
  \item \code{x} the local trace position (along profile)
  \item \code{crs} the coordinate reference system.
  \item \code{spunit} the spatial units are updated accroding to the new
                      coordinate reference system.
}
}
