% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spunit.R
\name{spunit}
\alias{spunit}
\alias{spunit<-}
\alias{spunit<-,GPR-method}
\alias{spunit,GPR-method}
\alias{spunit,GPRsurvey-method}
\alias{spunit<-,GPRsurvey-method}
\title{Spatial unit}
\usage{
spunit(x)

spunit(x) <- value

\S4method{spunit}{GPR}(x)

\S4method{spunit}{GPR}(x) <- value

\S4method{spunit}{GPRsurvey}(x)

\S4method{spunit}{GPRsurvey}(x) <- value
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}

\item{value}{[\code{character(1)}] Spatial units: "m", "feet", etc.
see units package.}
}
\value{
[\code{GPR class}] An object of the class \code{GPR}
}
\description{
Spatial unit of the trace coordinates
}
