% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{strTensor}
\alias{strTensor}
\title{Structure tensor field of GPR data}
\usage{
\S4method{strTensor}{GPR}(x, blksze = c(2, 4), kBlur = list(n = 1, m = 1, sd
  = 1), kEdge = list(n = 5, m = 5, sd = 1), kTensor = list(n = 5, m = 5, sd
  = 1), thresh = 0.02, what = c("tensor", "mask"), ...)
}
\description{
Structure tensor field of GPR data
}

