% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{strucTensor}
\alias{strucTensor}
\title{Structure tensor of GPR data}
\usage{
strucTensor(P, blksze = c(5, 10), thresh = 0.1, winEdge = c(7, 7),
  winBlur = c(3, 3), winTensor = c(5, 10), sdTensor = 2, ...)
}
\description{
Structure tensor of GPR data
}

