% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_GPR.R
\docType{methods}
\name{[,GPR,ANY,ANY,ANY-method}
\alias{[,GPR,ANY,ANY,ANY-method}
\title{Extract parts of a GPR object}
\usage{
\S4method{[}{GPR,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{[\code{GPR}]}

\item{i}{[\code{integer}] Indices specifying elements to extract or replace.}

\item{j}{[\code{integer}] Indices specifying elements to extract or replace.}

\item{...}{Not used.}

\item{drop}{Not used.}
}
\value{
[\code{GPR}|\code{numeric}] Returns a numeric vector
       only if \code{x[]}.
}
\description{
Extract parts of a GPR object
}
