% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_GPR.R
\docType{methods}
\name{[<-,GPR,ANY,ANY,ANY-method}
\alias{[<-,GPR,ANY,ANY,ANY-method}
\title{Replace parts of a GPR object}
\usage{
\S4method{[}{GPR,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{[\code{GPR}]}

\item{i}{[\code{integer}] Indices specifying elements to extract or replace.}

\item{j}{[\code{integer}] Indices specifying elements to extract or replace.}

\item{...}{Not used.}

\item{value}{[\code{numeric}] Value to set.}
}
\value{
[\code{GPR}|\code{numeric}] Returns a numeric vector
       only if \code{x[]}.
}
\description{
Replace parts of a GPR object
}
