% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\docType{methods}
\name{summary,GPR-method}
\alias{summary,GPR-method}
\title{Summary of a GPR object}
\usage{
\S4method{summary}{GPR}(object, ...)
}
\arguments{
\item{object}{[\code{GPR}]}

\item{...}{Additional parameters to be passed (see \code{\link{summary}}).}
}
\description{
Summary of all the values contained in \code{object}.
}
