% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\name{summary,GPRvirtual-method}
\alias{summary,GPRvirtual-method}
\title{Summary of a GPR object}
\usage{
\S4method{summary}{GPRvirtual}(object, ...)
}
\arguments{
\item{object}{(\code{GPR})}

\item{...}{Additional parameters to be passed (see \code{\link[base:summary]{base::summary()}}).}
}
\description{
Summary of all the values contained in \code{object}.
}
\concept{statistics}
