% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupSummary_GPRvirtual.R
\name{Summary,GPRvirtual-method}
\alias{Summary,GPRvirtual-method}
\alias{Summary-GPRvirtual-method}
\title{Basic summary functions}
\usage{
\S4method{Summary}{GPRvirtual}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An object of the class GPRvirtual}

\item{...}{further arguments}

\item{na.rm}{[\code{logical(1)}] should missing values be removed?}
}
\description{
Methods for the base Summary methods \link[methods]{S4groupGeneric}
}
\details{
Currently implemented methods include:
\itemize{
 \item{all, any, sum, prod, min, max, range}
 }
}
