% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/getterSetter.R
\name{surveymode}
\alias{surveymode}
\alias{surveymode<-}
\title{Survey mode of the GPR data}
\usage{
\S4method{surveymode}{GPR}(x)

\S4method{surveymode}{GPR}(x) <- value
}
\description{
Survey mode of the GPR data
}
