% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R, R/ClassGPRsurvey.R
\docType{methods}
\name{svAngle}
\alias{svAngle}
\title{Angle of the GPR survey}
\usage{
\S4method{svAngle}{GPR}(x)

\S4method{svAngle}{GPR}(x)
}
\description{
Angle of the GPR survey
}
