% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/getterSetter.R
\docType{methods}
\name{svDate}
\alias{svDate}
\alias{svDate<-}
\title{Survey date}
\usage{
\S4method{svDate}{GPR}(x)

\S4method{svDate}{GPR}(x) <- value
}
\arguments{
\item{x}{An object of the class 'GPR'}

\item{value}{An object of the class 'Date'}
}
\description{
Return NULL if no date exists, else an object of the class 'Date'
}
