% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchPolarity.R
\name{switchPolarity}
\alias{switchPolarity}
\alias{switchPolarity,GPR-method}
\alias{switchPolarity,GPRsurvey-method}
\title{Switch polarity}
\usage{
switchPolarity(obj, ..., track = TRUE)

\S4method{switchPolarity}{GPR}(obj, track = TRUE)

\S4method{switchPolarity}{GPRsurvey}(obj, sel, track = TRUE)
}
\arguments{
\item{obj}{(\code{GPR}) An object of the class GPR.}

\item{track}{(\code{logical[1]}) Should the processing step be tracked?}
}
\value{
(\code{GPR}) An object of the class GPR whose traces are dewowed.
}
\description{
Switch polarity
}
\concept{processing}
