% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time0.R
\name{time0}
\alias{time0}
\alias{time0<-}
\alias{setTime0}
\alias{time0,GPR-method}
\alias{time0<-,GPR-method}
\alias{setTime0,GPR-method}
\title{Return time-zero}
\usage{
time0(obj)

time0(obj) <- value

setTime0(obj, t0, track = TRUE)

\S4method{time0}{GPR}(obj)

\S4method{time0}{GPR}(obj) <- value

\S4method{setTime0}{GPR}(obj, t0, track = TRUE)
}
\arguments{
\item{obj}{(\code{GPR* object})}

\item{value}{(\code{numeric[n]}) Time-zero with \code{n = 1} or \code{n = ncol(obj)}}

\item{t0}{(\code{numeric[n]}) Time-zero with \code{n = 1} or \code{n = ncol(obj)}}

\item{track}{(\code{logical[1]}) Should the processing step be tracked?}
}
\value{
A vector containing the time-zero values of each traces.
}
\description{
\code{time0} returns the 'time-zero' of every traces. Generally,
'time-zero' corresponds to the first wave arrival (also called first wave
break).
}
\seealso{
\code{\link[=pickFirstBreak]{pickFirstBreak()}}  to estimate the first wave break.
}
