% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{time0Cor}
\alias{time0Cor}
\title{Time zero correction}
\usage{
\S4method{time0Cor}{GPR}(x, method = c("none", "linear", "nearest", "pchip",
  "cubic", "spline"), keep = NULL, crop = TRUE, c0 = 0.299)
}
\arguments{
\item{x}{A object of the class GPR}

\item{keep}{A length-one numeric vector indicating how much of the trace...}

\item{crop}{If TRUE (defaults), remove the rows containing only zero's 
(no data). If FALSE,}

\item{ts}{A numeric vector defining the amount of time the traces have to
shifted}
}
\description{
\code{time0Cor} shift the traces vertically
}

