% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{time0Estimation}
\alias{time0Estimation}
\title{Estimate and set time-zero}
\usage{
\S4method{time0Estimation}{GPR}(x, method = c("coppens", "coppens2",
  "threshold", "MER"), thr = 0.12, w = 11, ns = NULL, bet = NULL,
  c0 = 0.299, FUN)
}
\description{
Estimate and set time-zero
}
