% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R, R/ClassGPRsurvey.R
\name{tpOBB2D}
\alias{tpOBB2D}
\title{Oriented bounding box (2D)}
\source{
source "whuber" from stackexchange.com, 
https://gis.stackexchange.com/questions/22895/finding-minimum-area-rectangle-for-given-points/181883#181883
}
\usage{
\S4method{tpOBB2D}{GPR}(x)

\S4method{tpOBB2D}{GPRsurvey}(x)
}
\description{
Returns the oriented bounding box of the trace position of the survey.
}
\details{
The algorithm you are looking for is known in polygon generalisation as 
"smallest surrounding rectangle".
Compute the convex hull of the cloud.
For each edge of the convex hull:
compute the edge orientation (with arctan),
rotate the convex hull using this orientation in order to compute easily 
the bounding rectangle area with min/max of x/y of the rotated convex hull,
Store the orientation corresponding to the minimum area found,
Return the rectangle corresponding to the minimum area found.
In 3D, the same applies, except:
  The convex hull will be a volume,
  The orientations tested will be the orientations (in 3D) of the convex hull faces.
}
