% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/ClassGPR.R
\docType{methods}
\name{tpPos}
\alias{tpPos}
\alias{tpPos<-}
\alias{tpPos}
\alias{tpPos<-}
\title{Position of the GPR traces}
\usage{
\S4method{tpPos}{GPR}(x)

\S4method{tpPos}{GPR}(x) <- value
}
\description{
Position of the GPR traces
}
