% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRsurvey.R
\docType{methods}
\name{tpShift}
\alias{tpShift}
\title{Shift trace positions of one GPR data}
\usage{
\S4method{tpShift}{GPRsurvey}(x, i, dx = 0, dy = 0)
}
\description{
Shift trace positions of GPR data \code{i} by \code{dx} along x-axis and
by \code{dy} along y-axis.
}
