% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/ClassGPRvirtual.R
\docType{methods}
\name{tpUnit}
\alias{tpUnit}
\alias{tpUnit<-}
\alias{tpUnit}
\alias{tpUnit<-}
\title{Position unit of the GPR data}
\usage{
\S4method{tpUnit}{GPRvirtual}(x)

\S4method{tpUnit}{GPRvirtual}(x) <- value
}
\description{
Position unit of the GPR data
}
