% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{traceAverage}
\alias{traceAverage}
\title{Trace average}
\usage{
\S4method{traceAverage}{GPR}(x, w = NULL, FUN = mean, ...)
}
\arguments{
\item{x}{An object of the class GPR}

\item{w}{A window length defining width of the average window. Default value
is \code{NULL} and with this default value the function returns a 
single trace that is the average trace of the radargram.}

\item{FUN}{A function to compute the average (default is \code{mean})}

\item{...}{Additional parameters for the FUN functions}
}
\description{
Compute the average trace of a radargram (resulting in a single trace) or
a moving average of the trace.
}

