% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{traceScaling}
\alias{traceScaling}
\title{Trace scaling}
\usage{
\S4method{traceScaling}{GPR}(x, type = c("stat", "min-max", "95", "eq", "sum",
  "rms"))
}
\description{
Trace scaling
}

