% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{traceShift}
\alias{traceShift}
\title{Shift the traces}
\usage{
\S4method{traceShift}{GPR}(x, ts, keep = 0, method = c("linear", "nearest",
  "pchip", "cubic", "spline"), crop = TRUE)
}
\arguments{
\item{x}{A object of the class GPR}

\item{ts}{A numeric vector defining the amount of time the traces have to
shifted}

\item{keep}{A length-one numeric vector indicating how much of the trace...}

\item{crop}{If TRUE (defaults), remove the rows containing only zero's 
(no data). If FALSE,}
}
\description{
Shift the traces
}

