% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_string.R
\name{trimStr}
\alias{trimStr}
\title{Trim string}
\usage{
trimStr(x)
}
\arguments{
\item{x}{(\code{character}) a string or a vector of strings (characters)}
}
\description{
returns string w/o leading or trailing whitespace
}
