% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRsurvey.R
\docType{methods}
\name{trOBB2D}
\alias{trOBB2D}
\title{Oriented bounding box (2D)}
\usage{
\S4method{trOBB2D}{GPRsurvey}(x)
}
\description{
Returns the oriented bounding box of the trace position of the survey.
}
