% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\name{trPlot}
\alias{trPlot}
\title{Plot all the traces in one 1D plot}
\usage{
trPlot(x, ...)

\S4method{trPlot}{GPR}(x, ...)
}
\arguments{
\item{x}{Object of the class GPR}

\item{...}{Arguments to be passed to \code{plot}/\code{line}}
}
\description{
Plot all the traces in one 1D plot
}
