% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\name{trProject}
\alias{trProject}
\title{Trace projection}
\usage{
\S4method{trProject}{GPR}(x, CRSobj)
}
\arguments{
\item{x}{Object of the class GPR}

\item{CRSobj}{[\code{character(1)}] A string accepted by GDAL 
(e.g., \code{"EPSG:2056"}, WKT-string).}
}
\description{
Project the trace coordinates give a coordinate reference system.
}
