% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{trRmDuplicates}
\alias{trRmDuplicates}
\alias{trRmDuplicates}
\title{Remove traces with duplicated trace positions}
\usage{
\S4method{trRmDuplicates}{GPR}(x, tol = NULL)
}
\arguments{
\item{x}{An object of the class GPR}

\item{tol}{Length-one numeric vector: if the horizontal distance between two 
consecutive traces is smaller than \code{tol}, then
the second trace is removed.
If \code{tol = NULL}, \code{tol} is set equal to
\code{sqrt(.Machine$double.eps)}.}
}
\description{
Remove traces with duplicated trace positions
}
