% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getterSetter.R
\name{trTime,GPR-method}
\alias{trTime,GPR-method}
\title{#' Velocity model of the GPR data
#' 
#' @name vel
#' @rdname vel
#' @export
setMethod("vel", "GPR", function(x){
  if(length(x@vel) == 1){
    return(x@vel[[1]])
  }else{
    return(x@vel)
  }
})}
\usage{
\S4method{trTime}{GPR}(x)
}
\description{
#' @name vel<-
#' @rdname vel
#' @export
setGenericVerif("vel<-",function(x,value){standardGeneric("vel<-")})
}
\details{
#' @name vel<-
#' @rdname vel
#' @export
setReplaceMethod(
  f="vel",
  signature="GPR",
  definition=function(x, value){
    if(typeof(value) != "list"){
      value <- list(as.numeric(value))
    }
    x@vel <- value
    x@proc <- c(x@proc, "vel<-")
    return(x)
})
#' @name setVel
#' @rdname vel
#' @export
setGenericVerif("setVel", function(x, v) standardGeneric("setVel"))

#' @name setVel
#' @rdname vel
#' @export
setMethod("setVel", "GPR", function(x, v){
 vel(x) <- v 
 return(x)
})
}
