% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/ClassGPRvirtual.R
\docType{methods}
\name{trZ}
\alias{trZ}
\alias{trZ<-}
\alias{trZ}
\alias{trZ<-}
\title{Depth/time of the GPR data}
\usage{
\S4method{trZ}{GPRvirtual}(x)

\S4method{trZ}{GPRvirtual}(x) <- value
}
\description{
Depth/time of the GPR data
}
