% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/ClassGPR.R
\docType{methods}
\name{trZUnit}
\alias{trZUnit}
\alias{trZUnit<-}
\alias{trZUnit}
\alias{trZUnit<-}
\title{Depth unit of the GPR data}
\usage{
\S4method{trZUnit}{GPR}(x)

\S4method{trZUnit}{GPR}(x) <- value
}
\description{
Depth unit of the GPR data
}
