% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS.R
\name{UMTStringToEPSG}
\alias{UMTStringToEPSG}
\title{EPGS code from UTM zone string}
\usage{
UMTStringToEPSG(x)
}
\arguments{
\item{x}{(\code{character[1]}) The EPSG code string (e.g. 32N).}
}
\value{
(\code{integer[1]}) The EPSG code.
}
\description{
Returns the EPSG code from UTM zone string (e.g., '32N'). EPSG code is:
32600+zone for positive latitudes and  32700+zone for negatives latitudes.
}
\concept{CRS}
