% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_processing.R
\name{unscale}
\alias{unscale}
\title{Unscale}
\usage{
unscale(x, y)
}
\arguments{
\item{x}{(\code{numeric[n]}) A numerical vector}

\item{y}{(\code{numeric[n]}) A numerical vector}
}
\value{
(\code{numeric[n]})
}
\description{
Back-transform/unscale from \code{scale}
}
