% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample.R
\name{upsample}
\alias{upsample}
\title{Up-sample the GPR data (1D and 2D sinc-interpolation)}
\usage{
\S4method{upsample}{GPR}(x, n)
}
\arguments{
\item{x}{[\code{GPR class}]}

\item{n}{[\code{integer(1|2)}] Two positive integer values. The GPR data are
upsampled in the time/depth domain by a factor of \code{n[1]} and 
in the spatial domain (trace sampling) by a factor of \code{n[2]}. 
If only one value is passed, then \code{n[2] <- n[1]}.}
}
\description{
When upsampling is performed on a GPR data, it produces an approximation of 
the data that would have been collected by sampling them at a higher rate in
time/depth domain and/or in spatial domain (trace sampling).
}
