% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS.R
\name{UTMToEPSG}
\alias{UTMToEPSG}
\title{EPGS code from UTM zone}
\usage{
UTMToEPSG(zone, south = FALSE)
}
\arguments{
\item{zone}{(\code{integer[1]}) the UTM zone.}

\item{south}{(\code{integer[1]}) \code{TRUE} if the UTM is located in
southern hemisphere.}
}
\value{
(\code{integer[1]}) The EPSG code.
}
\description{
Returns the EPSG code from UTM zone. EPSG code is:
32600+zone for positive latitudes and  32700+zone for negatives latitudes.
}
\concept{CRS}
