% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPS.R
\name{UTMTolonlat}
\alias{UTMTolonlat}
\title{UTM to latitude-longitude}
\usage{
UTMTolonlat(xy, CRSobj = NULL)
}
\arguments{
\item{xy}{(\code{matrix[,2]}) Columns = x and y coordinates.}

\item{CRSobj}{(\code{character[1]}) Coordinate reference system
(proj4string)}
}
\value{
a 2-column-matrix (longitude N, latitude (E))
}
\description{
UTM to latitude-longitude
}
\concept{CRS}
