% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{values}
\alias{values}
\alias{values<-}
\title{Values of the GPR data}
\usage{
\S4method{values}{GPR}(x)

\S4method{values}{GPR}(x) <- value
}
\description{
Values of the GPR data
}
