% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getterSetter.R
\docType{methods}
\name{vel}
\alias{vel}
\alias{vel<-}
\alias{setVel}
\title{Velocity model of the GPR data}
\usage{
\S4method{vel}{GPR}(x)

\S4method{vel}{GPR}(x) <- value

\S4method{setVel}{GPR}(x, v)
}
\description{
Velocity model of the GPR data
}
