% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velDix.R
\name{velDix}
\alias{velDix}
\title{Computes Dix velocities}
\usage{
velDix(twt, v)
}
\arguments{
\item{twt}{(\code{numeric[n]}) $n$ two-way travel times ($n > 0$)}

\item{v}{[\code{numeric[n]}] $n$ corresponding velocities}
}
\value{
(\code{list}) List with two elements: \code{t} (time) and
\code{v} (corresponding Dix velocities)
}
\description{
Computes Dix velocities from root-mean-square velocity and corresponding
two-way travel times
}
