% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velInterp.R
\name{velInterp}
\alias{velInterp}
\alias{velInterp,GPR-method}
\title{Interpolate velocity model}
\usage{
velInterp(
  x,
  type = c("vrms", "vint"),
  method = c("stairs", "linear", "nearest", "pchip", "cubic", "spline")
)

\S4method{velInterp}{GPR}(
  x,
  type = c("vrms", "vint"),
  method = c("stairs", "linear", "nearest", "pchip", "cubic", "spline")
)
}
\arguments{
\item{x}{(\verb{GPR class}) An object of the class \code{GPR}}

\item{type}{(\code{character[1]}) Which type of velocity values has to be
updated? The root-mean-square velocity (\code{vrms}) or the
internal velocity (\code{vint})?}

\item{method}{(\code{character[1]}) Interpolation method to be applied:
one of \code{stairs}, \code{linear}, \code{nearest},
\code{pchip}, \code{spline}, \code{cubic}
(see also \code{\link[signal:interp1]{signal::interp1()}}).}
}
\value{
(\verb{GPR class}) An object of the class GPR.
}
\description{
Define the interpolation parameters that will be used when the velocities
will be plotted or used in other functions.
To undo the interpolation, simply set \code{method = stairs}.
}
