% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velLayers.R
\name{velLayers}
\alias{velLayers}
\title{Set layer velocities}
\usage{
velLayers(x, v, twt = NULL, method = "pchip", clean = TRUE)

\S4method{velLayers}{GPR}(x, v, twt = NULL, method = "pchip", clean = TRUE)
}
\arguments{
\item{x}{GPR object}

\item{v}{\link{\code{numeric}} velocities (length equal number of delineations plus one)}

\item{twt}{vertical resolution of the migrated data}

\item{method}{\link{\code{character(1)}} interpolation method.
One of \code{linear}, \code{nearest}, \code{pchip},
\code{cubic}, \code{spline}.}
}
\description{
Given delineation or the output of the function... and velocity values,
set 2D velocity model
}
