% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velPick.R
\name{velPick}
\alias{velPick}
\alias{velPick,GPR-method}
\title{Pick velocity interactively}
\usage{
velPick(x, ...)

\S4method{velPick}{GPR}(x, ...)
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}

\item{...}{additional graphics parameters used if type != "n" for plotting the locations}
}
\value{
[\code{GPR class}] An object of the class \code{GPR}.
}
\description{
Pick velocity interactively on a previously plotted velocity spectrum. The
picked velocity are NMO velocities which are a good approximation for
RMS velocities. The internal velocities are estimated from the NMO 
velocities through the Dix's formula \code{velPick()} add the NMO velocities
as RMS velocity as well as the internal velocities to \code{x}.
}
\details{
Application of the Dix's formula can provide non-real velocities, 
if the travel time intervals are small or
if the NMO velocity change is large. In this case, this
}
