% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velSetLayers.R
\name{velSetLayers}
\alias{velSetLayers}
\alias{velSetLayers,GPR-method}
\title{Set layer velocities}
\usage{
velSetLayers(obj, v, twt = NULL, method = "pchip", clean = TRUE)

\S4method{velSetLayers}{GPR}(obj, v, twt = NULL, method = "pchip", clean = TRUE)
}
\arguments{
\item{obj}{(\code{GPR* object}) An object of the class \code{GPR}}

\item{v}{(\code{numeric[n]})  velocities (length \code{n} equal number of delineations plus one)}

\item{twt}{(FIXME)      Two-way travel time (optinal)}

\item{method}{(\code{character[1]})  interpolation method.
One of \code{linear}, \code{nearest}, \code{pchip},
\code{cubic}, \code{spline}.}

\item{clean}{(\code{logical[1]}) When the interface crosses, should the older interfaces be clipped by the new ones?}
}
\description{
Given delineation or the output of the function... and velocity values,
set 2D velocity model
}
