% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{wapply}
\alias{wapply}
\title{wapply: A faster (but less functional) "rollapply" for vector setups
April 23, 2013.
By A.N. Spiess, senior scientist at the Department of Andrology at the
University Hospital Hamburg-Eppendorf.
This is what turned out (wapply for "window apply").}
\usage{
wapply(x = NULL, width = NULL, by = NULL, FUN = NULL, ...)
}
\description{
wapply: A faster (but less functional) "rollapply" for vector setups
April 23, 2013.
By A.N. Spiess, senior scientist at the Department of Andrology at the
University Hospital Hamburg-Eppendorf.
This is what turned out (wapply for "window apply").
}
