% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{wapplyMat}
\alias{wapplyMat}
\title{windowing with centered window}
\usage{
wapplyMat(x = NULL, width = NULL, by = NULL, FUN = NULL,
  MARGIN = 1, ...)
}
\description{
based on wapply and modified by Manu.
centered moving window.
return a matrix of the same dimension than x.
some border effect at a distance < width/2 at the first and last col/row
}
