% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{wapplyMat2}
\alias{wapplyMat2}
\title{windowing with not centered window}
\usage{
wapplyMat2(x = NULL, width = NULL, by = NULL, FUN = NULL, MARGIN = 1, ...)
}
\description{
based on wapply and modified by Manu.
not centered moving window! start first row/column and 
stop when the extremity of the windwo reach the last row/column.
return a matrix of with smaller dimension than x (margin - 2*width)
}
