% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{wapplyRowC}
\alias{wapplyRowC}
\title{Wapply on the row of a matrix (windowed + CENTERED)}
\usage{
wapplyRowC(x = NULL, width = NULL, by = NULL, FUN = NULL, ...)
}
\description{
NOT CURRENTLY USED
mod by MANU
}
