% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPRsim.R
\name{wavGPR}
\alias{wavGPR}
\title{Simulate GPR wavelet}
\usage{
wavGPR(xt, q, fc)
}
\description{
GPR wavelet: Check chap 11 of Annan in the book Near-Surface Geophysics
"The radiated wavelet from a GPR system is a complicated function of the 
antenna construction and the electronics drive circuitry. For impulse 
style ultra wideband systems using short electric dipole antennas, 
a simple mathematical model is helpful for numerical simulation."
}
