% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window}
\alias{window}
\alias{window.GPRvirtual}
\alias{window.GPR}
\title{Subset GPR data}
\usage{
\method{window}{GPRvirtual}(x, xlim, ylim, zlim, ...)

\method{window}{GPR}(x, xlim, zlim, ...)
}
\arguments{
\item{x}{(\code{GPR})}

\item{xlim}{(\code{numeric[2]}) Data range along x-axis}

\item{ylim}{(\code{numeric[2]}) Data range along y-axis}

\item{zlim}{(\code{numeric[2]}) Data range along z-axis}

\item{...}{Not used.}
}
\value{
(\code{GPR})
}
\description{
Based on the units...
}
