% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRsurvey.R
\docType{methods}
\name{writeSurvey}
\alias{writeSurvey}
\title{Write GPRsurvey object}
\usage{
\S4method{writeSurvey}{GPRsurvey}(x, fPath, overwrite = FALSE)
}
\description{
Write GPRsurvey object
}
