% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeProfileVTK.R
\name{writeVTK}
\alias{writeVTK}
\alias{writeVTK,GPR-method}
\alias{writeVTK,GPRsurvey-method}
\title{Write the GPRsurvey objects in VTK format.}
\usage{
writeVTK(obj, dsn = NULL, overwrite = TRUE)

\S4method{writeVTK}{GPR}(obj, dsn = NULL, overwrite = FALSE)

\S4method{writeVTK}{GPRsurvey}(obj, dsn = NULL, overwrite = FALSE)
}
\arguments{
\item{obj}{Object of the class \code{GPR} or \code{GPRsurvey}}

\item{dsn}{Filepath (Length-one character vector). If \code{dsn = NULL},
the file will be save in the current working directory with
the name of obj (\code{name(obj)}) with the extension depending
of \code{format}.}

\item{overwrite}{Boolean. If \code{TRUE} existing files will be overwritten,
if \code{FALSE} an error will be thrown if the file(s)
already exist(s).}
}
\description{
Write the GPRsurvey objects in VTK format.
}
\seealso{
\code{\link[=readGPR]{readGPR()}}
}
