% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpos.R
\name{xpos}
\alias{xpos}
\alias{xpos<-}
\alias{xpos,GPR-method}
\alias{xpos<-,GPR-method}
\title{x-position}
\usage{
xpos(x)

xpos(x) <- value

\S4method{xpos}{GPR}(x)

\S4method{xpos}{GPR}(x) <- value
}
\arguments{
\item{x}{[\code{GPR}] An object of the class GPR.}

\item{value}{[\code{numeric(m)}]}
}
\value{
[\code{GPR}]
}
\description{
Antenna separation distance(s)
}
\details{
Modified slots
\itemize{
  \item \code{x} the x-position 
  \item \code{xpos} the antenna separation distance only for CMP/WARR (plot twt as a function of antenna separation)
}
}
