% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R, R/CCF.R
\docType{methods}
\name{ACF}
\alias{ACF}
\alias{CCF}
\title{Auto-correlation function}
\usage{
\S4method{ACF}{GPR}(x, lag.max = 100, type = c("correlation",
  "covariance", "partial"), MARGIN = 2, na.action = na.fail,
  demean = TRUE, ...)

\S4method{CCF}{GPR}(x, lag.max = 100, type = c("correlation",
  "covariance", "partial"), na.action = na.fail, demean = TRUE, ...)
}
\description{
Auto-correlation function

Auto-correlation function
}
