% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{ann}
\alias{ann}
\alias{ann<-}
\title{Annotations of the GPR data}
\usage{
\S4method{ann}{GPR}(x)

\S4method{ann}{GPR}(x, values) <- value
}
\description{
Annotations of the GPR data
}

