% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsep.R
\name{antsep}
\alias{antsep}
\alias{antsep<-}
\alias{antsep,GPR-method}
\alias{antsep<-,GPR-method}
\title{Antenna separation distance(s)}
\usage{
antsep(x)

antsep(x) <- value

\S4method{antsep}{GPR}(x)

\S4method{antsep}{GPR}(x) <- value
}
\arguments{
\item{x}{[\code{GPR}] An object of the class GPR.}

\item{value}{[\code{numeric(1)|numeric(m)}]}
}
\value{
[\code{GPR}]
}
\description{
Antenna separation distance(s)
}
\details{
Modified slots
\itemize{
  \item \code{antsep} the antenna separation distance
  \item \code{x} the x-position only for CMP/WARR (plot twt as a function of antenna separation)
}
}
\concept{getters/setters}
