% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRvirtual.R
\docType{methods}
\name{apply,GPRvirtual-method}
\alias{apply,GPRvirtual-method}
\title{Apply Functions Over Array Margins}
\usage{
\S4method{apply}{GPRvirtual}(X, MARGIN, FUN, ...)
}
\description{
Returns a vector or array or list of values obtained by applying a function 
to margins of an array or matrix.
}
\examples{
data(frenkeLine00)
apply(frenkeLine00, 2, mean)
}
