% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupArith.R, R/GroupArith_GPRslice.R,
%   R/GroupArith_GPRcube.R
\docType{methods}
\name{Arith,GPR,ANY-method}
\alias{Arith,GPR,ANY-method}
\alias{Arith}
\alias{Arith,GPR,GPR-method}
\alias{Arith,ANY,GPR-method}
\alias{Arith,GPRslice,ANY-method}
\alias{Arith,GPRslice,GPRslice-method}
\alias{Arith,ANY,GPRslice-method}
\alias{Arith,GPRcube,ANY-method}
\alias{Arith,GPRcube,GPRcube-method}
\alias{Arith,ANY,GPRcube-method}
\title{Basic arithmetical functions}
\usage{
\S4method{Arith}{GPR,ANY}(e1, e2)

\S4method{Arith}{GPR,GPR}(e1, e2)

\S4method{Arith}{ANY,GPR}(e1, e2)

\S4method{Arith}{GPRslice,ANY}(e1, e2)

\S4method{Arith}{GPRslice,GPRslice}(e1, e2)

\S4method{Arith}{ANY,GPRslice}(e1, e2)

\S4method{Arith}{GPRcube,ANY}(e1, e2)

\S4method{Arith}{GPRcube,GPRcube}(e1, e2)

\S4method{Arith}{ANY,GPRcube}(e1, e2)
}
\arguments{
\item{e1}{An object of the class GPR}

\item{e2}{An object of the class GPR}

\item{e1}{An object of the class GPRslice}

\item{e2}{An object of the class GPRslice}

\item{e1}{An object of the class GPRcube}

\item{e2}{An object of the class GPRcube}
}
\description{
Basic arithmetical functions

Basic arithmetical functions

Basic arithmetical functions
}
