% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRvirtual.R
\docType{methods}
\name{Arith}
\alias{Arith}
\alias{Arith}
\alias{Arith}
\title{Basic arithmetical functions}
\usage{
\S4method{Arith}{GPRvirtual,ANY}(e1, e2)

\S4method{Arith}{GPRvirtual,GPRvirtual}(e1, e2)

\S4method{Arith}{ANY,GPRvirtual}(e1, e2)
}
\arguments{
\item{e1}{An object of the class GPR, GPRstack or GPRcube.}

\item{e2}{An object of the class GPR, GPRstack or GPRcube}
}
\description{
Basic arithmetical functions
}
\examples{
data(frenkeLine00)
A <- exp(frenkeLine00)
B <- A + frenkeLine00
}
