% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{Arith}
\alias{Arith}
\title{Basic arithmetical functions}
\usage{
\S4method{Arith}{GPR,ANY}(e1, e2)

\S4method{Arith}{GPR,GPR}(e1, e2)

\S4method{Arith}{ANY,GPR}(e1, e2)
}
\arguments{
\item{e1}{An object of the class RGPR.}

\item{e2}{An object of the class RGPR.}
}
\description{
Basic arithmetical functions
}
\examples{
data(frenkeLine00)
A <- exp(frenkeLine00)
B <- A + frenkeLine00
}

