% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion_spatial.R
\docType{methods}
\name{as.SpatialLines}
\alias{as.SpatialLines}
\alias{as.SpatialLines,GPRsurvey-method}
\alias{as.SpatialLines,GPR-method}
\title{Coerce object to an object of the class SpatialLines}
\usage{
as.SpatialLines(x)

\S4method{as.SpatialLines}{GPRsurvey}(x)

\S4method{as.SpatialLines}{GPR}(x)
}
\arguments{
\item{x}{[\code{class GPR|GPRsurvey}]}
}
\value{
[\code{SpatialLines object}]
}
\description{
Coerce object to an object of the class SpatialLines
}
