% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion_GPRsurvey.R
\docType{methods}
\name{as.SpatialPoints}
\alias{as.SpatialPoints}
\alias{as.SpatialPoints,GPRsurvey-method}
\title{Coerce object to an object of the class SpatialPoints}
\usage{
as.SpatialPoints(x)

\S4method{as.SpatialPoints}{GPRsurvey}(x)
}
\arguments{
\item{x}{[\code{class GPR|GPRsurvey}]}
}
\value{
[\code{SpatialPoints object}] An object of the class GPR.
}
\description{
Coerce object to an object of the class SpatialPoints
}
