% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion_spatial.R
\name{as.spatialPoints}
\alias{as.spatialPoints}
\alias{as.spatialPoints,GPR-method}
\alias{as.spatialPoints,GPRsurvey-method}
\title{Coerce object to an object of the class SpatialPoints}
\usage{
as.spatialPoints(x)

\S4method{as.spatialPoints}{GPR}(x)

\S4method{as.spatialPoints}{GPRsurvey}(x)
}
\arguments{
\item{x}{[\code{class GPR|GPRsurvey}]}
}
\value{
[\code{SpatialPoints}]
}
\description{
Coerce object to an object of the class SpatialPoints
}
\concept{coercion}
