% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{bits2volt}
\alias{bits2volt}
\title{Bytes to volt conversion
      
Convert bytes to volt values}
\usage{
bits2volt(Vmax = 50, Vmin = 50, nbits = 16)
}
\arguments{
\item{Vmax}{length-one numeric vector: maximal nominal analog input voltage. 
If \code{Vmax = NULL} it returns \code{1} (no bytes to volt 
transformation)}

\item{Vmin}{length-one numeric vector: minimal nominal analog input voltage. 
If missing, then \code{Vmin = -Vmax}.}

\item{nbits}{Length-one numeric vector: number of bits}
}
\description{
Bytes to volt conversion
      
Convert bytes to volt values
}
