% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{byte2volt}
\alias{byte2volt}
\title{Bytes to volt conversion}
\usage{
byte2volt(Vmax = 50, Vmin = 50, nBytes = 16)
}
\arguments{
\item{Vmax}{length-one numeric vector: maximal nominal analog input voltage.
If \code{Vmax = NULL} it returns \code{1} (no bytes to volt
transformation)}

\item{Vmin}{length-one numeric vector: minimal nominal analog input voltage.
If missing, then \code{Vmin = -Vmax}.}

\item{nBytes}{Length-one numeric vector: number of bytes}
}
\description{
Convert bytes to volt values
}
