% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipData.R
\name{clipData}
\alias{clipData}
\alias{clipData,GPRvirtual-method}
\title{Clip the amplitude}
\usage{
clipData(x, cl = NULL, track = TRUE)

\S4method{clipData}{GPRvirtual}(x, cl = NULL, track = TRUE)
}
\arguments{
\item{x}{(\code{GPR}) GPR object.}

\item{cl}{(\code{numeric[1|2]}) Value above and below which the signal has
to be clipDataped. If \code{cl} is a length-one
vector, the signal outside the range
\code{-cl}, \code{cl} will be clipDataped.
If \code{cl} is a length-one
vector, the signal outside the range
\code{cl[1]}, \code{cl[2]} will be
clipDataped.}

\item{track}{(\code{logical[1]}) If \code{TRUE}, processing will be tracked.}
}
\value{
(\verb{GPR class}) clipDataped GPR object.
}
\description{
Clip the amplitude

Clip the amplitude
}
\concept{signal processing}
