% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMPAnalysis.R
\name{CMPAnalysis}
\alias{CMPAnalysis}
\alias{CMPAnalysis,GPR-method}
\title{Velocity Analysis of CMP Gather}
\usage{
\S4method{CMPAnalysis}{GPR}(
  x,
  method = c("semblance", "winsemblance", "wincoherence", "wincoherence2"),
  v = NULL,
  w = NULL
)
}
\arguments{
\item{x}{An object of the class \code{GPR}}

\item{method}{A length-one character vector}

\item{v}{A numeric vector defining at which velocities the analysis is
performed}

\item{w}{A length-one numeric vector defining the window length for the
methods 'wincoherence' and 'wincoherence2'.}
}
\description{
Transform the space-time domain of the radargram into a velocity-time domain to obtain 
the velocity spectrum (i.e. change in wave velocity with depth or time). This is achieved by applying
Normal Move-Out (NMO) corrections to the radargram for the range of selected velocities 
and computing a coherency measure for each result. In RGPR, the coherency measure can be defined using 
different functions: "semblance", "winsemblance", "wincoherence", "wincoherence2".
}
\details{
either use 'rec' and 'trans' to compute the distance between the antennas
or give the distance between the antennas (asep)
or seq(x@antsep, by = x@dx, length.out = length(x))

\describe{
  \item{semblance}{also described as the ratio of input to output
        energy (Niedell and Taner, 1971)}
  \item{winsemblance}{windowed semblance}       
  \item{wincoherence}{Windowed coherence measure based on 
        eigen-decomposition that estimates the 
        signal-to-noise ratio for high resolution velocity analysis
        (Sacchi, 2002)}
  \item{wincoherence2}{Windowed coherence measure based on a log-generalized
        likelihood ratio which tests the hypothesis of equality of 
        eigenvalues (Key and Smithson, 1990)}
}
}
\references{
\itemize{
  \item{Neidell and Taner (1971) Semblance and other coherency measures
        for multichannel data. 
        Geophysics, 36(3):482-497.}
  \item{Key and Smithson (1990) New approach to seismic-reflection
        event detection and velocity determination. 
        Geophysics, 55(8):1057-1069.}
  \item{Textbook: Sacchi (2002) Statistical and Transform Methods
        in Geophysical Signal Processing}
}
}
