% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\name{colSums,GPR-method}
\alias{colSums,GPR-method}
\alias{rowSums,GPR-method}
\alias{colMeans,GPR-method}
\alias{rowMeans,GPR-method}
\title{Form Row and Column Sums and Means}
\usage{
\S4method{colSums}{GPR}(x, na.rm = FALSE, dims = 1)

\S4method{rowSums}{GPR}(x, na.rm = FALSE, dims = 1)

\S4method{colMeans}{GPR}(x, na.rm = FALSE, dims = 1)

\S4method{rowMeans}{GPR}(x, na.rm = FALSE, dims = 1)
}
\arguments{
\item{x}{[\code{GPR}]}

\item{na.rm}{[\code{logical(1)}]. Should missing values (including 
\code{NaN}) be omitted from the calculations?}

\item{dims}{[\code{integer(1)}]  Which dimensions are regarded as ‘rows’ or 
‘columns’ to sum over.(see \code{\link{colSums}}).}
}
\description{
Form row and column sums and means
}
