% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\name{colSums,GPRvirtual-method}
\alias{colSums,GPRvirtual-method}
\alias{rowSums,GPRvirtual-method}
\alias{colMeans,GPRvirtual-method}
\alias{rowMeans,GPRvirtual-method}
\title{Form Row and Column Sums and Means}
\usage{
\S4method{colSums}{GPRvirtual}(x, na.rm = FALSE, dims = 1)

\S4method{rowSums}{GPRvirtual}(x, na.rm = FALSE, dims = 1)

\S4method{colMeans}{GPRvirtual}(x, na.rm = FALSE, dims = 1)

\S4method{rowMeans}{GPRvirtual}(x, na.rm = FALSE, dims = 1)
}
\arguments{
\item{x}{(\code{GPR})}

\item{na.rm}{(\code{logical[1]}). Should missing values (including
\code{NaN}) be omitted from the calculations?}

\item{dims}{(\code{integer[1]})  Which dimensions are regarded as ‘rows’ or
‘columns’ to sum over.(see \code{\link[=colSums]{colSums()}}).}
}
\description{
Form row and column sums and means
}
\concept{arithmetic}
