% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupCompare.R
\name{Compare,GPR,ANY-method}
\alias{Compare,GPR,ANY-method}
\alias{Compare}
\alias{Compare,GPR,GPR-method}
\alias{Compare,ANY,GPR-method}
\title{Comparison operations}
\usage{
\S4method{Compare}{GPR,ANY}(e1, e2)

\S4method{Compare}{GPR,GPR}(e1, e2)

\S4method{Compare}{ANY,GPR}(e1, e2)
}
\arguments{
\item{e1}{An object of the class GPR}

\item{e2}{An object of the class GPR}
}
\description{
Methods for the base Compare methods \link[methods]{S4groupGeneric}: "==", ">", "<", "!=", "<=", ">="
}
