% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTimeToDepth.R
\name{convertTimeToDepth}
\alias{convertTimeToDepth}
\alias{convertTimeToDepth,GPR-method}
\title{Time to depth conversion}
\usage{
convertTimeToDepth(
  x,
  dz = NULL,
  dmax = NULL,
  method = c("pchip", "linear", "nearest", "spline", "cubic")
)

\S4method{convertTimeToDepth}{GPR}(
  x,
  dz = NULL,
  dmax = NULL,
  method = c("pchip", "linear", "nearest", "spline", "cubic")
)
}
\arguments{
\item{x}{[\code{GPR* object}] An object of the class \code{GPR}}

\item{dz}{[\code{numeric(1)}] Depth resolution for the time to depth 
conversion. If \code{dz = NULL}, then \code{dz} is set equal 
to the smallest depth resolution computed from time.}

\item{dmax}{[\code{numeric(1)}] Maximum depth for the time to depth 
conversion. If \code{dmax = NULL}, then \code{dmax} is set 
equal to the largest possible depth.}

\item{method}{[\code{character(1)}]Interpolation method to be applied:
one of \code{pchip}, \code{linear}, \code{nearest}, 
\code{spline}, or \code{cubic}
(see also \code{\link[signal]{interp1}}).}
}
\value{
[\code{GPR* object}] with signal as a function of depth.
}
\description{
Convert two-way travel time to depth. This is a non-linear operation
}
