% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTimeToDepth.R
\name{convertTimeToDepth}
\alias{convertTimeToDepth}
\alias{convertTimeToDepth,GPR-method}
\title{Time to depth conversion}
\usage{
convertTimeToDepth(
  x,
  dz = NULL,
  zmax = NULL,
  method = c("pchip", "linear", "nearest", "spline", "cubic")
)

\S4method{convertTimeToDepth}{GPR}(
  x,
  dz = NULL,
  zmax = NULL,
  method = c("pchip", "linear", "nearest", "spline", "cubic")
)
}
\arguments{
\item{x}{[\code{GPR* object}] An object of the class \code{GPR}}

\item{dz}{[\code{numeric(1)}] Desired depth resolution. 
If \code{dz = NULL}, then \code{dz} is set equal 
to the smallest depth resolution inferred from the data.}

\item{zmax}{[\code{numeric(1)}] Maximum Desired depth.
If \code{zmax = NULL}, then \code{zmax} is set 
equal to the largest depth inferred from the data.}

\item{method}{[\code{character(1)}] Interpolation method to be applied:
one of \code{pchip}, \code{linear}, \code{nearest}, 
\code{spline}, or \code{cubic}
(see also \code{\link[signal]{interp1}}).}
}
\value{
[\code{GPR* object}] with signal as a function of depth.
}
\description{
Convert the two-way travel time of the recorded waves into depth. It does
not account for the topography. To add the topography, 
use \code{\link{migrate}} instead. This is a non-linear operation
}
