% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{crs}
\alias{crs}
\alias{crs<-}
\alias{crs<-,GPR-method}
\alias{crs,GPR-method}
\alias{crs,GPRsurvey-method}
\alias{crs<-,GPRsurvey-method}
\title{Coordinate reference system (CRS) of the GPR data}
\usage{
crs(x)

crs(x) <- value

\S4method{crs}{GPR}(x)

\S4method{crs}{GPR}(x) <- value

\S4method{crs}{GPRsurvey}(x)

\S4method{crs}{GPRsurvey}(x) <- value
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}

\item{value}{[\code{character(1)}] A string accepted by GDAL 
(e.g., \code{"EPSG:2056"}, WKT-string).}
}
\value{
[\code{GPR class}] An object of the class \code{GPR}
}
\description{
Coordinate reference system (CRS) of the GPR data
}
\details{
Modified slots
\itemize{
  \item \code{crs} the coordinate reference system.
  \item \code{spunit} the spatial units are updated accroding to the new
                      coordinate reference system.
}
}
\examples{
\dontrun{
x <- readGPR("LINE.DT1")
crs(x) <- "EPSG:3857"
}
}
\concept{getters/setters}
