% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRsurvey.R, R/getterSetter.R
\docType{methods}
\name{crs}
\alias{crs}
\alias{crs<-}
\title{Coordinate reference system (CRS) of the GPR data}
\usage{
\S4method{crs}{GPRsurvey}(x)

\S4method{crs}{GPRsurvey}(x) <- value

\S4method{crs}{GPR}(x)

\S4method{crs}{GPR}(x) <- value
}
\description{
Coordinate reference system (CRS) of the GPR data
}
