% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\docType{methods}
\name{crs}
\alias{crs}
\alias{crs<-}
\alias{crs<-,GPR-method}
\alias{crs,GPR-method}
\alias{crs,GPRsurvey-method}
\alias{crs<-,GPRsurvey-method}
\title{Coordinate reference system (CRS) of the GPR data}
\usage{
crs(x)

crs(x) <- value

\S4method{crs}{GPR}(x)

\S4method{crs}{GPR}(x) <- value

\S4method{crs}{GPRsurvey}(x)

\S4method{crs}{GPRsurvey}(x) <- value
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}

\item{value}{[\code{sp::CRS|character}] CRS object (one or more)}
}
\value{
[\code{GPR class}] An object of the class \code{GPR}
}
\description{
Coordinate reference system (CRS) of the GPR data
}
\details{
Modified slots
\itemize{
  \item \code{crs} the coordinate reference system.
  \item \code{spunit} the spatial units are updated accroding to the new
                      coordinate reference system.
}
}
