% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/ClassGPRvirtual.R,
%   R/ClassGPRsurvey.R
\docType{methods}
\name{crs}
\alias{crs}
\alias{crs<-}
\alias{crs}
\alias{crs<-}
\alias{crs}
\alias{crs}
\title{Coordinate reference system (CRS) of the GPR data}
\usage{
\S4method{crs}{GPRvirtual}(x)

\S4method{crs}{GPRvirtual}(x) <- value

\S4method{crs}{GPRsurvey}(x)

\S4method{crs}{GPRsurvey}(x) <- value
}
\description{
Coordinate reference system (CRS) of the GPR data
}
