% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsUnit.R
\name{crsUnit}
\alias{crsUnit}
\title{Unit of Coordinate Reference System (CRS)}
\usage{
crsUnit(crs)
}
\arguments{
\item{crs}{(\code{character[1]}) A string accepted by GDAL
(e.g., \code{"EPSG:2056"}, WKT-string).}
}
\value{
(\code{character}) The unit abbreviation (except for degree)
}
\description{
Returns the unit of the CRS
}
\concept{units}
