% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{dcshift}
\alias{dcshift}
\title{First wave break}
\usage{
\S4method{dcshift}{GPR}(x, u = 1:10, FUN = mean)

\S4method{firstBreack}{GPR}(x, w = 11, ns = NULL, bet = NULL)
}
\description{
Compute the first wave break
}
\details{
Jaun I. Sabbione and Danilo Velis (2010). Automatic first-breaks picking: 
New strategies and algorithms. Geophysics, 75 (4): v67-v76
-> modified Coppens's Method
nl = length leading window: about one period of the firs-arrival waveform
ns = length eps (edge preserving smoothing) window: good results with ns 
between one and two signal periods
       -> default values ns= 1.5*nl
bet = stabilisation constant, not critical, set to 0.2*max(amplitude)
}

