% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declip.R
\docType{methods}
\name{declip,GPR-method}
\alias{declip,GPR-method}
\title{Declip the GPR signal}
\usage{
\S4method{declip}{GPR}(x, xclip = NULL, xrange = NULL, lambda = 1)
}
\description{
Use constrained least squares. Based on the code of Ivan Selesnick:
 "we minimize
  the energy of the the third derivative. This encourages the filled in data 
 to have the form of a parabola (second order polynomial), because the third
  derivative of a parabola is zero."
+ constraint (i.e., constrained least squares) like in
http://www.cs.cmu.edu/~robust/Papers/HarvillaStern15a.pdf
}
