% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{delineate}
\alias{delineate}
\alias{rmDelineations<-}
\alias{delineations}
\alias{addDelineation}
\alias{exportDelineations}
\alias{plotDelineations3D}
\alias{plotDelineations}
\alias{identifyDelineation}
\title{Delineate structure on GPR data}
\usage{
\S4method{delineate}{GPR}(x, name = NULL, type = c("raster",
  "wiggles"), addTopo = FALSE, nupspl = NULL, n = 10000, ...)

\S4method{addDelineation}{GPR}(x, itp, name = NULL, type = c("raster",
  "wiggles"), addTopo = FALSE, ...)

\S4method{rmDelineations}{GPR}(x) <- value

\S4method{delineations}{GPR}(x, sel = NULL, ...)

\S4method{exportDelineations}{GPR}(x, dirpath = "")

\S4method{plotDelineations3D}{GPR}(x, sel = NULL, col = NULL,
  add = TRUE, ...)

\S4method{plotDelineations}{GPR}(x, sel = NULL, col = NULL, ...)

\S4method{identifyDelineation}{GPR}(x, sel = NULL, ...)
}
\description{
Delineate structure on GPR data
}
