% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineation.R
\docType{methods}
\name{delineate}
\alias{delineate}
\alias{delineations}
\alias{plotDelineations}
\alias{rmDelineations<-}
\alias{exportDelineations}
\alias{plot3DDelineations}
\alias{plotDelineations3D}
\alias{identifyDelineation}
\title{Delineate structure on GPR data}
\usage{
delineate(x, name = NULL, n = 10000, plot_del = NULL, ...)

\S4method{delineate}{GPR}(x, name = NULL, n = 10000, plot_del = NULL,
  ...)

delineations(x)

\S4method{delineations}{GPR}(x)

plotDelineations(x, col = NULL, ...)

\S4method{plotDelineations}{GPR}(x, col = NULL, ...)

rmDelineations(x) <- value

\S4method{rmDelineations}{GPR}(x) <- value

\S4method{exportDelineations}{GPR}(x, dirpath = "")

\S4method{plotDelineations3D}{GPR}(x, col = NULL, add = TRUE, ...)

\S4method{identifyDelineation}{GPR}(x, ...)
}
\description{
Print and invisible returns the GPR data delineations.

Works only close to the points !!!
}
