% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate.R
\name{delineate}
\alias{delineate}
\alias{plotDelineations}
\title{Delineate structure on GPR data}
\usage{
delineate(x, name = NULL, values = NULL, n = 10000, plotDel = NULL, ...)

\S4method{delineate}{GPR}(x, name = NULL, values = NULL, n = 10000, plotDel = NULL, ...)

plotDelineations(
  x,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  ...
)

\S4method{plotDelineations}{GPR}(
  x,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  ...
)
}
\arguments{
\item{x}{(\code{GPR})}

\item{name}{(\code{character[1]}) Names of the delineated line}

\item{values}{(\code{list}) list of x and y coordinates (optional). If \code{NULL},
\code{locator()} function is run.}

\item{n}{(\code{numeric[1]}) the maximum number of points to locate. Valid values start at 1.}

\item{plotDel}{(\code{logical[1]}) If \code{TRUE}plot delineation.}
}
\description{
Delineate structure on GPR data

Plot the delineation on a 2D plot
}
