% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineation.R
\name{delineate}
\alias{delineate}
\alias{delineations}
\alias{plotDelineations}
\alias{rmDelineations<-}
\alias{exportDelineations}
\alias{plot3DDelineations}
\alias{identifyDelineation}
\title{Delineate structure on GPR data}
\usage{
delineate(x, name = NULL, n = 10000, plot_del = NULL, ...)

\S4method{delineate}{GPR}(x, name = NULL, n = 10000, plot_del = NULL, ...)

delineations(x, name = NULL)

\S4method{delineations}{GPR}(x, name = NULL)

plotDelineations(
  x,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  ...
)

\S4method{plotDelineations}{GPR}(
  x,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  ...
)

rmDelineations(x) <- value

\S4method{rmDelineations}{GPR}(x) <- value

\S4method{exportDelineations}{GPR}(x, dirpath = "")

plot3DDelineations(
  x,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  add = TRUE,
  ...
)

\S4method{plot3DDelineations}{GPR}(
  x,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  add = TRUE,
  ...
)

\S4method{identifyDelineation}{GPR}(
  x,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  ...
)
}
\description{
Print and invisible returns the GPR data delineations.

Works only close to the points !!!
}
