% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/ClassGPRvirtual.R
\docType{methods}
\name{description}
\alias{description}
\alias{description<-}
\alias{description}
\alias{description<-}
\title{Description of the GPR data}
\usage{
\S4method{description}{GPRvirtual}(x)

\S4method{description}{GPRvirtual}(x) <- value
}
\description{
Description of the GPR data
}
