% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{dewow}
\alias{dewow}
\title{Trace dewowing}
\usage{
\S4method{dewow}{GPR}(x, type = c("MAD", "Gaussian"), w)
}
\arguments{
\item{x}{An object of the class GPR.}

\item{type}{A length-one character vector, either \code{MAD} (Median
Absolute Deviation filter) or \code{Gaussian} (Gaussian
filter)}

\item{w}{A length-one numeric vector equal to the window length 
of the filter. Per default, the filter length is five times
the GPR pulse width.}
}
\value{
An object of the class GPR whose traces are dewowed.
}
\description{
\code{dewow} remove the low-frequency component (the so-called 'wow') of 
every trace..
}
\examples{
data(frenkeLine00)
A <- dewow(frenkeLine00, type = "Gaussian")
A
}
