% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewow.R
\name{dewow}
\alias{dewow}
\alias{dewow,GPR-method}
\title{Trace dewowing}
\usage{
dewow(x, type = c("runmed", "runmean", "gaussian"), w = NULL, track = TRUE)

\S4method{dewow}{GPR}(x, type = c("runmed", "runmean", "gaussian"), w = NULL, track = TRUE)
}
\arguments{
\item{x}{(\code{GPR}) An object of the class GPR.}

\item{type}{(\code{character[1]}) Dewow method,
one of \code{runmed} (running median),
\code{runmean} (running mean),
\code{Gaussian} (Gaussian smoothing).}

\item{w}{(\code{numeric[1]}) If \code{type} = \code{runmed},
\code{MAD} or \code{runmean}, window length of the filter in
trace unit;
If \code{type} = \code{Gaussian}, standard deviation in trace
unit.
If \code{w = NULL}, \code{w} is estimated as five times the
wavelength corresponding to the maximum frequency of x
(estimated with \link{spec})}

\item{track}{(\code{logical[1]}) Should the processing step be tracked?}
}
\value{
(\code{GPR}) An object of the class GPR whose traces are dewowed.
}
\description{
\code{dewow} remove the low-frequency component (the so-called 'wow') of
every traces.
}
\details{
The low-frequency component is computed by different methods:
\itemize{
\item \code{runmed} running median based on \link[stats:runmed]{stats::runmed}
\item \code{runmean} running mean based on \link[stats:filter]{stats::filter}
\item \code{MAD} DEPRECATED - Median Absolute Deviation filter
\item \code{Gaussian} Gaussian smoothing applied to the trace samples
after time-zero based on \link[mmand:gaussianSmooth]{mmand::gaussianSmooth}
}

Modified slots:
\itemize{
\item \code{data}: trace dewowed.
\item \code{proc}: updated with function name and arguments.
}
}
\concept{processing}
