% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{dim,GPR-method}
\alias{dim,GPR-method}
\title{Dimenstionf an object of the class \code{GPR}}
\usage{
\S4method{dim}{GPR}(x)
}
\description{
= number of time samples, number of traces.
}
\examples{
data(frenkeLine00)
dim(frenkeLine00)
}
