% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRvirtual.R
\docType{methods}
\name{dim,GPRvirtual-method}
\alias{dim,GPRvirtual-method}
\title{Dimenstionf an object of the class \code{GPR}}
\usage{
\S4method{dim}{GPRvirtual}(x)
}
\description{
= number of time samples, number of traces.
}
\examples{
data(frenkeLine00)
dim(frenkeLine00)
}
