% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHDF5.R
\name{.write_GPR_line_hdf5}
\alias{.write_GPR_line_hdf5}
\title{Write a single GPR line into an open HDF5 group}
\usage{
.write_GPR_line_hdf5(parent_grp, name, gpr, compress = 5L)
}
\arguments{
\item{parent_grp}{An open \code{hdf5r} group object (the \verb{/lines} group).}

\item{name}{(\code{character(1)}) Name for the new sub-group.}

\item{gpr}{Object of class \code{GPR}.}

\item{compress}{(\code{integer(1)}) gzip level 0–9.}
}
\value{
Invisibly returns the created group object.
}
\description{
Write a single GPR line into an open HDF5 group
}
\keyword{internal}
