% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPRsurvey.R
\name{.write_intersections_hdf5}
\alias{.write_intersections_hdf5}
\title{Write line intersection data into the HDF5 file}
\usage{
.write_intersections_hdf5(h5, survey)
}
\arguments{
\item{h5}{Open \code{hdf5r} H5File object.}

\item{survey}{\code{GPRsurvey} object returned by \code{intersect()}.}
}
\description{
Writes the \code{@intersections} slot (if non-empty) to
\code{/survey/intersections/} as a dataset of crossing coordinates.
Called once at the end of the constructor, after \code{intersect()}.
}
\keyword{internal}
