% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{estimateTime0}
\alias{estimateTime0}
\title{Estimate and set time-zero}
\usage{
\S4method{estimateTime0}{GPR}(x, method = c("coppens", "threshold",
  "MER"), thr = 0.12, w = 11, ns = NULL, bet = NULL, c0 = 0.299,
  FUN = NULL, ...)
}
\arguments{
\item{x}{An object of the class \code{GPR}}

\item{method}{A length-one character vector. \code{"coppens"} corresponds to
the modified Coppens method, \code{"threshold"} to the 
threshold method, and \code{"MER"} to the modified energy ratio
method.}

\item{thr}{A length-one numeric vector defining the threshold  signal 
amplitude (in \%) at which time zero is picked (only for the
threshold method).}

\item{w}{A length-one numeric vector defining the length of leading window 
(only for the modified Coppens and modified energy ratio 
methods). Recommended value: about one period of the first-arrival 
waveform. w is defined on a time basis.}

\item{ns}{A length-one numeric vector defining the length of the edge 
preserving smoothing window (only for the modified Coppens 
method). Recommended value: between one and two signal periods.
When \code{ns = NULL} the value of \code{ns} is set to 
\code{1.5 * w}.}

\item{bet}{A length-one numeric vector defining the stabilisation 
constant (only for the modified Coppens method). Not critical. 
When \code{bet = NULL} the value of \code{bet} is set to 
20\% of the maximal signal amplitude.}

\item{c0}{Propagation speed of the GPR wave through air (used only when
\code{keep = NULL}).}

\item{FUN}{Optional: function to apply to the estimated time-zero (e.g., 
\code{mean} or \code{median} to get set a single time-zero 
value to the data).}

\item{...}{Arguments of \code{FUN}.}
}
\description{
Pick the time corresponding to the first break of each trace in the GPR profile.
Return a vector containing the first break times.
}
\details{
This function is a wrapper for the following commands
\itemize{
  \item \code{tfb <- firstBreak(x, ...)}
  \item \code{t0 <- firstBreakToTime0(tfb, x)}
  \item \code{time0(x) <- t0}
}
}
\references{
\describe{
  \item{Modified Coppens method}{Sabbione J.I. and Velis D. (2010) 
       Automatic first-breaks picking: New strategies and algorithms. 
       Geophysics, 75(4): 67-76.}
  \item{Modified Energy Ratio (MER) method}{Han L., Wong J., and John C. 
       (2010) Time picking on noisy microseismograms. In: Proceedings of the
       GeoCanada 2010 Convention - Working with the Earth, Calgary, AB, 
       Canada, p. 4}
}
}
\seealso{
\code{\link{time0}} to set time zero and 
         \code{\link{time0Cor}} to shift the traces such that they start
         at time zero.

\code{\link{time0}} and \code{\link{setTime0}} to set time zero,
         \code{\link{firstBreak}} to estimate the first wave break, and
         \code{\link{firstBreakToTime0}} to convert the first wave break
         into time zero.
}
