% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTime0.R
\name{estimateTime0}
\alias{estimateTime0}
\title{Estimate and set time-zero}
\usage{
estimateTime0(
  x,
  method = c("coppens", "threshold", "MER"),
  thr = NULL,
  w = NULL,
  ns = NULL,
  bet = NULL,
  shorten = TRUE,
  c0 = 0.299,
  FUN = NULL,
  ...,
  track = TRUE
)

\S4method{estimateTime0}{GPR}(
  x,
  method = c("coppens", "threshold", "MER"),
  thr = NULL,
  w = NULL,
  ns = NULL,
  bet = NULL,
  shorten = TRUE,
  c0 = 0.299,
  FUN = NULL,
  ...,
  track = TRUE
)
}
\arguments{
\item{x}{(\code{GPR}) An object of the class `GPR}

\item{method}{(\verb{character[1]) Method to be applied (either }"coppens"\verb{, }"threshold"\code{or}"MER"\verb{).  }"coppens"\verb{corresponds to the modified Coppens method, }"threshold"\verb{to the threshold method,  and}"MER"` to the modified energy ratio method.}

\item{thr}{(\code{numeric[1]}] Threshold for the signal
amplitude (in percent) at which time zero is picked (only for the
threshold method).
\code{thr} ranges between 0 (0 percent) and 1 (100 percent).}

\item{w}{(\code{numeric[1]}) Length of the short-term leading window in
unit of time (only for \code{method = "coppens"} or
\code{method = "MER"}).
Recommended value: about one period of the first-arrival
waveform.}

\item{ns}{(\code{numeric[1]}] Length of the edge preserving smoothing
window in unit of time (only for  \code{method = "coppens"}).
Recommended value: between one and two signal periods.
When \code{ns = NULL} the value of \code{ns} is set to
\code{1.5 * w}.}

\item{bet}{(\code{numeric[1]}) Stabilisation constant (only for
\code{method = "coppens"}). Not critical.
When \code{bet = NULL} the value of \code{bet} is set to
20 percent of the maximal signal amplitude.}

\item{shorten}{(\code{logical[1]}) If \code{TRUE}, each trace is shortened
by removing the samples that are \eqn{2 \times w} after the
maximum value (only for \code{method = "coppens"} or
\code{method = "MER"}). You may set
\code{shorten = FALSE} if the first wave break occurs
after the maximum absolute amplitude time.}

\item{c0}{(\code{numeric[1]}) Propagation speed of the GPR wave
through air in unit of space per unit of time
(generally in m/ns).}

\item{FUN}{(\verb{function()}) A function to apply on the
estimated time-zero of every traces (e.g., \code{mean} or
\code{median} to get set a single time-zero value to the data).}

\item{...}{Further arguments to be passed to \code{FUN}.}
}
\value{
(\code{GPR}) An object of the class \code{GPR}.
}
\description{
\code{estimateTime0} estimates for each trace individually the first wave
break, computes the corresponding time-zero knowing the propagation speed
of the electromagnetic wave through air and returns an object of the class
\code{GPR} with updated time-zero. It is possible to apply a function
provided by the user (e.g., \code{FUN}) on time-zero (e.g., to set time-zero
equal to the average value of the time-zeros computed for every traces; in
this case, all traces would have the same time-zero).
}
\details{
This function is a wrapper for the following commands:
\itemize{
\item \code{tfb <- pickFirstBreak(x, ...)}
\item \code{t0 <- firstBreakToTime0(x, tfb)}
\item \code{time0(x) <- t0} (if \code{FUN} is not \code{NULL} else \code{time0(x) <- FUN(t0, ...)})
}

Modified slots:
\itemize{
\item \code{time0}: new estimated time-zero.
\item \code{proc}: updated with function name and arguments.
}
}
\examples{
data("frenkeLine00")
x <- frenkeLine00
x1 <- estimateTime0(x, w = 10)
time0(x1)
x2 <- estimateTime0(x, w = 10, FUN = mean)
time0(x2)

}
\seealso{
\code{\link[=pickFirstBreak]{pickFirstBreak()}} to estimate the first wave break;
\code{\link[=firstBreakToTime0]{firstBreakToTime0()}} to convert the first wave break
into time zero.
\code{\link[=time0]{time0()}} and \code{\link[=setTime0]{setTime0()}} to set time-zero;
\code{\link[=shiftToTime0]{shiftToTime0()}} to shift the traces such that they start
at time-zero.
}
\concept{processing}
