% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{exportProc}
\alias{exportProc}
\title{Export the process steps.}
\usage{
\S4method{exportProc}{GPR}(x, fPath = NULL, sep = "\\t",
  row.names = FALSE, col.names = FALSE, ...)
}
\description{
Export the process steps.
}
