% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{fFilter}
\alias{fFilter}
\title{Frequency filter}
\usage{
\S4method{fFilter}{GPR}(x, f = 100, type = c("low", "high",
  "bandpass"), L = 257, plotSpec = FALSE)
}
\description{
Frequency filter
}
