% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{filter1D}
\alias{filter1D}
\alias{filter1D}
\title{One dimensional filters}
\usage{
\S4method{filter1D}{GPR}(x, type = c("median", "hampel", "Gaussian"), ...)
}
\description{
One dimensional filters
}
