% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\name{filter1D}
\alias{filter1D}
\title{One dimensional filters}
\usage{
\S4method{filter1D}{GPR}(
  x,
  type = c("runmed", "runmean", "mad", "gaussian", "hampel"),
  w = NULL,
  track = TRUE
)
}
\description{
One dimensional filters
}
