% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2D.R
\name{filter2D}
\alias{filter2D}
\title{Two-dimensional filters}
\usage{
filter2D(x, type = c("median3x3", "adimpro", "smooth"), ..., track = TRUE)

\S4method{filter2D}{GPR}(x, type = c("median3x3", "adimpro", "smooth"), ..., track = TRUE)
}
\description{
Two-dimensional filters
}
\concept{processing}
