% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter2D.R
\name{filter2D}
\alias{filter2D}
\alias{filter2D,GPR-method}
\title{Two-dimensional filters}
\usage{
filter2D(obj, type = c("median3x3", "adimpro", "gaussian"), ..., track = TRUE)

\S4method{filter2D}{GPR}(obj, type = c("median3x3", "adimpro", "gaussian"), ..., track = TRUE)
}
\arguments{
\item{obj}{(\code{GPR* object})}

\item{type}{(\code{character[1]}) Filter method.}

\item{...}{(\code{*}) Additional parameters}

\item{track}{(\code{logical[1]}) Should the processing step be tracked?}
}
\description{
Two-dimensional filters
}
\concept{processing}
