% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{filter2D}
\alias{filter2D}
\title{Two-dimensional filters}
\usage{
\S4method{filter2D}{GPR}(x, type = c("median3x3", "adimpro"), ...)
}
\description{
Two-dimensional filters
}
