% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIntersection.R
\name{findIntersection}
\alias{findIntersection}
\alias{findIntersection,GPRsurvey-method}
\title{Compute GPR profile intersections}
\usage{
findIntersection(x)

\S4method{findIntersection}{GPRsurvey}(x)
}
\arguments{
\item{x}{\code{\link{GPRsurvey}} An object of the class \code{GPRsurvey}}
}
\value{
(\verb{object GPRsurvey}) An object of the class GPRsurvey.
}
\description{
Compute GPR profile intersections
}
\details{
Modified slots
\itemize{
\item \code{intersects}: trace shifted. The number of rows of data may
be smaller if \code{crop = TRUE}.
}
}
\concept{spatial computing}
