% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstBreak.R
\name{firstBreak}
\alias{firstBreak}
\title{Time of first wave break}
\usage{
\S4method{firstBreak}{GPR}(
  x,
  method = c("coppens", "threshold", "MER"),
  thr = NULL,
  w = NULL,
  ns = NULL,
  bet = NULL,
  shorten = TRUE
)
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}

\item{method}{[\code{character(1)}] Method to be applied (either
\code{coppens}, \code{threshold} or \code{MER}). 
\code{"coppens"} corresponds to the modified Coppens method, 
\code{"threshold"} to the threshold method, 
and \code{"MER"} to the modified energy ratio method.}

\item{thr}{[\code{numeric(1)}] Threshold for the signal 
amplitude (in \%) at which time zero is picked (only for the
threshold method). 
\code{thr} ranges between 0 (0%) and 1 (100%).}

\item{w}{[\code{numeric(1)}] Length of the short-term leading window in 
unit of time (only for \code{method = "coppens"} or 
\code{method = "MER"}). 
Recommended value: about one period of the first-arrival 
waveform.}

\item{ns}{[\code{numeric(1)}] Length of the edge preserving smoothing 
window in unit of time (only for  \code{method = "coppens"}). 
Recommended value: between one and two signal periods.
When \code{ns = NULL} the value of \code{ns} is set to 
\code{1.5 * w}.}

\item{bet}{[\code{numeric(1)}] Stabilisation constant (only for 
\code{method = "coppens"}). Not critical. 
When \code{bet = NULL} the value of \code{bet} is set to 
20\% of the maximal signal amplitude.}

\item{shorten}{[\code{logical(1)}] If \code{TRUE}, each trace is shortened
by removing the samples that are \eqn{2 \times w} after the
maximum value (only for \code{method = "coppens"} or 
\code{method = "MER"}). You may set 
\code{shorten = FALSE} if the first wave break occurs 
after the maximum absolute amplitude time.}
}
\value{
[\code{numeric(n)}] The time of the first wave break for every
        traces in unit of time (\code{n = ncol(x) =} number of traces).
}
\description{
Pick the time of the first wave break in each trace 
(trace-by-trace function).
}
\details{
The Modified Coppens's method (\code{coppens}) computes the energy ratio 
between a long term window (with increasing length) and a short-term leading 
window (fixed length).
Edge-preserving smoothing is then applied and the first wave break is 
assigned to the sample in which the derivative of the output is largest.

The modified energy ratio method (\code{MER}) computes the energy ratio between
a preceding and trailing windows of equal length. The energy ratio is
then multiplied by the absolute values of the trace and the output to the 
power of three is returned.

In the threshold method (\code{"threshold"}), the sample before the first 
sample that is larger than the threshold times the maximum absolute 
amplitude is picked. Then the time of the first wave break is linearly 
interpolated between these two samples.
}
\references{
\describe{
  \item{Modified Coppens method}{Sabbione J.I. and Velis D. (2010) 
       Automatic first-breaks picking: New strategies and algorithms. 
       Geophysics, 75(4): 67-76.}
  \item{Modified Energy Ratio (MER) method}{Han L., Wong J., and John C. 
       (2010) Time picking on noisy microseismograms. In: Proceedings of the
       GeoCanada 2010 Convention - Working with the Earth, Calgary, AB, 
       Canada, p. 4}
}

#' @examples 
data("frenkeLine00")
fb <- firstbreak(frenkeLine00, w = 10)
plot(seq_along(frenkeLine00), fb)
}
\seealso{
\code{\link{firstBreakToTime0}} to convert time of first wave break
         into time-zero; 
         \code{\link{time0}} and \code{\link{setTime0}} to set time-zero;
         \code{\link{estimateTime0}} to estimate first wave break, convert
         it to time-zero and set time zero (all in one step);
         \code{\link{time0Cor}} to shift the traces such that they start
         at time-zero.
}
