% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{firstBreak}
\alias{firstBreak}
\title{First wave break}
\usage{
\S4method{firstBreak}{GPR}(x, method = c("coppens", "coppens2",
  "threshold", "MER"), thr = 0.12, w = 11, ns = NULL, bet = NULL)
}
\arguments{
\item{x}{An object of the class \code{GPR}}

\item{method}{A length-one character vector. \code{"coppens"} corresponds to
the modified Coppens method, \code{"threshold"} to the 
threshold method, and \code{"MER"} to the modified energy ratio
method.}

\item{thr}{A length-one numeric vector defining the threshold  signal 
amplitude (in \%) at which time zero is picked (only for the
threshold method).}

\item{w}{A length-one numeric vector defining the length of leading window 
(only for the modified Coppens and modified energy ratio 
methods). Recommended value: about one period of the first-arrival 
waveform. w is defined on a time basis.}

\item{ns}{A length-one numeric vector defining the length of the edge 
preserving smoothing window (only for the modified Coppens 
method). Recommended value: between one and two signal periods.
When \code{ns = NULL} the value of \code{ns} is set to 
\code{1.5 * w}.}

\item{bet}{A length-one numeric vector defining the stabilisation 
constant (only for the modified Coppens method). Not critical. 
When \code{bet = NULL} the value of \code{bet} is set to 
20\% of the maximal signal amplitude.}
}
\description{
Pick the time corresponding to the first break of each trace in the GPR profile.
Return a vector containing the first break times.
}
\references{
\describe{
  \item{Modified Coppens method}{Sabbione J.I. and Velis D. (2010) 
       Automatic first-breaks picking: New strategies and algorithms. 
       Geophysics, 75(4): 67-76.}
  \item{Modified Energy Ratio (MER) method}{Han L., Wong J., and John C. 
       (2010) Time picking on noisy microseismograms. In: Proceedings of the
       GeoCanada 2010 Convention - Working with the Earth, Calgary, AB, 
       Canada, p. 4}
}
}
\seealso{
\code{\link{time0}} to set time zero and 
         \code{\link{time0Cor}} to shift the traces such that they start
         at time zero.
}
