% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{firstBreakToTime0}
\alias{firstBreakToTime0}
\title{Convert first wave break time to time-zero}
\usage{
firstBreakToTime0(fb, x, c0 = 0.299)
}
\description{
Account for the delay time between time of wave emission and time of first
wave break recording due to the antenna separation (offset).
}
