% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{freqFromString}
\alias{freqFromString}
\title{Extract frequency from string}
\usage{
freqFromString(s)
}
\arguments{
\item{s}{[\code{character}] Character string that may contain an 
indication of a frequency.}
}
\value{
[\code{numeric}] The frequency (\code{NA} if no frequency value
                            is found)
}
\description{
Extract with regex the antenna frequency in a string
}
\examples{
s <- "1230 fds 200-MHZ 12.3"
freqFromString(s) 
s <- "1230MLF"
freqFromString(s) 
s <- "D1230MLF"
freqFromString(s) 
}
