% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainAGC.R
\name{gainAGC}
\alias{gainAGC}
\title{Automatic Gain Control (AGC) gain}
\usage{
\S4method{gainAGC}{GPR}(x, w = 10, p = 2, r = 0.5, track = TRUE)
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class GPR.}

\item{w}{[\code{numeric(1)}] Standard deviation of the 
Gaussian smoother (in trace unit).}

\item{p}{[\code{numeric(1)}] Parameter of the power filter
(\code{b} \eqn{\geq} 0). Usually, \code{b = 1}.}

\item{r}{[\code{numeric}] Start time of the gain filter
(if \code{t0 = NULL}, \code{t0} is set equal to \code{time0(x)}).}
}
\value{
[\code{GPR class}] An object of the class GPR.
}
\description{
\code{gainAGC} applies an AGC (Automatic Gain Control) gain.
The trace signal is smoothed with a Gaussian filter. The smoothed trace
is substracted from the original trace, raised to power \code{p}, smoothed
by a Gaussian filter and raised to power \code{r} to obtain the gain.
}
\details{
Spreading and Exponential Compensation (SEC) gain can be written as 
\eqn{\exp(a \cdot t) \cdot t^b}, where \eqn{t^b} is the power gain
(set \eqn{b = 1} to get a linear gain) and \eqn{\exp(a \cdot t)} is the
exponential gain.

Modified slots
\itemize{
  \item \code{data}: trace gained.
  \item \code{proc}: updated with function name and arguments.
}
}
\seealso{
\code{\link{gainSEC}}
}
