% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainEnv.R
\name{gainEnv}
\alias{gainEnv}
\alias{gainEnv,GPR-method}
\title{Gain compensation based on envelop}
\usage{
gainEnv(x, FUN = mean, track = TRUE)

\S4method{gainEnv}{GPR}(x, FUN = mean, track = TRUE)
}
\description{
Gain compensation based on envelop
}
