% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_utils.R
\name{getAntFreqGSSI}
\alias{getAntFreqGSSI}
\title{Frequency of GSSI antenna}
\usage{
getAntFreqGSSI(x)
}
\arguments{
\item{x}{(\code{character}) Name(s) of the antenna(e)}
}
\value{
(\code{list}) List of numeric values corresponding to the
frequency/frequencies
}
\description{
Given the antenna name, returns the frequency of GSSI antenna
}
