% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_files.R
\name{getFName}
\alias{getFName}
\title{Filepath(s) with correct extension(s)}
\usage{
getFName(fPath, ext = c(".hd", ".dt1"), throwError = TRUE)
}
\arguments{
\item{fPath}{[\code{character(1)}] Filepath.}

\item{ext}{[\code{character}] Extensions to check 
(e.g., \code{".hd"}).}

\item{throwError}{[\code{logical(1)}] If TRUE, an error is thrown if the 
filepath with one of the extension does not exist. 
If FALSE, it returns NULL for the missing extension}
}
\value{
[\code{list}] The list keys correspond to \code{ext} and 
                  the values to the filepaths 
                  (e.g., \code{$hd  -> xline01.hd}).
}
\description{
Returns the filepaths with the correct extension and check for 
upper and lower case extension (e.g., ".txt" or ".TXT")
}
\concept{files}
