% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGPR.R
\name{getGPR}
\alias{getGPR}
\alias{getGPR,GPRsurvey-method}
\title{Extract GPR object from GPRsurvey object}
\usage{
getGPR(x, id, verbose = FALSE)

\S4method{getGPR}{GPRsurvey}(x, id, verbose = FALSE)
}
\arguments{
\item{x}{(\code{GPRsurvey})}

\item{id}{(\code{integer[1]|character[1]}) Indice or name of the GPR line to
extract.}

\item{verbose}{(\code{logical[1]}) If \code{TRUE} the function prints some
information.}
}
\value{
(\verb{GPR class}) An object of the class GPR.
}
\description{
Extract GPR object from GPRsurvey object
}
