% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRsurvey.R
\docType{methods}
\name{getGPR}
\alias{getGPR}
\alias{getGPR,GPRsurvey-method}
\title{Extract GPR object from GPRsurvey object}
\usage{
\S4method{getGPR}{GPRsurvey}(x, id)
}
\description{
Extract GPR object from GPRsurvey object
}
