% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVel.R
\name{getVel}
\alias{getVel}
\alias{getVel,GPR-method}
\title{Get velocity model as GPR object}
\usage{
getVel(x, type = c("vrms", "vint"))

\S4method{getVel}{GPR}(x, type = c("vrms", "vint"))
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}

\item{type}{[\code{vrm|vint}] Set the velocity you want (either root-mean square or internal velocity)}
}
\value{
[\code{GPR class}] An object of the class \code{GPR} containing the velocity model.
}
\description{
Return the velocity model (either the root-mean square or internal velocity).
}
