% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{class}
\name{GPR-class}
\alias{GPR-class}
\title{An S4 class to represent a ground-penetrating radar (GPR) data.}
\description{
An S4 class to represent a ground-penetrating radar (GPR) data.
}
\section{Slots}{

\describe{
\item{\code{version}}{A length-one character vector indicating the version of RGPR}

\item{\code{data}}{A \eqn{m \time n} numeric matrix consiting of a 
cross-section of signal amplitudes as a function of the GPR 
position. The columns of \code{data} correspond to the GPR 
traces and the row of \code{data} to the time/depth samples.}

\item{\code{traces}}{A length-m numeric vector corresponding to the trace number.}

\item{\code{depth}}{A length-n numeric vector indicating the sampling time or
the vertical position of the trace samples.}

\item{\code{pos}}{A length-m numeric vector indicating the relative position of
the trace along the survey profile.}

\item{\code{time0}}{A length-m numeric vector containing the 'time-zero' of every
trace.}

\item{\code{time}}{A length-m numeric vector containing the recording time of 
every trace.}

\item{\code{fid}}{A length-m character vector containing fiducial markers associated
with the traces.}

\item{\code{ann}}{A length-m character vector containing annotations associated 
with the traces.}

\item{\code{coord}}{A \eqn{m \time 3} matrix containing the (x, y, z) positions of
every trace.}

\item{\code{rec}}{A \eqn{m \time 3} matrix containing the (x, y, z) positions of
the receiver for every trace.}

\item{\code{trans}}{A \eqn{m \time 3} matrix containing the (x, y, z) positions of
the transmitter for every trace.}

\item{\code{coordref}}{A length-3 numeric vector containing the coordinates of a
local reference.}

\item{\code{freq}}{A length-one numeric vector corresponding to the GPR antennae
frequency (in MHz).}

\item{\code{dz}}{A length-one numeric vector corresponding to the time or depth
sampling step.}

\item{\code{dx}}{A length-one numeric vector corresponding to the trace step.}

\item{\code{antsep}}{A length-one numeric vector corresponding to the antenna
separation.}

\item{\code{name}}{A length-one character vector containing the name of the GPR
data.}

\item{\code{description}}{A length-one character vector containing the description 
of the GPR data.}

\item{\code{filepath}}{A length-one character vector containing the file path of 
the original GPR data.}

\item{\code{depthunit}}{A length-one character vector corresponding to the 
time/depth unit (e.g., "ns", "m").}

\item{\code{posunit}}{A length-one character vector corresponding to the 
(x, y)-unit (e.g., "m").}

\item{\code{surveymode}}{A length-one character vector containing the survey mode
(e.g., "Reflection", "CMP")}

\item{\code{date}}{A length-one character vector containing the date of the survey
in the format "yyyy-mm-dd".}

\item{\code{crs}}{A length-one character vector containing the coordinate 
reference system following the R notation of proj4string 
from the PROJ.4 library.}

\item{\code{proc}}{A length-varying character vector whose each element correspond
to a processing step applied to the data.}

\item{\code{vel}}{A list containing the velocity model.}

\item{\code{delineations}}{A list containing delineated structures.}

\item{\code{hd}}{A list containing less relevant additional informations.}
}}

