% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion_GPRvirtual.R
\name{as.matrix}
\alias{as.matrix}
\alias{as.matrix,GPRvirtual-method}
\alias{as.array}
\alias{as.array,GPRvirtual-method}
\alias{as.vector,GPRvirtual-method}
\alias{as.numeric}
\alias{as.numeric,GPRvirtual-method}
\alias{as.integer}
\alias{as.integer,GPRvirtual-method}
\alias{as.GPR.data.frame}
\alias{as.GPR.matrix}
\alias{as.GPR.list}
\title{Coercion to matrix, vector or numeri}
\usage{
as.matrix(x, ...)

\S4method{as.matrix}{GPRvirtual}(x)

\S4method{as.array}{GPRvirtual}(x, ...)

\S4method{as.vector}{GPRvirtual}(x, mode = "any")

\S4method{as.numeric}{GPRvirtual}(x, ...)

\S4method{as.integer}{GPRvirtual}(x, ...)

as.GPR.data.frame(x, ...)

as.GPR.matrix(x, ...)

as.GPR.list(x, ...)
}
\arguments{
\item{x}{[\code{GPR}]}

\item{...}{Not used}

\item{mode}{Not used}
}
\description{
Coercion to matrix, vector or numeri

Coercion from data.frame to GPR

Coercion from matrix to GPR

Coercion from list to GPR
}
\concept{coercion}
