% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R, R/ClassGPRset.R, R/ClassGPRcube.R
\name{as.matrix}
\alias{as.matrix}
\alias{as.vector}
\alias{as.numeric}
\alias{as.integer}
\alias{as.GPR.data.frame}
\alias{as.GPR.matrix}
\alias{as.GPR.list}
\alias{as.GPRset.GPR}
\alias{as.raster}
\title{Coercion to matrix}
\usage{
\S4method{as.matrix}{GPR}(x)

\S4method{as.vector}{GPR}(x, mode = "any")

\S4method{as.numeric}{GPR}(x, ...)

\S4method{as.integer}{GPR}(x, ...)

as.GPR.data.frame(x, ...)

as.GPR.matrix(x, ...)

as.GPR.list(x, ...)

as.GPRset.GPR(x, ...)

\S4method{as.raster}{GPRslice}(x)

\S4method{as.raster}{GPRcube}(x)
}
\description{
Coercion to matrix

Coercion to vector

Coercion to numeric

Coercion to integer

Coercion from data.frame to GPR

Coercion from matrix to GPR

Coercion from list to GPR

Coercion from matrix to GPR

Coercion to rasterstack

Coercion to rasterstack
}
