% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion_GPRvirtual.R, R/coercion_GPRset.R,
%   R/coercion_spatial.R, R/coercion_GPRcube.R
\name{as.matrix}
\alias{as.matrix}
\alias{as.matrix,GPRvirtual-method}
\alias{as.array}
\alias{as.array,GPRvirtual-method}
\alias{as.vector,GPRvirtual-method}
\alias{as.numeric}
\alias{as.numeric,GPRvirtual-method}
\alias{as.integer}
\alias{as.integer,GPRvirtual-method}
\alias{as.GPR.data.frame}
\alias{as.GPR.matrix}
\alias{as.GPR.list}
\alias{as.GPRset.data.frame}
\alias{as.GPRset.matrix}
\alias{as.GPRset.array}
\alias{as.GPRset.GPR}
\alias{as.raster}
\alias{as.GPRcube.array}
\alias{as.GPRcube.list}
\title{Coercion to matrix, vector or numeri}
\usage{
as.matrix(x, ...)

\S4method{as.matrix}{GPRvirtual}(x)

\S4method{as.array}{GPRvirtual}(x, ...)

\S4method{as.vector}{GPRvirtual}(x, mode = "any")

\S4method{as.numeric}{GPRvirtual}(x, ...)

\S4method{as.integer}{GPRvirtual}(x, ...)

as.GPR.data.frame(x, ...)

as.GPR.matrix(x, ...)

as.GPR.list(x, ...)

as.GPRset.data.frame(x, ...)

as.GPRset.matrix(x, ...)

as.GPRset.array(x, ...)

as.GPRset.GPR(x, ...)

\S4method{as.raster}{GPRslice}(x)

\S4method{as.raster}{GPRcube}(x)

as.GPRcube.array(x, ...)

as.GPRcube.list(x, ...)
}
\arguments{
\item{x}{(\code{GPR})}

\item{...}{Not used}

\item{mode}{Not used}
}
\description{
Coercion to matrix, vector or numeri

Coercion from data.frame to GPR

Coercion from matrix to GPR

Coercion from list to GPR

Coercion from data.frame to GPRset

Coercion from matrix to GPRset

Coercion from array to GPRset

Coercion from GPR to GPRset

Coercion to SpatRaster

Coercion to SpatRaster (multi-layer)

Coercion from matrix to GPR

Coercion from list to GPR
}
\concept{coercion}
