% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRslice.R
\docType{class}
\name{GPRslice-class}
\alias{GPRslice-class}
\title{Class GPRslice}
\description{
An S4 class to represent time/depth slices of 
ground-penetrating radar (GPR) data.
Array of dimension \eqn{1 \times m \time p}  
(\eqn{m} traces or A-scans along \eqn{x}, 
and \eqn{p} traces or A-scans along \eqn{y}), with grid cell sizes 
\eqn{dx}, \eqn{dy}.
We assume that the unit along y is the same as the unit along x.
}
