% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRsurvey.R
\docType{class}
\name{GPRsurvey-class}
\alias{GPRsurvey-class}
\title{GPR survey data}
\description{
An S4 class to represent a collection of \eqn{k} GPR data.
An object of the class \code{GPRsurvey} does only contain the links to
the GPR data files as well as information about the GPR data such as
coordinates, frequencies, antenna separation etc.
The methods of the \code{GPRsurvey} class exclusively manipulate the trace
(A-scan) coordinates. More specifically, the methods currently allow to:
\itemize{
\item display map of the GPR measurement lines that can be superimposed
on any raster or vector data,
\item display GPR data in three-dimensional interactive graphics
(e.g., fence  diagram) with openGL,
\item add or interpolate trace coordinates (the elevation coordinates
can be interpolated from raster data),
\item georeference the trace coordinates (e.g., from a local coordinate
reference system to a regional reference coordinate system),
\item project the GPR data to any coordinate reference system,
\item estimate the spatial shift between two (parallel)GPR profiles.
}
}
\section{Slots}{

\describe{
\item{\code{version}}{(\code{character[1]}) Version of RGPR.}

\item{\code{name}}{(\code{character[1]}) Name of the GPR survey}

\item{\code{desc}}{(\code{character[1]}) Description of the GPR survey}

\item{\code{path}}{(\code{character[1]}) File path of the original GPR survey}

\item{\code{names}}{(\code{character[k]}) Names of the GPR data.}

\item{\code{descs}}{(\code{character[k]}) Descriptions of the GPR data.}

\item{\code{paths}}{(\code{character[k]}) File paths of the original GPR data.}

\item{\code{modes}}{(\code{character[k]}) Survey modes of the GPR data
(e.g., \code{"CO"}, \code{"CMP"}).}

\item{\code{dates}}{(\code{Date[k]}) Date of the GPR data (class \code{Date},
e.g., \code{Sys.Date()}).}

\item{\code{freqs}}{(\code{numeric[k]}) Antennae frequency of the GPR data
(in MHz).}

\item{\code{antseps}}{(\code{numeric[k]}) Antennae separation of the GPR data}

\item{\code{spunit}}{(\code{character[1]}) Spatial units.}

\item{\code{crs}}{(\code{character[1]}) Coordinate reference system in wkt format (sf package)}

\item{\code{coords}}{(\code{list[k]}) GPR data coordinates}

\item{\code{intersections}}{(\code{list[k]}) Intersections coordinates of the GPR data}

\item{\code{markers}}{(\code{list[k]}) Fiducial markers of the GPR data}

\item{\code{nz}}{(\code{integer[k]}) Number of samples in each GPR data.}

\item{\code{zlengths}}{(\code{numeric[k]}) Range along \code{z} in each GPR data.}

\item{\code{zunits}}{(\code{character[1]}) Unit of \code{z}.}

\item{\code{nx}}{(\code{integer[k]}) Number of traces in each GPR data.}

\item{\code{xlengths}}{(\code{numeric[k]}) Length of the GPR data.}

\item{\code{transf}}{(\code{numeric[2]}) Translation vector before rotation, after and rotation angle}

\item{\code{view}}{(\code{logical[1]}) TRUE if only a view.}
}}

