% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPRsurvey.R
\name{GPRsurvey}
\alias{GPRsurvey}
\title{Create an object of the class GPRsurvey}
\usage{
GPRsurvey(x, verbose = FALSE)
}
\arguments{
\item{x}{[\code{character(k)}]     Vector of \code{k} file paths of GPR data.}

\item{verbose}{[\code{logical(1)}] If \code{TRUE} the function prints some
information.}
}
\description{
Create an object of the class GPRsurvey using a vector of GPR data filepath
}
