% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRsurvey.R
\name{GPRsurvey}
\alias{GPRsurvey}
\title{Create an object of the class GPRsurvey}
\usage{
GPRsurvey(LINES, verbose = TRUE)
}
\description{
Create an object of the class GPRsurvey using a vector of GPR data filepath
}
