% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRsurvey.R
\name{GPRsurvey.as.SpatialLines}
\alias{GPRsurvey.as.SpatialLines}
\alias{GPRsurvey.as.SpatialPoints}
\alias{as.SpatialLines}
\alias{as.SpatialPoints}
\title{Coerce to SpatialLines}
\usage{
as.SpatialLines(x, ...)

as.SpatialPoints(x, ...)
}
\description{
Coerce to SpatialLines

Coerce to SpatialPoints
}

