% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRsurvey.R
\docType{methods}
\name{GPRsurvey.as.SpatialLines}
\alias{GPRsurvey.as.SpatialLines}
\alias{GPRsurvey.as.SpatialPoints}
\title{Coerce to SpatialLines}
\usage{
\S4method{as.SpatialLines}{GPRsurvey}(x)

\S4method{as.SpatialPoints}{GPR}(x)
}
\description{
Coerce to SpatialLines

Coerce to SpatialPoints
}

