% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{interpPos}
\alias{interpPos}
\alias{interpPos}
\title{Interpolate the trace position.}
\usage{
\S4method{interpPos}{GPR}(x, topo, plot = FALSE, r = NULL, ...)
}
\arguments{
\item{x}{An object of the class GPR.}

\item{topo}{A (mx4) numeric matrix, with m the number of traces 
in x. The columns names of topo must be
"E", "N", "Z" and "TRACE".}

\item{plot}{A length-one boolean vector. If TRUE some control
plots are displayed.}
}
\description{
Interpolate the trace position.
}
