% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpSlices.R
\name{interpSlices}
\alias{interpSlices}
\alias{interpSlices,GPRsurvey-method}
\title{Interpolate horizontal slices}
\usage{
interpSlices(
  obj,
  dx = NULL,
  dy = NULL,
  dz = NULL,
  h = 6,
  extend = c("bbox", "obbox", "chull", "buffer"),
  buffer = NULL,
  shp = NULL,
  rot = FALSE,
  verbose = TRUE
)

\S4method{interpSlices}{GPRsurvey}(
  obj,
  dx = NULL,
  dy = NULL,
  dz = NULL,
  h = 6,
  extend = c("bbox", "obbox", "chull", "buffer"),
  buffer = NULL,
  shp = NULL,
  rot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{(\code{GPRsurvey})}

\item{dx}{(\code{numeric[1]}) x-resolution}

\item{dy}{(\code{numeric[1]}) y-resolution}

\item{dz}{(\code{numeric[1]}) z-resolution}

\item{h}{(\code{numeric[1]}) FIXME: Number of levels in MBA hierarchy (see function...)}

\item{extend}{(\code{character[1]}) FIXME: Method to define interpolation extent.}

\item{buffer}{(\code{numeric[1]}) FIXME: Buffer distance around survey lines.}

\item{shp}{(\code{matrix[n,2]|list[2]|sf}) FIXME: Shape/polygon defining interpolation bounds.}

\item{rot}{(\code{logical[1]|numeric[1]}) If \code{TRUE} the GPR lines are
fist rotated such to minimise their axis-aligned bounding box.
If \code{rot} is numeric, the GPR lines is rotated first
rotated by \code{rot} (in radian).}

\item{verbose}{(\code{logical[1]}) If TRUE, verbose.}
}
\value{
(\code{GPRcube|GPRslice})
}
\description{
Interpolate horizontal slices
}
\concept{3D}
