% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRcube.R
\docType{methods}
\name{interpSlices}
\alias{interpSlices}
\title{Interpolate horizontal slices}
\usage{
\S4method{interpSlices}{GPRsurvey}(x, dx = NULL, dy = NULL,
  dz = NULL, h = 6)
}
\description{
Interpolate horizontal slices
}
