% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpSlices.R
\docType{methods}
\name{interpSlices}
\alias{interpSlices}
\alias{interpSlices,GPRsurvey-method}
\title{Interpolate horizontal slices}
\usage{
interpSlices(x, dx = NULL, dy = NULL, dz = NULL, h = 6,
  extend = TRUE, buffer = NULL)

\S4method{interpSlices}{GPRsurvey}(x, dx = NULL, dy = NULL,
  dz = NULL, h = 6, extend = TRUE, buffer = NULL)
}
\arguments{
\item{x}{[\code{GPRsurvey}]}

\item{dx}{[\code{numeric(1)}] x-resolution}

\item{dy}{[\code{numeric(1)}] y-resolution}

\item{dz}{[\code{numeric(1)}] z-resolution}

\item{h}{[\code{numeric(1)}] FIXME: see function...}

\item{extend}{[\code{numeric(1)}] FIXME: see function...}

\item{buffer}{[\code{numeric(2)}] FIXME: see function...}
}
\description{
Interpolate horizontal slices
}
