% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCRSGeographic.R
\name{isCRSGeographic}
\alias{isCRSGeographic}
\alias{isCRSGeographic,numeric-method}
\alias{isCRSGeographic,integer-method}
\alias{isCRSGeographic,character-method}
\alias{isCRSGeographic,crs-method}
\alias{isCRSGeographic,GPR-method}
\alias{isCRSGeographic,GPRsurvey-method}
\title{Check wether the Coordinate Reference Systems is lon/lat}
\usage{
isCRSGeographic(x)

\S4method{isCRSGeographic}{numeric}(x)

\S4method{isCRSGeographic}{integer}(x)

\S4method{isCRSGeographic}{character}(x)

\S4method{isCRSGeographic}{crs}(x)

\S4method{isCRSGeographic}{GPR}(x)

\S4method{isCRSGeographic}{GPRsurvey}(x)
}
\arguments{
\item{x}{(\code{character}) CRS (one or more)}
}
\value{
(\code{logical}) \code{TRUE} if lon/lat else \code{FALSE}
}
\description{
Returns \code{TRUE} or \code{FALSE}.
}
