% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCRSLonLat.R
\name{isCRSLonLat}
\alias{isCRSLonLat}
\title{Check wether the Coordinate Reference Systems is lon/lat}
\usage{
isCRSLonLat(x)
}
\arguments{
\item{x}{[\code{character}] CRS (one or more)}
}
\value{
[\code{logical}] \code{TRUE} if lon/lat else \code{FALSE}
}
\description{
Returns \code{TRUE} or \code{FALSE}.
}
