% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R, R/ClassGPRsurvey.R,
%   R/ClassGPRcube.R
\docType{methods}
\name{isLengthUnit}
\alias{isLengthUnit}
\title{Return TRUE if the data are a function of length}
\usage{
\S4method{isLengthUnit}{GPR}(x)

\S4method{isLengthUnit}{GPRsurvey}(x)

\S4method{isLengthUnit}{GPRslice}(x)
}
\description{
Return TRUE if the data are a function of length

Return TRUE is the data are a function of length

Return TRUE if the data are a function of length
}
