% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R, R/ClassGPRsurvey.R
\docType{methods}
\name{isTimeUnit}
\alias{isTimeUnit}
\title{Return TRUE is the data are a function of time}
\usage{
\S4method{isTimeUnit}{GPR}(x)

\S4method{isTimeUnit}{GPRsurvey}(x)
}
\description{
Return TRUE is the data are a function of time

Return TRUE is the data are a function of time
}
