% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isZunitTime.R
\name{isZunitTime}
\alias{isZunitTime}
\alias{isZunitTime,GPRvirtual-method}
\alias{isZunitTime,GPRsurvey-method}
\alias{isZunitLength}
\alias{isZunitLength,GPRsurvey-method}
\alias{isZunitLength,GPRvirtual-method}
\title{Oriented bounding box}
\usage{
isZunitTime(x)

\S4method{isZunitTime}{GPRvirtual}(x)

\S4method{isZunitTime}{GPRsurvey}(x)

isZunitLength(x)

\S4method{isZunitLength}{GPRsurvey}(x)

\S4method{isZunitLength}{GPRvirtual}(x)
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}
}
\value{
[\code{matrix(5,2)}] The coordinates of the corners of the oriented 
         bounding box, whereby the last row is identical to the first row.
         FIXME!!
}
\description{
source "whuber" from stackexchange.com
}
