% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\name{length,GPR-method}
\alias{length,GPR-method}
\title{Return length of a GPR object (number of traces)}
\usage{
\S4method{length}{GPR}(x)
}
\arguments{
\item{x}{(\code{GPR})}
}
\description{
Return length of a GPR object (number of traces)
}
\concept{array}
