% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R, R/plot.R
\name{lines}
\alias{lines}
\alias{lines.GPR}
\alias{lines.GPRsurvey}
\title{Add a GPR trace on a plot}
\usage{
\method{lines}{GPR}(x, relTime0 = FALSE, ...)

\method{lines}{GPRsurvey}(x, ...)
}
\description{
Plot the GPR survey as lines
}
