% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R, R/ClassGPRsurvey.R
\name{lines}
\alias{lines}
\alias{lines.GPR}
\alias{lines}
\alias{lines.GPRsurvey}
\title{Add a GPR trace on a plot}
\usage{
\method{lines}{GPR}(x, ...)

\method{lines}{GPRsurvey}(x, ...)
}
\description{
Add a GPR trace on a plot

Plot the GPR survey as lines
}
