% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GPR.R, R/plot_GPRsurvey.R
\name{lines}
\alias{lines}
\alias{lines.GPR}
\alias{lines.GPRsurvey}
\title{Add a GPR trace on a plot}
\usage{
\method{lines}{GPR}(x, relTime0 = FALSE, ...)

\method{lines}{GPRsurvey}(x, ...)
}
\arguments{
\item{x}{(\verb{GPR class})}

\item{relTime0}{(\code{logical[1]}) If \code{TRUE}, shift \code{x} to time-
zero.}

\item{...}{Additional parameters to be passed to \code{\link[=lines]{lines()}}.}
}
\description{
Add a GPR trace on a plot
}
