% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{llToUTM}
\alias{llToUTM}
\title{latitude-longitude to UTM}
\usage{
llToUTM(lat, lon, zone = NULL, south = NULL)
}
\description{
see https://stackoverflow.com/a/30225804
https://stackoverflow.com/questions/18639967/converting-latitude-and-longitude-points-to-utm
}
