% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupMath_GPRvirtual.R
\docType{methods}
\name{Math,GPRvirtual-method}
\alias{Math,GPRvirtual-method}
\alias{Math-GPRvirtual-method}
\title{Basic mathematical functions}
\usage{
\S4method{Math}{GPRvirtual}(x)
}
\arguments{
\item{x}{An object of the class GPRvirtual}
}
\description{
Methods for the base Math methods \link[methods]{S4groupGeneric}
}
\details{
Currently implemented methods include:
\itemize{
 \item{"abs", "sign", "sqrt", "ceiling", "floor", "trunc", "cummax", 
       "cummin", "cumprod", "cumsum", "log", "log10", "log2", "log1p", 
       "acos", "acosh", "asin", "asinh", "atan", "atanh", "exp", "expm1", 
       "cos", "cosh", "cospi", "sin", "sinh", "sinpi", "tan", "tanh", 
       "tanpi", "gamma", "lgamma", "digamma", "trigamma"}
 }
}
