\name{Math-methods}
\docType{methods}
\alias{Math-methods}
\alias{Math,GPR-method}
\title{ Mathematical functions applicable to the GPR objects}
\description{
	 Mathematical functions applicable to the GPR objects
}
\section{Methods}{
\describe{

\item{\code{signature(x = "GPR")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
