% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_GPR.R
\name{mean,GPRvirtual-method}
\alias{mean,GPRvirtual-method}
\title{Mean of a GPR object}
\usage{
\S4method{mean}{GPRvirtual}(x, ...)
}
\arguments{
\item{x}{(\code{GPR})}

\item{...}{Additional parameters to be passed (see \code{\link[=mean]{mean()}}).}
}
\description{
Mean of all the values contained in \code{x}.
}
\concept{statistics}
