% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRvirtual.R
\docType{methods}
\name{mean,GPRvirtual-method}
\alias{mean,GPRvirtual-method}
\title{Mean of GPR data}
\usage{
\S4method{mean}{GPRvirtual}(x)
}
\description{
Mean of GPR data
}
\examples{
data(frenkeLine00)
mean(frenkeLine00)
}
