% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{median,GPR-method}
\alias{median,GPR-method}
\title{Median of GPR data}
\usage{
\S4method{median}{GPR}(x)
}
\description{
Median of GPR data
}
\examples{
data(frenkeLine00)
median(frenkeLine00)
}
