% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{name}
\alias{name}
\alias{name<-}
\alias{name}
\alias{name<-}
\title{Name of the GPR data}
\usage{
\S4method{name}{GPR}(x)

\S4method{name}{GPR}(x) <- value
}
\description{
Name of the GPR data
}
