% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRvirtual.R
\docType{methods}
\name{ncol,GPRvirtual-method}
\alias{ncol,GPRvirtual-method}
\title{Number of columns of an object of the class \code{GPR}}
\usage{
\S4method{ncol}{GPRvirtual}(x)
}
\description{
= number of traces.
}
\examples{
data(frenkeLine00)
ncol(frenkeLine00)
}
