% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRvirtual.R
\docType{methods}
\name{nrow,GPRvirtual-method}
\alias{nrow,GPRvirtual-method}
\title{Number of rows of an object of the class \code{GPR}}
\usage{
\S4method{nrow}{GPRvirtual}(x)
}
\description{
= number of time samples
}
\examples{
data(frenkeLine00)
nrow(frenkeLine00)
}
