% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{palCol}
\alias{palCol}
\title{Return color from palette}
\usage{
palCol(x, col = palGPR(n = 101), sym = TRUE)
}
\arguments{
\item{x}{Values}

\item{col}{[\code{character}] Colors to be used.}
}
\description{
Return color from palette
}
