% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{palGPR}
\alias{palGPR}
\alias{plotPal}
\alias{displayPalGPR}
\title{Plot single colour palette}
\usage{
palGPR(colPal = "default", n = 101, power = 1, returnNames = FALSE)

plotPal(col, border = NA)

displayPalGPR()
}
\description{
source: vignette of the R-package "colorspace" (Color Space Manipulation)
}
\examples{
plotPal(palGPR("hcl_5"))
displayPalGPR()
}
