% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{palGPR}
\alias{palGPR}
\alias{palPlot}
\alias{palDisplay}
\title{Color palettes for GPR data}
\usage{
palGPR(pal = "default", n = 101, power = 1, returnNames = FALSE)

palPlot(col, border = NA)

palDisplay()
}
\arguments{
\item{pal}{(\code{character[1]}) Name of the color palette.}

\item{n}{(\code{integer[1]}) Number of colors to be in the palette.}

\item{power}{(\code{integer[1]}) Control parameter determining how chroma
should be increased (1 = linear, 2 = quadratic, etc.).}

\item{returnNames}{(\code{logical[1]}) If \code{TRUE}, returns only the
color palette names.}

\item{col}{(\code{character}) Colors to `be plotted.}

\item{border}{(\code{character}) color for rectangle border(s).
The default means par("fg"). Use border = NA to omit borders.
If there are shading lines, border = TRUE means use the same colour
for the border as for the shading lines.}
}
\description{
source: vignette of the R-package "colorspace" (Color Space Manipulation)
}
\examples{
palPlot(palGPR("hcl_5"))
palDisplay()
}
\concept{plot}
