% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papply.R
\name{papply}
\alias{papply}
\title{Apply processing to GPRsurvey object}
\usage{
\S4method{papply}{GPRsurvey}(x, prc = NULL)

\S4method{papply}{GPR}(x, prc = NULL)
}
\description{
Apply processing to GPRsurvey object
}
