% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRsurvey.R
\docType{methods}
\name{papply}
\alias{papply}
\title{Apply processing to GPRsurvey object}
\usage{
\S4method{papply}{GPRsurvey}(x, prc = NULL)
}
\description{
Apply processing to GPRsurvey object
}
