% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papply.R
\name{papply}
\alias{papply}
\alias{papply,GPRsurvey-method}
\title{Apply bacth processing to GPRsurvey object}
\usage{
papply(obj, prc = NULL)

\S4method{papply}{GPRsurvey}(obj, prc = NULL)

\S4method{papply}{GPR}(obj, prc = NULL)
}
\arguments{
\item{obj}{(\code{GPR}) An object of the class GPR.}

\item{prc}{(\code{list}) A list of processing funtions with their arguments}
}
\value{
(\code{GPRsurvey})
}
\description{
Apply bacth processing to GPRsurvey object
}
\concept{processing}
