% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_path.R
\name{pathLength}
\alias{pathLength}
\title{Path length}
\usage{
pathLength(xy, lonlat = FALSE)
}
\arguments{
\item{xy}{[\code{matrix}] Each column corresponding to a coordinate}

\item{lonlat}{[\code{logical(1)}] If \code{TRUE} computes geodistance. 
Otherwise Euclidean distance.}
}
\value{
Path length.
}
\description{
Path length
}
