% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_path.R
\name{pathRelPos}
\alias{pathRelPos}
\title{Relative position on a path}
\usage{
pathRelPos(xy, lonlat = FALSE)
}
\arguments{
\item{xy}{(\code{matrix}) Each column corresponding to a coordinate}

\item{lonlat}{(\code{logical[1]}) If \code{TRUE} computes geodistance.
Otherwise Euclidean distance.}
}
\value{
Relative position of the coordinates (cumulative distance).
}
\description{
Relative position of each coordinates (knots) along a path (polyline).
}
