% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GPR.R, R/plot_GPRset.R,
%   R/plot_GPRsurvey.R, R/plot_GPRcube.R, R/plot_GPRslice.R
\name{plot}
\alias{plot}
\alias{plot.GPR}
\alias{contour}
\alias{contour.GPR}
\alias{plot.GPRset}
\alias{plot.GPRsurvey}
\alias{plot.GPRcube}
\alias{plot.GPRslice}
\title{Plot the GPR object.}
\usage{
\method{plot}{GPR}(
  x,
  col = NULL,
  NAcol = "white",
  clim = NULL,
  add = FALSE,
  relTime0 = FALSE,
  note = NULL,
  addMarkers = TRUE,
  addAnn = TRUE,
  addTime0 = TRUE,
  addDepth0 = TRUE,
  addAmpl0 = TRUE,
  addTopo = FALSE,
  xclip = NULL,
  ratio = 1,
  barscale = TRUE,
  wsize = 1,
  wside = 1,
  pdfName = NULL,
  horiz = TRUE,
  ...
)

\method{contour}{GPR}(
  x,
  col = NULL,
  NAcol = "white",
  clim = NULL,
  add = FALSE,
  relTime0 = FALSE,
  note = NULL,
  addMarkers = TRUE,
  addAnn = TRUE,
  addTime0 = TRUE,
  addDepth0 = TRUE,
  addAmpl0 = TRUE,
  addTopo = FALSE,
  xclip = NULL,
  ratio = 1,
  barscale = TRUE,
  wsize = 1,
  wside = 1,
  pdfName = NULL,
  ...
)

\method{plot}{GPRset}(
  x,
  col = NULL,
  NAcol = "white",
  clim = NULL,
  add = FALSE,
  relTime0 = FALSE,
  note = NULL,
  addMarkers = TRUE,
  addAnn = TRUE,
  addTime0 = TRUE,
  addDepth0 = TRUE,
  addAmpl0 = TRUE,
  addTopo = FALSE,
  xclip = NULL,
  ratio = 1,
  barscale = TRUE,
  wsize = 1,
  wside = 1,
  pdfName = NULL,
  ...
)

\method{plot}{GPRsurvey}(x, ...)

\method{plot}{GPRcube}(x, ...)

\method{plot}{GPRslice}(
  x,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  asp = 1,
  xclip = NULL,
  clim = NULL,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class \code{GPR}}

\item{col}{[\code{character}] FIXME Color palette.}

\item{NAcol}{color of NA values}

\item{clim}{[\code{numeric(2)}] Value range for the color palette.}

\item{add}{[\code{logical(1)}] If \code{TRUE}, add to current plot.}

\item{relTime0}{logical. If \code{TRUE}, adjust vertical axis to time-zero.
If time-zero varies from trace to trace, the vertical axis
is adjusted to the mean time-zero. Apply first the function
\code{time0Cor()} to shift the traces to their time-zero.}

\item{note}{note of something...}

\item{addMarkers}{logical. Add fiducial marks}

\item{addAnn}{logical. Add GPR annotations (line intersections)}

\item{addTime0}{logical. Add time-zero line}

\item{addDepth0}{logical. Add depth-zero line}

\item{addAmpl0}{logical. Add an horizontal line on 1D plot}

\item{addTopo}{logical. For 2D plot, add topography (if the data are sampled
in time unit, the data are migrated with a static
migration)}

\item{xclip}{numeric. If length-one numeric vector, clip the amplitudes 
larger than \code{xclip} and smaller than \code{-xclip}.
If length-two numeric vector, xclip the amplitudes
smaller than \code{xclip[1]} and larger than 
\code{xclip[2]}. Per default, values below the 
0.01-quantile and above the 0.99-quantile are clipped.
If \code{xclip = FALSE} the data are not clipped.}

\item{ratio}{logical.}

\item{barscale}{logical. Add a colorbar scale}

\item{wsize}{length-one numeric. Size of the wiggles (default = \code{1}).}

\item{wside}{length-one numeric. If positive the right part of the wavelet
is colored. If negative, the left part of the wavelet is
colored.}

\item{pdfName}{length-one character. Name/path of the PDF to export 
without extension}

\item{...}{additional arguments passed to the plotting methods 
\code{\link[graphics]{plot}} for 1D plot and 
\code{\link[plot3D]{Image}} for 2D plot. See also  \code{details}.}

\item{main}{[\code{character(1)}] A main title for the plot,.}

\item{xlab}{[\code{character(1)}] A title for the x axis.}

\item{ylab}{[\code{character(1)}] A title for the y axis.}

\item{asp}{[\code{numeric(1)}] The y/x aspect ratio.}
}
\description{
\code{plot}: If the GPR object consists of a single trace, wiggle plot 
is shown. For CMP, the position of the traces on the x-axis is defined
by the antenna separation (\code{antsep(x)}).
}
\details{
Additional arguments
\itemize{
  \item \code{type}:
  \itemize{
     \item 1D plot: \code{p}, \code{l}, 
            \code{b}, \code{c}, \code{o}, \code{h}, \code{s}, \code{S}, 
            \code{n} (see argument \code{type} in
            \code{\link[graphics]{plot}}).
     \item  2D plot: \code{"raster"} (default), \code{"wiggles"} or
            \code{contour}.
  }
  \item \code{col}: Color. Default 1D: "black". 2D: \code{palGPR(n = 101)}
}
}
\examples{
\dontrun{
## PLOT GPRsurvey data (trace positions of several GPR data)
plot(x, parLines = NULL, parMarkers = NULL, parArrows = NULL, parIntersect = NULL)
# plot only arrows
plot(x, parLines = NULL, parMarkers = NULL, parIntersect = NULL)  
# plot only lines
plot(x, parMarkers = NULL, parArrows = NULL, parIntersect = NULL) 
# plot only markers
plot(x, parLines = NULL, parArrows = NULL, parIntersect = NULL)   
# plot only intersections
plot(x, parLines = NULL, parArrows = NULL, parMarkers = NULL)   
# plot only markers with other parameters
plot(x, parLines = NULL, parArrows = NULL, parMarkers = list(pch = 3), parIntersect = NULL)   
plot(x, parMarkers = NULL)
}
}
