% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPRset.R, R/ClassGPRcube.R, R/plot.R
\name{plot}
\alias{plot}
\alias{plot.GPRset}
\alias{plot.GPRcube}
\alias{plot.GPRslice}
\alias{plot.GPR}
\alias{contour}
\alias{contour.GPR}
\alias{plot.GPRsurvey}
\title{Plot a GPR cube}
\usage{
\method{plot}{GPRset}(
  x,
  add = FALSE,
  relTime0 = FALSE,
  note = NULL,
  addFid = TRUE,
  addAnn = TRUE,
  addTime0 = TRUE,
  addDepth0 = TRUE,
  addAmpl0 = TRUE,
  addTopo = FALSE,
  clip = NULL,
  ratio = 1,
  barscale = TRUE,
  wsize = 1,
  wside = 1,
  pdfName = NULL,
  ...
)

\method{plot}{GPRcube}(
  x,
  i = NULL,
  j = NULL,
  k = NULL,
  xlim = NULL,
  ylim = NULL,
  zlim = NULL,
  clim = NULL,
  colkey = NULL,
  add = FALSE,
  col = NULL,
  inttype = 2,
  ...
)

\method{plot}{GPRslice}(
  x,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  clim = NULL,
  relCoords = FALSE,
  ...
)

\method{plot}{GPR}(
  x,
  add = FALSE,
  relTime0 = FALSE,
  note = NULL,
  addFid = TRUE,
  addAnn = TRUE,
  addTime0 = TRUE,
  addDepth0 = TRUE,
  addAmpl0 = TRUE,
  addTopo = FALSE,
  add2ndYAxis = TRUE,
  elev = FALSE,
  clip = NULL,
  barscale = TRUE,
  wsize = 1,
  wside = 1,
  pdfName = NULL,
  pngName = NULL,
  NAcol = "white",
  fast = FALSE,
  ...
)

\method{contour}{GPR}(
  x,
  relTime0 = FALSE,
  add = FALSE,
  note = NULL,
  addFid = TRUE,
  addAnn = TRUE,
  addTime0 = TRUE,
  addDepth0 = TRUE,
  addAmpl0 = TRUE,
  addTopo = FALSE,
  clip = NULL,
  asp = 1,
  barscale = TRUE,
  pdfName = NULL,
  ...
)

\method{plot}{GPRsurvey}(x, y, ...)
}
\arguments{
\item{x}{Object of class \code{GPR}}

\item{add}{logical. If \code{TRUE}, add to current plot}

\item{relTime0}{logical. If \code{TRUE}, adjust vertical axis to time-zero.
If time-zero varies from trace to trace, the vertical axis
is adjusted to the mean time-zero. Apply first the function
\code{time0Cor()} to shift the traces to their time-zero.}

\item{addFid}{logical. Add fiducial marks}

\item{addAnn}{logical. Add GPR annotations (line intersections)}

\item{addTime0}{logical. Add time-zero line}

\item{addDepth0}{logical. Add depth-zero line}

\item{addAmpl0}{logical. Add an horizontal line on 1D plot}

\item{addTopo}{logical. For 2D plot, add topography (if the data are sampled
in time unit, the data are migrated with a static
migration)}

\item{clip}{numeric. If length-one numeric vector, clip the amplitudes 
larger than \code{clip} and smaller than \code{-clip}.
If length-two numeric vector, clip the amplitudes
smaller than \code{clip[1]} and larger than 
\code{clip[2]}. Per default, values below the 
0.01-quantile and above the 0.99-quantile are clipped.
If \code{clip = FALSE} the data are not clipped.}

\item{ratio}{logical.}

\item{barscale}{logical. Add a colorbar scale}

\item{wsize}{length-one numeric. Size of the wiggles (default = \code{1}).}

\item{wside}{length-one numeric. If positive the right part of the wavelet
is colored. If negative, the left part of the wavelet is
colored.}

\item{pdfName}{length-one character. Name/path of the PDF to export 
without extension}

\item{...}{additional arguments passed to the plotting methods 
\code{\link[graphics]{plot}} for 1D plot and 
\code{\link[plot3D]{Image}} for 2D plot. See also  \code{details}.}

\item{main}{character. Plot title.}

\item{add2ndYAxis}{logical. Should be a second (i.e., right) y-axis added?}

\item{elev}{logical. If \code{TRUE} the elevation instead of depth is
displayed on y-axis (and therefore, time-to-depth
conversion is applied).}

\item{pngName}{length-one character. Name/path of the PDF to export 
without extension}

\item{NAcol}{lengthe-one vector: color to be used.}

\item{fast}{logical: if \code{TRUE} plots only a subset of the data (max. 
1000 traces) to speed up plotting.}
}
\description{
Plot GPR data cube.

\code{plot}: If the GPR object consists of a single trace, wiggle plot 
is shown. For CMP, the position of the traces on the x-axis is defined
by the antenna separation (\code{antsep(x)}).

Plot GPR suvey lines
}
\details{
i, j and k define the sections of the cube that are plotted. 
With the default value (e.g., i = NULL, the first and the last sections 
of all dimensions are plotted). The additional arguments (...) correspond 
to the arguments of the function 'plot3D::surf3D()' (with argument 'theta' 
and 'phi' you can define the orientation of the cube).

Additional arguments
\itemize{
  \item \code{type}: Possible values for 1D plot: see argument \code{type} in
            \code{\link[graphics]{plot}}.
             For 2D plot: \code{"raster"} (default) or \code{"wiggles"}.
  \item \code{col}: Color. Default 1D: "black". 2D: \code{palGPR(n = 101)}
}
}
