% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R, R/ClassGPRsurvey.R
\name{plot}
\alias{plot}
\alias{plot.GPR}
\alias{plot}
\alias{plot.GPRsurvey}
\title{Plot the GPR object.}
\usage{
\method{plot}{GPR}(x, y, ...)

\method{plot}{GPRsurvey}(x, y, ...)
}
\description{
If the GPR object consists of a single trace, wiggle plot is shown.

Plot GPR suvey lines
}
