% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3DRGL.R
\docType{methods}
\name{plot3DRGL}
\alias{plot3DRGL}
\title{Three-dimensional plot of the GPR data with Open-GL}
\usage{
\S4method{plot3DRGL}{GPR}(x, addTopo = FALSE, clip = NULL,
  normalize = NULL, nupspl = NULL, add = TRUE, xlim = NULL,
  ylim = NULL, zlim = NULL, ...)
}
\description{
Three-dimensional plot of the GPR data with Open-GL
}
