% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCMPhyperbolas.R
\name{plotCMPhyperbolas}
\alias{plotCMPhyperbolas}
\alias{plotCMPhyperbolas,GPR-method}
\title{Plot hyperbolas associated with Vrms velociies}
\usage{
plotCMPhyperbolas(x, ...)

\S4method{plotCMPhyperbolas}{GPR}(x, ...)
}
\arguments{
\item{x}{[\code{GPR class}] A CMP object of the class \code{GPR}}

\item{...}{see arguments of \code{\link[graphics]{matplot}}}
}
\value{
[\code{list}] A list element key \code{antsep} containing a numeric 
       vector of \eqn{n} antenna separation values, and element \code{twt} 
       containing a \eqn{n \times m} matrix, where \eqn{m} is the
       number of hyperbolas (i.e., the number of velocities).
}
\description{
Plot hyperbolas associated with root-mean-square velocities stored in
the CMP data.
}
