% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFast.R
\name{plotFast}
\alias{plotFast}
\title{Faster plot}
\usage{
plotFast(
  x,
  col = palGPR(),
  colorbar = TRUE,
  interpolate = TRUE,
  xlab = NULL,
  ylab = NULL,
  sym = TRUE,
  ylim = NULL,
  xlim = NULL
)
}
\arguments{
\item{x}{[\code{class GPR}]}

\item{col}{Color palette}

\item{colorbar}{[\code{color palette}]}

\item{interpolate}{[\code{logical(1)}]}

\item{xlab}{[\code{charachter(1)}]}

\item{sym}{[\code{logical(1)}] if \code{TRUE} the colorbar is symmetric}

\item{ylim}{[\code{numeric(2)}] the y limits (y1, y2) of the plot.}

\item{xlim}{[\code{numeric(2)}] the x limits (x1, x2) of the plot.}
}
\description{
Function to plot faster GPR data
}
