% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFast.R
\name{plotFast}
\alias{plotFast}
\title{Faster plot}
\usage{
plotFast(x, col = palGPR(), colorbar = TRUE, interpolate = TRUE)
}
\arguments{
\item{x}{[GPR]}

\item{col}{Color palette}

\item{colorbar}{[logical]}

\item{interpolate}{[logical]}
}
\description{
Function to plot faster GPR data
}
