% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raytracing.R
\name{rayCMP}
\alias{rayCMP}
\title{CMP ray tracing}
\usage{
rayCMP(av, h, v)
}
\arguments{
\item{av}{\code{\link{numeric(m)}} antenna separation}

\item{h}{\code{\link{numeric(n)}} Layer heights}

\item{v}{\code{\link{numeric(n)}} Layer velocity}
}
\description{
Smulate ray tracing
}
