% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raytracing.R
\name{raytracehalf}
\alias{raytracehalf}
\title{Ray tracing}
\usage{
raytracehalf(h, v, theta)
}
\arguments{
\item{h}{\code{\link{numeric(n)}} Layer heights}

\item{v}{\code{\link{numeric(n)}} Layer velocity}

\item{theta}{\code{\link{numeric(n)}} Incidence angle above first layer}
}
\description{
Ray tracing
}
