% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readDT1}
\alias{readDT1}
\title{Read ".DT1" GPR files}
\usage{
readDT1(fPath)
}
\arguments{
\item{fPath}{A length-one character vector: the GPR data file path}
}
\value{
A list with all the elements contained in the GPR data file.
}
\description{
This function read *.HD and associated *.DT1 files from Sensors
and Software.
}
