% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDZT.R
\name{readDZG}
\alias{readDZG}
\title{Read GSSI GPS data}
\usage{
readDZG(dsn)
}
\arguments{
\item{dsn}{(\code{character[1]|connection}) data source name:
either the filepath to the GPR data (character),
or an open file connection.}

\item{UTM}{(\code{logical[1]}) If \code{TRUE} project coordinates to
the corresponding UTM zone.}
}
\value{
(\code{data.frame(,5)}) position (\code{x}, \code{y}, \code{z}),
trace id (\code{id}), and time (\code{time}).
}
\description{
Read GSSI GPS data
}
\seealso{
\code{\link[=readDZT]{readDZT()}}, \code{\link[=readDZX]{readDZX()}}
}
