% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDZT.R
\name{readDZX}
\alias{readDZX}
\title{Read GSSI's .dzx file}
\usage{
readDZX(dsn)
}
\arguments{
\item{dsn}{(\verb{character(1)|connection object}) data source name:
either the filepath to the GPR data (character),
or an open file connection.}
}
\value{
(\code{list}) contains the markers, the trace position and
the spatial sampling.
}
\description{
.dzx files are xml files
}
\seealso{
\code{\link[=readDZT]{readDZT()}}, \code{\link[=readDZG]{readDZG()}}
}
