% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDZT.R
\name{readDZX}
\alias{readDZX}
\title{Read GSSI's .dzx file}
\usage{
readDZX(dsn)
}
\arguments{
\item{dsn}{connection or filepath}
}
\value{
a list containing the markers, the trace position and the spatial
        sampling.
}
\description{
.dzx files are xml files
}
