% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGPGGA.R
\name{readGPGGA}
\alias{readGPGGA}
\title{Read GPS file with GPGGA string}
\usage{
readGPGGA(dsn, sep = ",", returnSf = TRUE)
}
\arguments{
\item{dsn}{[\code{character(1)|connection object}] Data source name: 
either the filepath to the GPR data (character),
or an open file connection.}

\item{sep}{[\code{character(1)}] The field separator character
(see\code{\link{read.table}}).}

\item{returnSf}{[\code{logical(1)}] If \code{TRUE} returns an object of class
\code{sf} (see \code{\link{sf}{sf}}). If \code{FALSE} returns
a \code{data.frame}.}
}
\value{
[\code{SpatialPoints}]
}
\description{
Read GPS file with GPGGA string
}
