% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\name{readGPR}
\alias{readGPR}
\title{Read a GPR data file}
\usage{
readGPR(dsn, desc = "", dsn2 = NULL, format = NULL, Vmax = 50,
  fPath, ch = 1, verbose = TRUE)
}
\arguments{
\item{dsn}{data source name: either the filepath to the GPR data (character),
or an open file connection.}

\item{desc}{Short description of the file (character).}

\item{dsn2}{data source name for additional file connection if \code{dsn} is
an open file connection (e.g., open file connection to '*.hd'
file if \code{ds} is an open file connection to a '*.dt1' file).}

\item{format}{lenth-one character vector required if the file extension is
not appearent in the filepath or the connection (either
\code{dt1}, \code{rad}, \code{dzt}, \code{sgy}, \code{iprb},
\code{txt}, \code{rds})}

\item{Vmax}{length-one numeric vector: nominal analog input voltage used 
for the byte to volt transformation. 
It assumes that \code{Vmin = -Vmax}. If \code{Vmax = NULL},
no bytes to Volt transformation is applied.}

\item{fPath}{Filepath (character). DEPRECATED. Use \code{dsn} instead.}

\item{ch}{For multi-frequency GSSI files (*.dzt), which channel is red.}

\item{verbose}{(boolean). If \code{FALSE}, all messages and warnings are
suppressed (use with care).}
}
\value{
The GPR data as object of the class RGPR.
}
\description{
Note: argument \code{fPath} is depreacted. Use \code{dsn} instead.
}
\details{
Supported file format
\itemize{
  \item Sensors & Software file format (*.dt1 , *.hd).
        \code{readGPR(dsn = 'xline.dt1')}
  \item MALA file format (*.rd3, *.rad).
        \code{readGPR(dsn = 'xline.rd3')}
  \item RadSys Zond GPR device (*.sgy). 
        \strong{Note: it is not the SEG-Y file format)}.
        \code{readGPR(dsn = 'xline.sgy')}  
  \item GSSI file format (*.dzt).
        \code{readGPR(dsn = 'xline.dzt')}
  \item ASCII file format (*.txt): either 4-column format 
        (x,t,amplitude) or matrix-format (without header/rownames).
        \code{readGPR(dsn = 'xline.txt')}  
  \item R object file format (*rds). These files are created by saving the
        \code{GPR} object with 
        \code{writeGPR(x, fPath = 'xline.rds', type = "rds")}.
        \code{readGPR(dsn = 'xline.txt')}  
}
}
\examples{
\dontrun{
# argument dsn is a file path
x1 <- readGPR(dsn = "data/RD3/DAT_0052.rd3")
y1 <- readGPR("data/FILE____050.DZT")

# argument dsn is a connection
con <- file("data/RD3/DAT_0052.rd3", "rb")   # binary mode
con2 <- file("data/RD3/DAT_0052.rad", "rt")  # text mode
x2 <- readGPR(dsn = con, dsn2 = con2)
close(con)
close(con2)

con <- file(dsn = "data/FILE____050.DZT", "rb")
y1 <- readGPR(con)
close(con)
}
}
\seealso{
\code{\link{writeGPR}}
}
