% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{readGPR}
\alias{readGPR}
\title{Read a GPR data file}
\usage{
\S4method{readGPR}{character}(fPath, desc = "", coordfile = NULL,
  crs = "", intfile = NULL)
}
\arguments{
\item{fPath}{Filepath (character).}

\item{desc}{Short description of the file (character).}

\item{coordfile}{Filepath of a text file containing the coordinates (x,y,z)
of each traces.}

\item{crs}{Coordinate reference system (character)}

\item{intfile}{Filepath of a text file containing the intersection.}
}
\value{
The GPR data as object of the class RGPR.
}
\description{
Read a GPR data file
}
\examples{
NULL
}

