% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/ClassGPR.R
\docType{methods}
\name{readGPR}
\alias{readGPR}
\alias{readGPR,character-method}
\title{Read a GPR data file}
\usage{
\S4method{readGPR}{character}(fPath, desc = "")
}
\arguments{
\item{fPath}{Filepath (character).}

\item{desc}{Short description of the file (character).}
}
\value{
An object of the class GPR.
}
\description{
Function to read GPR data
}
\details{
Supported format: 
\itemize{
  \item Sensors and Software (.DT1, .HD)
  \item MALA (.rd3, .rad)
  \item SEG-Y for RadSys Zond GPR device (.sgy)
  \item R (.rds)
}
If the antenna separation distance is missing, it will be estimated from
the GPR frequency.
}
\examples{
NULL
}
