% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGPRsurvey.R
\name{readGPRsurvey}
\alias{readGPRsurvey}
\title{Read a GPRsurvey object from an HDF5 file}
\usage{
readGPRsurvey(file)
}
\arguments{
\item{file}{(\code{character(1)}) Path to the \code{.h5} file previously written by
\code{\link{GPRsurvey}} or \code{\link{writeGPR}}.}
}
\value{
Object of class \code{GPRsurvey}.
}
\description{
Reconstructs the lightweight \code{GPRsurvey} index from the \code{/survey}
group of an RGPR HDF5 file.  Individual GPR lines are not loaded until
accessed via \code{[[}.
}
\seealso{
\code{\link[=GPRsurvey]{GPRsurvey()}}, \code{\link[=writeGPR]{writeGPR()}}
}
