% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDT1.R
\name{readGPS}
\alias{readGPS}
\title{Read .GPS files (Sensors and Software)}
\usage{
readGPS(dsn, toUTM = FALSE)
}
\arguments{
\item{dsn}{[\code{character(1)|connection object}]data source name: 
either the filepath to the GPR data (character),
or an open file connection.}
}
\description{
Read .GPS files (Sensors and Software)
}
