% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDT1.R
\name{readGPS}
\alias{readGPS}
\title{Read Sensors and Software .GPS file}
\usage{
readGPS(dsn)
}
\arguments{
\item{dsn}{[\code{character(1)|connection object}] data source name: 
 either the filepath to the GPR data (character),
or an open file connection.}
}
\value{
[\code{data.frame}]
}
\description{
Read Sensors and Software .GPS file
}
\seealso{
\code{\link{readDT1}}, \code{\link{readHD}}
}
