% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{readSGY}
\alias{readSGY}
\title{Read SEG-Y file}
\usage{
readSGY(dsn, ENDIAN = "big")
}
\arguments{
\item{dsn}{data source name: either the filepath to the GPR data (character),
or an open file connection.}

\item{ENDIAN}{The endian-ness ("big" or "little") of the target system for 
the file. Using "swap" will force swapping endian-ness.}
}
\description{
Read SEG-Y file
}
