% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\alias{reverse,GPR-method}
\title{Reverse the trace position.}
\usage{
reverse(x, id = NULL, tol = 0.3, track = TRUE)

\S4method{reverse}{GPR}(x, id = NULL, tol = 0.3, track = TRUE)

\S4method{reverse}{GPRsurvey}(x, id = NULL, tol = 0.3, track = TRUE)
}
\arguments{
\item{x}{(\code{GPR}|\code{GPRsurvey}) Object of the class \code{GPR} or \code{GPRsurvey}}

\item{id}{(\code{NULL}|\code{integer[n]}|\code{zigzag}) Wokrs only if \code{x} is an object of the class \code{GPRsurvey}.
Either the index of the GPR data to reverse (e.g., \code{id = c(1, 3, 4, 7)}) or
\verb{id = "zigzag} to reverse every second GPR data.
If \code{id = NULL} and \code{x} has coordinates, \code{reverse()}
will cluster the GPR data according to their names (e.g.,
cluster 1 = XLINE01, XLINE02, XLINE03; cluster 2 = YLINE01,
YLINE02; cluster 3 = XYLINE1, XYLINE2, XYLINE3) and reverse the
data such that all GPR lines within the same cluster have the
same orientation (up to a tolerance value \code{tol}).}

\item{to}{Length-one numeric vector. Tolerance angle in radian to determine
if the data have the same orientation. The first data of the
cluster is set as reference angle \eqn{\alpha_0}, then for data
\eqn{i} in the same cluster, if \eqn{\alpha_i} is not between
\eqn{\alpha_0 - \frac{tol}{2}} and
\eqn{\alpha_0 + \frac{tol}{2}}, then the data is reversed.}
}
\description{
Reverse the trace position.

Reverse the trace position.
}
\examples{
\dontrun{
# SU class GPRsurvey
SU <- reverse(SU, id = "zigzag")
# identical to above
SU <- reverse(SU, id = seq(from = 2, to = length(SU), by = 2))
}
}
\concept{processing}
