% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R, R/ClassGPRsurvey.R
\docType{methods}
\name{reverse}
\alias{reverse}
\alias{reverse}
\alias{reverse}
\title{Reverse the trace position.}
\usage{
\S4method{reverse}{GPR}(x, id = NULL, tol = 0.3)

\S4method{reverse}{GPRsurvey}(x, id = NULL, tol = 0.3)
}
\description{
Reverse the trace position.

Reverse the trace position.
}
