% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustSmooth.R
\name{robustSmooth}
\alias{robustSmooth}
\title{Robust smoothing}
\usage{
robustSmooth(
  x,
  spar = NULL,
  width,
  method = "PRMMH",
  extrapolate = TRUE,
  minNonNAs = 3
)
}
\arguments{
\item{x}{a numeric vector or (univariate) time series object.}

\item{spar}{smoothing parameter, typically (but not necessarily) in (0,1].
See \verb{[stats:smooth.spline()]}.}

\item{width}{an odd positive integer (>=3) defining the window width
used for fitting.
See \verb{[robfilter:hybrid.filter()]}.}

\item{method}{a (vector of) character string(s) containing the method(s)
to be used for the estimation of the signal level.
Method choice: "MED", "RM",
"MEAN", FMH, "PFMH", "CFMH", "MH", "PRMH", "CRMH", "MMH",
"PRMMH", "CRMMH".
See \verb{[robfilter:hybrid.filter()]}.}

\item{extrapolate}{a logical indicating whether the level estimations
should be extrapolated to the edges of the time series.
See \verb{[robfilter:hybrid.filter()]}.}

\item{minNonNAs}{a positive integer defining the minimum number of
non-missing observationswithin each window (half)
which is required for a 'sensible' estimation.
See \verb{[robfilter:hybrid.filter()]}.}
}
\description{
A wrapper for the functions \code{robfilter::hybrid.filter} and
\code{smooth.spline}
}
