% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{rotatePhase}
\alias{rotatePhase}
\title{Phase rotation}
\usage{
\S4method{rotatePhase}{GPR}(x, phi)
}
\arguments{
\item{x}{A GPR data}

\item{phi}{A length-one numeric vector defining the phase rotation in radian.}
}
\value{
The GPR data with rotated phase.
}
\description{
Rotate the phase of the GPR data by a given angle \code{phi}.
}
