% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complextrace.R
\name{rotatePhase}
\alias{rotatePhase}
\title{Phase rotation

shift the phase of signal by phi (in radian)}
\usage{
rotatePhase(x, phi)
}
\description{
Phase rotation

shift the phase of signal by phi (in radian)
}
