% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLayerVelocities.R
\name{plotVelocityLayers}
\alias{plotVelocityLayers}
\alias{setLayerVelocities}
\title{Plot the delineation on a 2D plot}
\usage{
plotVelocityLayers(
  x,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  ...
)

\S4method{plotVelocityLayers}{GPR}(
  x,
  method = c("linear", "nearest", "pchip", "cubic", "spline", "none"),
  col = NULL,
  ...
)

setLayerVelocities(x, v, twt = NULL, method = "pchip", clean = TRUE)

\S4method{setLayerVelocities}{GPR}(x, v, twt = NULL, method = "pchip", clean = TRUE)
}
\arguments{
\item{x}{GPR object}

\item{method}{[\code{character(1)}] interpolation method. 
One of \code{linear}, \code{nearest}, \code{pchip},
\code{cubic}, \code{spline}.}

\item{v}{[\code{numeric}] velocities (length equal number of delineations plus one)}

\item{twt}{vertical resolution of the migrated data}
}
\description{
Given delineation or the output of the function... and velocity values,
set 2D velocity model
}
