% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_GPRcube.R
\docType{methods}
\name{show,GPRcube-method}
\alias{show,GPRcube-method}
\title{Show some information on the GPRcube object}
\usage{
\S4method{show}{GPRcube}(object)
}
\arguments{
\item{object}{[\code{GPRcube object}]}
}
\description{
Identical to print().
}
