% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{print}
\alias{print}
\alias{print.GPR}
\alias{show}
\title{Print GPR}
\usage{
\method{print}{RGPR}(x, ...)

\S4method{show}{GPR}(object)
}
\description{
Print GPR

Identical to print().
}

