% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R, R/ClassGPRsurvey.R
\docType{methods}
\name{print}
\alias{print}
\alias{print.GPR}
\alias{print.GPRsurvey}
\alias{show}
\alias{show-method}
\title{Print GPR}
\usage{
\method{print}{GPR}(x, ...)

\S4method{show}{GPR}(object)

\method{print}{GPRsurvey}(x, ...)

\S4method{show}{GPRsurvey}(object)
}
\description{
Print GPR

Identical to print().

Print GPR survey

Identical to print().
}

