% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth2D.R
\name{smooth2D}
\alias{smooth2D}
\alias{smooth2D,GPR-method}
\title{Adapative weights 2D smoothing}
\usage{
smooth2D(x, method = c("awsaniso", "aws", "awsp"), ...)

\S4method{smooth2D}{GPR}(x, method = c("awsaniso", "aws", "awsp"), ...)
}
\arguments{
\item{x}{(\code{GPR})}

\item{method}{(\code{character[1]}) \code{awsaniso} stand for anisotropic
adaptive weights smoothing
(call the function \code{\link[adimpro:awsimage]{adimpro::awsaniso()}}),
\code{aws} stand for adaptive weights smoothing using
a local constant model
(call the function \code{\link[adimpro:awsimage]{adimpro::awsimage()}}), and
\code{awsp} stand for adaptive weights smoothing using a
local polynomial models up to a degree of 2.
(call the function \code{\link[adimpro:awsimage]{adimpro::awsimage()}}).}

\item{...}{additional parameters to be passed to the \code{adimpro}
functions: \code{\link[adimpro:awsimage]{adimpro::awsaniso()}},
\code{\link[adimpro:awsimage]{adimpro::awsimage()}}, and
\code{\link[adimpro:awsimage]{adimpro::awsimage()}}.}
}
\value{
(\code{GPR})
}
\description{
A wrapper for the following \code{adimpro} functions:
\code{\link[adimpro:awsimage]{adimpro::awsaniso()}}, \code{\link[adimpro:awsimage]{adimpro::awsimage()}}, and
\code{\link[adimpro:awsimage]{adimpro::awsimage()}}.
}
