% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spInterpToRaster.R
\docType{methods}
\name{spInterpToRaster}
\alias{spInterpToRaster}
\title{spInterpToRaster}
\usage{
\S4method{spInterpToRaster}{GPRsurvey}(x, nx = 100, ny = 100, h = 5,
  ...)
}
\description{
spInterpToRaster
}
