% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{strucTensor}
\alias{strucTensor}
\title{Structure tensor of GPR data}
\usage{
strucTensor(P, winBlur = c(3, 3), winEdge = c(7, 7), winTensor = c(5, 10),
  sdTensor = 2, dxy = c(1, 1), ...)
}
\description{
Structure tensor of GPR data
}

