% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{strucTensor}
\alias{strucTensor}
\alias{.strucTensor}
\title{structure tensor for matrices}
\usage{
.strucTensor(
  P,
  dxy = c(1, 1),
  mask = c(2, 2),
  kBlur = list(n = 3, m = 3, sd = 1),
  kEdge = list(n = 7, m = 7, sd = 1),
  kTensor = list(n = 5, m = 10, sd = 2),
  thresh = 0.1,
  ...
)
}
\description{
structure tensor for matrices
}
