% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_GPR.R, R/subset_GPRsurvey.R
\name{[,GPR,ANY,ANY,ANY-method}
\alias{[,GPR,ANY,ANY,ANY-method}
\alias{[,GPR-method}
\alias{[<-,GPR,ANY,ANY,ANY-method}
\alias{[<-,GPR-method}
\alias{[<-,GPRsurvey,ANY,ANY,ANY-method}
\title{Extract and replace parts of a GPR object}
\usage{
\S4method{[}{GPR,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{GPR,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{GPRsurvey,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{(\code{GPR})}

\item{i}{(\code{integer}) Indices specifying elements to extract or replace.}

\item{j}{(\code{integer}) Indices specifying elements to extract or replace.}

\item{...}{Not used.}

\item{drop}{Not used.}

\item{value}{(\code{numeric}) Value to set.}
}
\value{
(\verb{GPR|})numeric\verb{] Returns a numeric vector only if }x[]`.
}
\description{
Extract parts of a GPR object
}
