% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_GPRcube.R
\name{[,GPRcube,ANY,ANY-method}
\alias{[,GPRcube,ANY,ANY-method}
\alias{[,GPRcube-method}
\title{Extract and replace parts of a GPRcube object}
\usage{
\S4method{[}{GPRcube,ANY,ANY}(x, i, j, k, drop = TRUE)
}
\arguments{
\item{x}{[class \code{GPRcube}]}

\item{i}{[\code{integer}] Indices specifying elements to extract or replace.}

\item{j}{[\code{integer}] Indices specifying elements to extract or replace.}

\item{k}{[\code{integer}] Indices specifying elements to extract or replace.}

\item{drop}{Not used.}
}
\value{
[\code{GPR}|\code{numeric}] Returns a numeric vector
       only if \code{x[]}.
}
\description{
Extract parts of a GPR object
}
