% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_GPRset.R
\name{[,GPRset,ANY,ANY,ANY-method}
\alias{[,GPRset,ANY,ANY,ANY-method}
\alias{[,GPRset-method}
\alias{[<-,GPRset,ANY,ANY-method}
\alias{[<-,GPRset-method}
\title{Extract and replace parts of a GPRset object}
\usage{
\S4method{[}{GPRset,ANY,ANY,ANY}(x, i, j, k, ..., drop)

\S4method{[}{GPRset,ANY,ANY}(x, i, j, k, ...) <- value
}
\arguments{
\item{x}{(\code{GPRset})}

\item{i}{(\code{integer}) Indices specifying elements to extract or replace.}

\item{j}{(\code{integer}) Indices specifying elements to extract or replace.}

\item{k}{(\code{integer}) Indices specifying elements to extract or replace.}

\item{...}{Not used.}

\item{drop}{Not used.}

\item{value}{(\code{numeric}) Value to set.}
}
\value{
(\code{GPR|numeric}) Returns a numeric vector
only if \code{x[]}.
}
\description{
Extract parts of a GPR object
}
