% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_GPRsurvey.R
\docType{methods}
\name{[,GPRsurvey,ANY,ANY,ANY-method}
\alias{[,GPRsurvey,ANY,ANY,ANY-method}
\alias{[,GPRsurvey-method}
\title{Extract and replace parts of a GPRsurvey object}
\usage{
\S4method{[}{GPRsurvey,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{[class \code{GPRsurvey}]}

\item{i}{[\code{integer}] Indices specifying elements to extract or replace.}

\item{j}{[\code{integer}] Not used.}

\item{...}{Not used.}

\item{drop}{Not used.}
}
\value{
[\code{GPR}|\code{numeric}] Returns a numeric vector
       only if \code{x[]}.
}
\description{
Extract parts of a GPRsurvey object
}
