% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{summary,GPR-method}
\alias{summary,GPR-method}
\title{Summary of GPR traces}
\usage{
\S4method{summary}{GPR}(object)
}
\description{
Summary of GPR traces
}
\examples{
data(frenkeLine00)
summary(frenkeLine00)
}
