% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupSummary.R, R/GroupSummary_GPRslice.R,
%   R/GroupSummary_GPRcube.R
\docType{methods}
\name{Summary,GPR-method}
\alias{Summary,GPR-method}
\alias{Summary-GPR-method}
\alias{Summary,GPRslice-method}
\alias{Summary-GPRslice-method}
\alias{Summary,GPRcube-method}
\alias{Summary-GPRcube-method}
\title{Basic summary functions}
\usage{
\S4method{Summary}{GPR}(x, ..., na.rm = FALSE)

\S4method{Summary}{GPRslice}(x, ..., na.rm = FALSE)

\S4method{Summary}{GPRcube}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An object of the class GPR}

\item{...}{further arguments}

\item{na.rm}{[\code{logical(1)}] should missing values be removed?}

\item{x}{An object of the class GPRslice}

\item{...}{further arguments}

\item{na.rm}{[\code{logical(1)}] should missing values be removed?}

\item{x}{An object of the class GPRcube}

\item{...}{further arguments}

\item{na.rm}{[\code{logical(1)}] should missing values be removed?}
}
\description{
Methods for the base Summary methods \link[methods]{S4groupGeneric}

Methods for the base Summary methods \link[methods]{S4groupGeneric}

Methods for the base Summary methods \link[methods]{S4groupGeneric}
}
\details{
Currently implemented methods include:
\itemize{
 \item{all, any, sum, prod, min, max, range}
 }

Currently implemented methods include:
\itemize{
 \item{all, any, sum, prod, min, max, range}
 }

Currently implemented methods include:
\itemize{
 \item{all, any, sum, prod, min, max, range}
 }
}
