% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPRsurvey.R
\name{surveyIntersect}
\alias{surveyIntersect}
\alias{surveyIntersect,GPRsurvey-method}
\title{Compute the survey intersections}
\usage{
\S4method{surveyIntersect}{GPRsurvey}(x)
}
\description{
Compute the survey intersections
}
