% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{time0}
\alias{time0}
\alias{time0<-}
\alias{setTime0}
\title{'time-zero' of every traces}
\usage{
\S4method{time0}{GPR}(x)

\S4method{time0}{GPR}(x) <- value

\S4method{setTime0}{GPR}(x, t0, track = TRUE)
}
\arguments{
\item{x}{An object of the class GPR.}
}
\value{
A vector containing the time-zero values of each traces.
}
\description{
\code{time0} returns the 'time-zero' of every traces. Generally, 
'time-zero' corresponds to the first wave arrival (also called first wave
break).
}
\examples{
data(frenkeLine00)
time0(frenkeLine00)
}
\seealso{
\code{\link{firstBreak}} to estimate the first wave break.
}
