% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{time0Cor}
\alias{time0Cor}
\title{Time zero correction}
\usage{
\S4method{time0Cor}{GPR}(x, t0 = NULL, method = c("pchip", "linear",
  "nearest", "spline", "cubic", "none"), crop = TRUE, keep = 0,
  track = TRUE)
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}

\item{t0}{[\code{DEPRECATED}] DEPRECATED - NO MORE USED.
Instead, set time-zero with either 
\code{time0(x) <- ...} or 
\code{x <- setTime0(x, ...)}.}

\item{method}{[\code{character(1)}] Interpolation method to be applied
(one of \code{pchip} \code{linear}, \code{nearest}, 
\code{spline}, \code{cubic}, \code{none}, 
see also \code{\link[signal]{interp1}}). 
 \code{"none"} means that the trace is shifted by the
amount of trace samples the closest to \code{ts} without
interpolation.}

\item{crop}{[\code{logical(1)}] 
If \code{TRUE} (default), remove the rows containing only 
zero's (no data).}

\item{keep}{[\code{DEPRECATED}] DEPRECATED - NO MORE USED.}
}
\value{
[\code{GPR class}] An object of the class \code{GPR}
}
\description{
\code{time0Cor} shift the traces vertically such that they start at
time zero (time zero of the data can be modified with the function).
New traces are interpolated.
}
\details{
This function is a wrapper for the following commands
\itemize{
  \item \code{ts <- -t0 + keep} (or if \code{t0} is \code{NULL}, 
        \code{ts <- -time0(x) + keep})
  \item \code{x <- traceShift( x,  ts, method = method, crop = crop)}
  \item \code{time0(x) <- time0(x) + ts}
}

Modified slots
\itemize{
  \item \code{data}: trace shifted. The number of rows of data may 
        be smaller if \code{crop = TRUE}.
  \item \code{time0}: set to 0.
  \item \code{proc}: updated with function name and arguments.
}
}
\examples{
data(frenkeLine00)
tfb <- firstBreak(frenkeLine00)
t0 <- firstBreakToTime0(tfb, frenkeLine00, c0 = 0.299)
time0(frenkeLine00) <- t0
frenkeLine00_2 <- time0Cor(frenkeLine00, method = "pchip")

}
\seealso{
\code{\link{firstBreak}} to estimate the first wave break;
         \code{\link{firstBreakToTime0}} to convert the first wave break
         into time zero.
         \code{\link{time0}} and \code{\link{setTime0}} to set time-zero;
         \code{\link{traceShift}} to shift the traces
}
