% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{firstBreakToTime0}
\alias{firstBreakToTime0}
\alias{timeCorOffset}
\title{Convert first wave break to time-zero}
\usage{
firstBreakToTime0(fb, x, c0 = 0.299)

\S4method{timeCorOffset}{GPR}(x, t0 = NULL, track = TRUE)
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}

\item{t0}{[\code{DEPRECATED}] DEPRECATED - NO MORE USED.
Instead, set time-zero with either 
\code{time0(x) <- ...} or 
\code{x <- setTime0(x, ...)}.}
}
\value{
[\code{GPR class}] An object of the class \code{GPR}
}
\description{
Account for the delay time between time of wave emission and time of first
wave break recording due to the antenna separation (offset).

\code{timeCorOffset} applies a time correction to every traces 
to compensate the offset between transmitter 
and receiver antennae (it converts the trace time of the data acquired with
a bistatic antenna system into trace time data virtually acquiered with 
a monostatic system under the assumption of horizontally layered structure).
If all the traces have the same time-zero, this function does not change the 
trace but only the time (time scale). If the traces have different
time-zero, the traces are first aligned to have the same time-zero 
(spline interpolation)
}
\details{
Modified slots
\itemize{
  \item \code{data}: trace shifted to time-zeor. The number of rows of data 
        may be smaller.
  \item \code{time0}: set to 0.
  \item \code{depth}: adapted to a virtual monostatic system. 
        FIXME: not regularly spaced!
  \item \code{antsep}: set to 0.
  \item \code{proc}: updated with function name and arguments.
}
}
\seealso{
\code{\link{time0}} to set time zero and 
         \code{\link{firstBreakToTime0}} to convert the first wave break
         into time zero.
}
