% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{timeCorOffset}
\alias{timeCorOffset}
\title{Constant-offset correction (time) of the GPR data}
\usage{
\S4method{timeCorOffset}{GPR}(x)
}
\description{
Time correction for each trace to compensate the offset between transmitter 
and receiver antennae (it converts the trace time of the data acquired with
a bistatic antenna system into trace time data virtually acquiered with 
a monostatic system under the assumption of horizontally layered structure).
Note that this function does not change the trace but only the time (time
scale).
}
