% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{traceAverage}
\alias{traceAverage}
\title{Trace average}
\usage{
\S4method{traceAverage}{GPR}(x, w = NULL, FUN = mean, ...)
}
\arguments{
\item{x}{An object of the class GPR}

\item{w}{A window length defining width of the average window. Default value
is \code{NULL} and with this default value the function returns a 
single trace that is the average trace of the radargram.}

\item{FUN}{A function to compute the average (default is \code{mean})}

\item{...}{Additional parameters for the FUN functions}
}
\description{
Compute the average trace of a radargram (resulting in a single trace) or
a moving average of the trace.
}
\examples{
data("frenkeLine00")

f0 <- frenkeLine00

f1 <- trAverage(f0)
plot(f1)
# substract the average trace
plot(f0 - f1)

f2 <- trAverage(f0, w = 20)
plot(f2)
plot(f0 - f2)

f3 <- trAverage(f0, w = 20, FUN = median)
plot(f3)
plot(f0 - f3)
}

