% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{traceAverage}
\alias{traceAverage}
\title{Trace average (DEPRECATED, use 'traceStat' instead)}
\usage{
\S4method{traceAverage}{GPR}(x, w = NULL, FUN = mean, ...)
}
\arguments{
\item{x}{An object of the class GPR}

\item{w}{A length-one integer vector equal to the window length of the 
average window. If \code{w = NULL} a single trace corresponding to
the average trace of the whole profile is returned.}

\item{FUN}{A function to compute the average (default is \code{mean})}

\item{...}{Additional parameters for the FUN functions}
}
\value{
An object of the class GPR. When \code{w = NULL}, this function 
        returns a GPR object with a single trace corresponding to the 
        average trace of the whole radargram. When \code{w} is equal to a
        strictly positive interger this function returns a GPR object with
        a size identical to x where each trace corresponds to the average
        of the \code{w} neighbouring traces centered on the considered trace.
}
\description{
Average traces in a radargram along the distance (horizontal) axis using
a moving window. This can be used to increase signal to noise ratio. Note that if 
the moving window length is not defined, all traces are averaged into one single trace.
}
\examples{
data("frenkeLine00")

f0 <- frenkeLine00

f1 <- traceAverage(f0)
plot(f1)
# substract the average trace
plot(f0 - f1)

f2 <- traceAverage(f0, w = 20)
plot(f2)
plot(f0 - f2)

f3 <- traceAverage(f0, w = 20, FUN = median)
plot(f3)
plot(f0 - f3)
}
