% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{traceShift}
\alias{traceShift}
\title{Shift traces vertically by an amount of depth (time) units. New traces 
are interpolated.}
\usage{
\S4method{traceShift}{GPR}(x, ts, method = c("spline", "linear", "nearest",
  "pchip", "cubic", "none"), crop = TRUE)
}
\arguments{
\item{x}{A object of the class GPR}

\item{ts}{A numeric vector defining the amount of depth the traces have to
be shifted}

\item{method}{A length-one character vector indicating the interpolation
method. \code{"none"} means that the trace is shifted by the
amount of points that is the closest to amount of depth 
\code{ts}.}

\item{crop}{If TRUE (defaults), remove the rows containing only zero's 
(no data).,}
}
\value{
An object of the class GPR.
}
\description{
Shift traces vertically by an amount of depth (time) units. New traces 
are interpolated.
}
\seealso{
\code{\link{time0Cor}} to shift the traces such that they start
         at time zero.
}
