% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapply.R
\name{trapply}
\alias{trapply}
\alias{trapply,GPR-method}
\title{Trace statistics}
\usage{
trapply(x, w = NULL, FUN = mean, ..., track = TRUE)

\S4method{trapply}{GPR}(x, w = NULL, FUN = mean, ..., track = TRUE)
}
\arguments{
\item{x}{An object of the class GPR}

\item{w}{A length-one integer vector equal to the window length of the
average window. If \code{w = NULL} similar to \code{apply(x, MARGIN = 2, FUN, ...)}}

\item{FUN}{A function to compute the average (default is \code{mean})}

\item{...}{Additional parameters for the FUN functions}

\item{track}{(\code{logical[1]}) Should the processing step be tracked?}
}
\value{
An object of the class GPR. When \code{w = NULL}, this function
returns a GPR object with a single trace corresponding to the
average trace of the whole radargram. When \code{w} is equal to a
strictly positive interger this function returns a GPR object with
a size identical to x where each trace corresponds to the average
of the \code{w} neighbouring traces centered on the considered trace.
}
\description{
\code{trapply} is a generic function used to produce results defined
by an user function. The user function is applied accross traces (horizontal)
using a moving window. Note that if the moving window length is not defined,
all traces are averaged into one single trace (the results is similar to
\code{apply(x, 1, FUN, ...)}.
}
\concept{processing}
