% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsample.R
\docType{methods}
\name{upsample}
\alias{upsample}
\title{Up-sample the GPR data (1D and 2D sinc-interpolation)}
\usage{
\S4method{upsample}{GPR}(x, n)
}
\description{
Up-sample the GPR data (1D and 2D sinc-interpolation)
}
