% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vel.R
\name{vel}
\alias{vel}
\alias{vel,GPR-method}
\title{Velocity}
\usage{
vel(x)

\S4method{vel}{GPR}(x)
}
\arguments{
\item{x}{[\code{GPR class}] An object of the class \code{GPR}}
}
\value{
[\code{list}|\code{numeric}|\code{matrix}] The velocities as they
        are stored in \code{x}.
}
\description{
Returns (raw) velocity data.
}
