% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/getterSetter.R
\docType{methods}
\name{vel}
\alias{vel}
\alias{vel<-}
\title{Velocity model of the GPR data}
\usage{
\S4method{vel}{GPR}(x)

\S4method{vel}{GPR}(x) <- value
}
\description{
Velocity model of the GPR data
}
