% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R, R/ClassGPRvirtual.R
\docType{methods}
\name{vel}
\alias{vel}
\alias{vel<-}
\alias{vel}
\alias{vel<-}
\title{Velocity model of the GPR data}
\usage{
\S4method{vel}{GPRvirtual}(x)

\S4method{vel}{GPRvirtual}(x) <- value
}
\description{
Velocity model of the GPR data
}
