% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velLayers.R
\name{velLayers}
\alias{velLayers}
\alias{velLayers,GPR-method}
\title{Print the list of delineation of the GPR data}
\usage{
velLayers(obj, name = NULL)

\S4method{velLayers}{GPR}(obj, name = NULL)
}
\arguments{
\item{obj}{(\code{GPR* object}) An object of the class \code{GPR}}

\item{name}{(\code{character[n]}) Names of the delineations.}
}
\description{
Print and invisible returns the GPR data delineations.
}
