% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassGPR.R
\docType{methods}
\name{writeGPR}
\alias{writeGPR}
\title{Write the GPR object in a file.}
\usage{
\S4method{writeGPR}{GPR}(x, fPath, format = c("DT1", "rds"),
  overwrite = FALSE)
}
\description{
Write the GPR object in a file.
}

