% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGPR.R
\name{writeGPR}
\alias{writeGPR}
\alias{writeGPR,GPR-method}
\alias{writeGPR,GPRsurvey-method}
\title{Write the GPR object in a file.}
\usage{
writeGPR(
  obj,
  dsn = NULL,
  format = c("rds", "dt1", "ascii", "xta", "xyza", "vtk"),
  overwrite = FALSE,
  ...
)

\S4method{writeGPR}{GPR}(
  obj,
  dsn = NULL,
  format = c("rds", "dt1", "ascii", "xta", "xyza", "vtk"),
  overwrite = FALSE,
  ...
)

\S4method{writeGPR}{GPRsurvey}(
  obj,
  dsn = NULL,
  format = c("DT1", "rds", "ASCII", "xta", "xyzv", "vtk"),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{obj}{Object of the class \code{GPR} or \code{GPRsurvey}}

\item{dsn}{Filepath (Length-one character vector). If \code{dsn = NULL},
the file will be save in the current working directory with
the name of obj (\code{name(obj)}) with the extension depending
of \code{format}.}

\item{format}{Format type. See Details.}

\item{overwrite}{Boolean. If \code{TRUE} existing files will be overwritten,
if \code{FALSE} an error will be thrown if the file(s)
already exist(s).}

\item{...}{additional parameters to be passed to \code{\link[=write.table]{write.table()}}
when \code{format = "ASCII"} or \code{format = "xyza"}.}
}
\description{
Write the GPR object in a file.
}
\seealso{
\code{\link[=readGPR]{readGPR()}}
}
