% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ClassGPR.R
\docType{methods}
\name{writeGPR}
\alias{writeGPR}
\alias{writeGPR}
\title{Write the GPR object in a file.}
\usage{
writeGPR(x, fPath = NULL, type = c("DT1", "rds", "ASCII", "xyzv"),
  overwrite = FALSE, ...)

\S4method{writeGPR}{GPR}(x, fPath = NULL, type = c("DT1", "rds", "ASCII",
  "xyzv"), overwrite = FALSE, ...)
}
\description{
Write the GPR object in a file.
}
