% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGPR.R
\name{writeGPR}
\alias{writeGPR}
\title{Write GPR data}
\usage{
writeGPR(
  x,
  fPath = NULL,
  type = c("rds", "DT1", "SGY", "ASCII", "xta", "xyza"),
  overwrite = FALSE,
  endian = .Platform$endian,
  ...
)

\S4method{writeGPR}{GPR}(
  x,
  fPath = NULL,
  type = c("rds", "DT1", "SGY", "ASCII", "xta", "xyza"),
  overwrite = FALSE,
  endian = .Platform$endian,
  ...
)
}
\arguments{
\item{x}{Object of the class \code{GPR} or \code{GPRsurvey}}

\item{fPath}{Filepath (Length-one character vector). If \code{fPath = NULL},
the file will be save in the current working directory with
the name of x (\code{name(x)}) with the extension depending 
of \code{type}.}

\item{type}{Format type. See Details.}

\item{overwrite}{Boolean. If \code{TRUE} existing files will be overwritten,
if \code{FALSE} an error will be thrown if the file(s) 
already exist(s).}
}
\description{
Write GPR data to a specified file format (see Details). Add automatically
the file extension if missing.
}
\details{
Modified slots
\itemize{
  \item \code{rds}: Internal R format (useful if you aim to save your data
         for later processing in R).
  \item \code{DT1}: Sensors & Software file format
  \item \code{ASCII}: ".txt" format; write the GPR data as matrix like format with 
        column names equal to trace position and row name equal to depth/time
  \item \code{xta}: write the GPR data as a 3-column ".txt" file. The columns
        correspond to 1) trace position, 2) time/depth, and 3) amplitude.
  \item \code{xyza}: write the GPR data as a 4-column ".txt" file.
        The columns correspond to 1) trace x-position, 2) trace y-position, 
        3) time/depth, and 4) amplitude.
}
}
\seealso{
\code{\link{readGPR}}
}
